Add table information to nlmixr fit object without tables

Add table information to nlmixr fit object without tables

addTable(
  object,
  updateObject = FALSE,
  data = object$dataSav,
  thetaEtaParameters = .foceiThetaEtaParameters(object),
  table = tableControl(),
  keep = NULL,
  drop = NULL,
  envir = parent.frame(1)
)

Arguments

object

nlmixr family of objects

updateObject

Update the object (default FALSE)

data

Saved data from

thetaEtaParameters

Intenral theta/eta parameters

table

a `tableControl()` list of options

keep

Character Vector of items to keep

drop

Character Vector of items to drop or NULL

envir

ENvironment to search for updating

Value

Fit with table information attached

Author

Matthew Fidler

Examples

# \donttest{

one.cmt <- function() {
  ini({
    ## You may label each parameter with a comment
    tka <- 0.45 # Log Ka
    tcl <- log(c(0, 2.7, 100)) # Log Cl
    ## This works with interactive models
    ## You may also label the preceding line with label("label text")
    tv <- 3.45; label("log V")
    ## the label("Label name") works with all models
    eta.ka ~ 0.6
    eta.cl ~ 0.3
    eta.v ~ 0.1
    add.sd <- 0.7
  })
  model({
    ka <- exp(tka + eta.ka)
    cl <- exp(tcl + eta.cl)
    v <- exp(tv + eta.v)
    linCmt() ~ add(add.sd)
  })
}

# run without tables step
f <- nlmixr(one.cmt, theo_sd, "saem", control=list(calcTables=FALSE))
#> ℹ parameter labels from comments will be replaced by 'label()'
#>  
#> 1:    0.2305   0.9671   3.4723   0.5700   0.2850   0.0950   4.4919
#> 2:    0.3651   0.9778   3.5295   0.5415   0.2707   0.0902   1.7823
#> 3:    0.4882   0.9147   3.4856   0.5144   0.2572   0.0857   0.9433
#> 4:    0.4388   0.9472   3.4840   0.4887   0.2444   0.0815   0.7109
#> 5:    0.4829   0.9508   3.4741   0.4643   0.2321   0.0774   0.6214
#> 6:    0.4920   0.9792   3.4720   0.4411   0.2205   0.0735   0.5153
#> 7:    0.4777   1.0028   3.4600   0.4269   0.2095   0.0698   0.5071
#> 8:    0.4391   0.9985   3.4583   0.4056   0.1990   0.0663   0.5103
#> 9:    0.4114   0.9775   3.4580   0.3853   0.1891   0.0630   0.5211
#> 10:    0.4107   0.9914   3.4474   0.4237   0.1796   0.0599   0.5031
#> 11:    0.4334   1.0173   3.4396   0.4026   0.1706   0.0569   0.5035
#> 12:    0.4362   1.0233   3.4396   0.3824   0.1621   0.0540   0.4856
#> 13:    0.3955   1.0403   3.4232   0.3633   0.1540   0.0513   0.4968
#> 14:    0.4067   1.0196   3.4387   0.4627   0.1463   0.0488   0.4935
#> 15:    0.4418   1.0214   3.4495   0.4395   0.1390   0.0463   0.4847
#> 16:    0.4555   1.0222   3.4523   0.4272   0.1320   0.0440   0.5063
#> 17:    0.4723   0.9985   3.4633   0.4328   0.1254   0.0418   0.4749
#> 18:    0.4713   0.9710   3.4543   0.4301   0.1192   0.0397   0.5001
#> 19:    0.4586   1.0053   3.4522   0.4213   0.1132   0.0377   0.4762
#> 20:    0.4686   1.0159   3.4491   0.4415   0.1075   0.0358   0.4926
#> 21:    0.4525   1.0063   3.4582   0.4194   0.1022   0.0341   0.4619
#> 22:    0.4383   1.0097   3.4606   0.4134   0.0971   0.0324   0.4611
#> 23:    0.5105   1.0017   3.4695   0.4324   0.0922   0.0307   0.4693
#> 24:    0.4888   1.0081   3.4691   0.4107   0.0876   0.0292   0.4601
#> 25:    0.4573   1.0045   3.4623   0.3902   0.0890   0.0277   0.4746
#> 26:    0.4185   1.0213   3.4423   0.3951   0.0845   0.0264   0.4757
#> 27:    0.4101   1.0423   3.4402   0.3753   0.0803   0.0250   0.5107
#> 28:    0.4670   1.0181   3.4478   0.3566   0.0763   0.0238   0.4552
#> 29:    0.4637   1.0191   3.4483   0.3514   0.0742   0.0226   0.4823
#> 30:    0.4823   1.0191   3.4489   0.3386   0.0743   0.0215   0.4951
#> 31:    0.4069   1.0082   3.4580   0.3217   0.0849   0.0204   0.4833
#> 32:    0.4274   0.9955   3.4673   0.3516   0.0806   0.0194   0.4877
#> 33:    0.4486   1.0011   3.4596   0.3814   0.0766   0.0184   0.5041
#> 34:    0.4835   1.0048   3.4582   0.4096   0.0728   0.0175   0.4855
#> 35:    0.4500   1.0009   3.4576   0.3892   0.0691   0.0166   0.4789
#> 36:    0.4521   1.0192   3.4494   0.3697   0.0690   0.0158   0.4873
#> 37:    0.4443   1.0203   3.4556   0.3676   0.0713   0.0195   0.4830
#> 38:    0.5028   1.0178   3.4465   0.4699   0.0773   0.0186   0.4955
#> 39:    0.4952   0.9817   3.4726   0.4464   0.0735   0.0237   0.4637
#> 40:    0.4771   0.9930   3.4763   0.4241   0.0698   0.0260   0.4682
#> 41:    0.4646   0.9895   3.4716   0.4029   0.0663   0.0247   0.4842
#> 42:    0.4724   1.0117   3.4602   0.3828   0.0748   0.0235   0.4609
#> 43:    0.4411   1.0275   3.4433   0.3636   0.0755   0.0223   0.4672
#> 44:    0.4393   1.0245   3.4456   0.3676   0.0718   0.0212   0.4858
#> 45:    0.4213   1.0442   3.4313   0.3932   0.0682   0.0201   0.4756
#> 46:    0.4131   1.0255   3.4391   0.4216   0.0779   0.0191   0.4802
#> 47:    0.4401   1.0325   3.4398   0.4268   0.0740   0.0181   0.4612
#> 48:    0.4603   1.0197   3.4460   0.4863   0.0726   0.0196   0.4734
#> 49:    0.4523   1.0101   3.4449   0.4619   0.0752   0.0186   0.4863
#> 50:    0.4857   0.9838   3.4442   0.4439   0.0821   0.0178   0.5032
#> 51:    0.4827   1.0036   3.4402   0.4614   0.0890   0.0190   0.5150
#> 52:    0.5207   0.9994   3.4615   0.5074   0.0845   0.0180   0.4780
#> 53:    0.5052   1.0095   3.4547   0.4987   0.0803   0.0171   0.4658
#> 54:    0.4724   1.0147   3.4560   0.4737   0.0763   0.0163   0.4847
#> 55:    0.4456   1.0103   3.4482   0.4500   0.0770   0.0155   0.4727
#> 56:    0.4556   0.9972   3.4503   0.4310   0.0731   0.0180   0.4873
#> 57:    0.4717   0.9961   3.4601   0.4094   0.0695   0.0178   0.5002
#> 58:    0.5111   0.9884   3.4595   0.3906   0.0689   0.0169   0.4843
#> 59:    0.5299   0.9601   3.4775   0.4401   0.0848   0.0160   0.5179
#> 60:    0.5258   0.9706   3.4786   0.4721   0.0806   0.0172   0.5463
#> 61:    0.5476   0.9869   3.4828   0.4989   0.0846   0.0175   0.5335
#> 62:    0.4957   0.9961   3.4819   0.4739   0.0873   0.0197   0.5230
#> 63:    0.5010   0.9844   3.4750   0.4502   0.0868   0.0211   0.4994
#> 64:    0.4980   0.9863   3.4719   0.4594   0.0856   0.0201   0.5006
#> 65:    0.4999   0.9879   3.4558   0.4840   0.0902   0.0191   0.4858
#> 66:    0.4855   1.0094   3.4708   0.4598   0.0965   0.0181   0.4656
#> 67:    0.4799   0.9869   3.4526   0.4368   0.0917   0.0172   0.4720
#> 68:    0.4261   1.0070   3.4548   0.4150   0.0871   0.0163   0.5139
#> 69:    0.4372   1.0071   3.4479   0.4108   0.0827   0.0155   0.5039
#> 70:    0.4523   1.0070   3.4606   0.4929   0.0786   0.0156   0.5108
#> 71:    0.4312   1.0303   3.4359   0.4682   0.0770   0.0181   0.5383
#> 72:    0.4553   0.9787   3.4556   0.4448   0.0749   0.0172   0.4921
#> 73:    0.4797   1.0105   3.4687   0.4226   0.0749   0.0163   0.4876
#> 74:    0.4795   1.0006   3.4598   0.4014   0.0712   0.0155   0.4624
#> 75:    0.5021   0.9811   3.4820   0.4242   0.0778   0.0187   0.4671
#> 76:    0.5250   1.0031   3.4901   0.4030   0.0739   0.0192   0.4885
#> 77:    0.5216   0.9688   3.4883   0.4116   0.0702   0.0199   0.5193
#> 78:    0.5108   0.9865   3.4845   0.4293   0.0756   0.0221   0.4785
#> 79:    0.5097   0.9947   3.4755   0.4346   0.0718   0.0254   0.5220
#> 80:    0.5230   0.9928   3.4795   0.4593   0.0682   0.0265   0.5049
#> 81:    0.5519   0.9995   3.4804   0.4364   0.0648   0.0272   0.4777
#> 82:    0.5369   0.9764   3.4881   0.4180   0.0616   0.0258   0.4909
#> 83:    0.5110   0.9914   3.4757   0.4266   0.0585   0.0245   0.4752
#> 84:    0.4835   0.9822   3.4678   0.4053   0.0630   0.0233   0.4759
#> 85:    0.4943   1.0028   3.4816   0.3851   0.0631   0.0221   0.4854
#> 86:    0.4739   1.0058   3.4669   0.3658   0.0737   0.0210   0.4683
#> 87:    0.4323   1.0174   3.4464   0.3475   0.0727   0.0200   0.4687
#> 88:    0.3573   1.0096   3.4444   0.3565   0.0739   0.0190   0.5053
#> 89:    0.4307   1.0124   3.4422   0.3416   0.0705   0.0180   0.5228
#> 90:    0.3981   1.0268   3.4291   0.3703   0.0670   0.0171   0.5147
#> 91:    0.4057   1.0303   3.4166   0.4033   0.0742   0.0183   0.5056
#> 92:    0.4643   1.0388   3.4407   0.4198   0.0922   0.0214   0.5000
#> 93:    0.3853   1.0218   3.4418   0.3988   0.0876   0.0203   0.4876
#> 94:    0.4144   1.0191   3.4380   0.3871   0.0971   0.0193   0.5084
#> 95:    0.4261   1.0446   3.4388   0.4601   0.0923   0.0184   0.4936
#> 96:    0.4135   1.0207   3.4357   0.4725   0.0877   0.0183   0.4864
#> 97:    0.4439   1.0279   3.4522   0.4489   0.0833   0.0181   0.4938
#> 98:    0.4766   1.0194   3.4610   0.4265   0.0791   0.0172   0.4916
#> 99:    0.4605   0.9854   3.4526   0.4580   0.0752   0.0164   0.4765
#> 100:    0.4330   0.9965   3.4527   0.4351   0.0714   0.0163   0.4743
#> 101:    0.3980   1.0110   3.4321   0.4134   0.0747   0.0170   0.4516
#> 102:    0.4152   1.0176   3.4287   0.3927   0.0710   0.0185   0.4891
#> 103:    0.4471   1.0166   3.4446   0.4006   0.0689   0.0192   0.4586
#> 104:    0.4308   1.0155   3.4490   0.4306   0.0654   0.0183   0.4998
#> 105:    0.4187   0.9930   3.4461   0.4091   0.0808   0.0174   0.5012
#> 106:    0.4735   0.9846   3.4557   0.4018   0.0812   0.0165   0.4898
#> 107:    0.5128   1.0054   3.4604   0.4465   0.0885   0.0157   0.4790
#> 108:    0.4794   0.9893   3.4599   0.4242   0.0841   0.0149   0.5088
#> 109:    0.4480   0.9997   3.4670   0.4030   0.0866   0.0160   0.4887
#> 110:    0.4750   1.0124   3.4507   0.3828   0.0823   0.0167   0.4980
#> 111:    0.4563   1.0069   3.4587   0.3637   0.0790   0.0160   0.4819
#> 112:    0.4692   0.9962   3.4481   0.3649   0.0770   0.0169   0.4872
#> 113:    0.4327   1.0425   3.4484   0.3708   0.0768   0.0161   0.4796
#> 114:    0.4215   1.0422   3.4462   0.3522   0.0953   0.0153   0.4642
#> 115:    0.4546   1.0155   3.4399   0.3469   0.0920   0.0169   0.4788
#> 116:    0.4536   1.0027   3.4477   0.3984   0.0874   0.0190   0.4825
#> 117:    0.4284   1.0386   3.4491   0.3851   0.0859   0.0181   0.4793
#> 118:    0.4485   1.0224   3.4483   0.3658   0.0841   0.0172   0.4807
#> 119:    0.4783   1.0052   3.4558   0.3999   0.0799   0.0216   0.4871
#> 120:    0.4478   0.9837   3.4651   0.4012   0.0759   0.0206   0.4723
#> 121:    0.4564   1.0135   3.4587   0.3812   0.0721   0.0199   0.4664
#> 122:    0.4615   0.9840   3.4500   0.3978   0.0685   0.0197   0.4688
#> 123:    0.5128   1.0198   3.4507   0.4289   0.0715   0.0196   0.4634
#> 124:    0.4910   1.0332   3.4499   0.4603   0.0679   0.0225   0.4727
#> 125:    0.4774   1.0296   3.4479   0.4373   0.0714   0.0214   0.4699
#> 126:    0.4323   1.0446   3.4350   0.4352   0.0782   0.0203   0.4749
#> 127:    0.4545   1.0391   3.4337   0.4134   0.0817   0.0193   0.4959
#> 128:    0.4108   1.0267   3.4404   0.4637   0.0777   0.0229   0.5114
#> 129:    0.4250   1.0351   3.4494   0.4985   0.0762   0.0217   0.4695
#> 130:    0.5269   1.0264   3.4570   0.5079   0.0745   0.0207   0.4488
#> 131:    0.4508   1.0179   3.4401   0.4825   0.0709   0.0196   0.4701
#> 132:    0.4299   1.0122   3.4282   0.4583   0.0734   0.0186   0.5105
#> 133:    0.4165   1.0089   3.4401   0.4354   0.0726   0.0177   0.4845
#> 134:    0.4491   0.9936   3.4602   0.4137   0.0749   0.0197   0.4931
#> 135:    0.4562   0.9995   3.4599   0.3930   0.0796   0.0191   0.4891
#> 136:    0.4833   0.9866   3.4660   0.3733   0.0756   0.0182   0.4804
#> 137:    0.4878   0.9942   3.4615   0.3783   0.0718   0.0173   0.4995
#> 138:    0.4698   1.0094   3.4572   0.4097   0.0682   0.0173   0.4824
#> 139:    0.4838   1.0133   3.4667   0.4908   0.0697   0.0213   0.4688
#> 140:    0.4906   1.0098   3.4702   0.4663   0.0693   0.0202   0.4961
#> 141:    0.4826   0.9823   3.4663   0.4430   0.0727   0.0212   0.4684
#> 142:    0.4752   0.9888   3.4644   0.4208   0.0746   0.0201   0.4767
#> 143:    0.4569   1.0147   3.4593   0.4186   0.0709   0.0208   0.4761
#> 144:    0.4881   1.0258   3.4572   0.3977   0.0676   0.0197   0.4520
#> 145:    0.4463   1.0301   3.4403   0.4076   0.0693   0.0193   0.4448
#> 146:    0.4252   1.0312   3.4518   0.4247   0.0663   0.0189   0.4628
#> 147:    0.4406   1.0123   3.4379   0.4034   0.0695   0.0179   0.4513
#> 148:    0.4028   1.0283   3.4515   0.3863   0.0660   0.0176   0.4600
#> 149:    0.4380   1.0134   3.4537   0.3705   0.0715   0.0195   0.4572
#> 150:    0.4460   1.0115   3.4409   0.3890   0.0695   0.0185   0.4753
#> 151:    0.3816   1.0358   3.4308   0.3696   0.0660   0.0189   0.4664
#> 152:    0.4094   1.0377   3.4363   0.3878   0.0736   0.0205   0.4603
#> 153:    0.4208   1.0246   3.4532   0.3954   0.0651   0.0225   0.4678
#> 154:    0.4760   1.0340   3.4597   0.3910   0.0733   0.0216   0.4652
#> 155:    0.4678   1.0123   3.4597   0.4495   0.0627   0.0242   0.4619
#> 156:    0.4882   1.0151   3.4639   0.4176   0.0778   0.0244   0.4968
#> 157:    0.4449   1.0079   3.4669   0.3499   0.0622   0.0259   0.4942
#> 158:    0.4593   1.0261   3.4542   0.3701   0.0581   0.0261   0.4708
#> 159:    0.4133   1.0247   3.4577   0.3636   0.0523   0.0238   0.5084
#> 160:    0.4843   0.9950   3.4734   0.5106   0.0621   0.0193   0.5039
#> 161:    0.4660   0.9900   3.4651   0.3748   0.0475   0.0230   0.5116
#> 162:    0.4052   1.0211   3.4512   0.4324   0.0512   0.0233   0.5002
#> 163:    0.4668   1.0048   3.4494   0.4723   0.0458   0.0217   0.4768
#> 164:    0.4520   1.0149   3.4540   0.4130   0.0464   0.0243   0.4751
#> 165:    0.4638   1.0211   3.4498   0.3954   0.0483   0.0214   0.4819
#> 166:    0.4716   1.0375   3.4603   0.3928   0.0440   0.0226   0.4920
#> 167:    0.4907   1.0337   3.4710   0.3186   0.0528   0.0252   0.4956
#> 168:    0.4645   1.0322   3.4643   0.3550   0.0579   0.0253   0.4877
#> 169:    0.4779   1.0502   3.4501   0.3294   0.0687   0.0195   0.5175
#> 170:    0.4641   1.0227   3.4524   0.3495   0.0525   0.0207   0.5195
#> 171:    0.4333   1.0251   3.4446   0.3388   0.0605   0.0212   0.4732
#> 172:    0.4484   1.0351   3.4497   0.3383   0.0633   0.0223   0.4638
#> 173:    0.4087   1.0325   3.4342   0.3153   0.0635   0.0287   0.4709
#> 174:    0.4407   1.0396   3.4500   0.3038   0.0544   0.0282   0.4815
#> 175:    0.4513   1.0338   3.4519   0.3305   0.0612   0.0279   0.4849
#> 176:    0.4291   1.0364   3.4646   0.3252   0.0608   0.0211   0.4784
#> 177:    0.4815   1.0242   3.4710   0.4081   0.0677   0.0254   0.4577
#> 178:    0.4585   1.0127   3.4598   0.3493   0.0742   0.0195   0.4455
#> 179:    0.4878   1.0230   3.4559   0.3733   0.0917   0.0161   0.4883
#> 180:    0.4427   1.0317   3.4615   0.2961   0.0775   0.0163   0.5045
#> 181:    0.4472   1.0155   3.4380   0.3325   0.0777   0.0178   0.4870
#> 182:    0.4136   1.0280   3.4362   0.3177   0.0863   0.0171   0.4951
#> 183:    0.4261   1.0369   3.4350   0.3631   0.0744   0.0179   0.4868
#> 184:    0.4362   1.0186   3.4395   0.4078   0.0759   0.0166   0.4679
#> 185:    0.3912   1.0324   3.4332   0.3703   0.0735   0.0181   0.4752
#> 186:    0.3899   1.0500   3.4340   0.3790   0.0719   0.0201   0.4749
#> 187:    0.4471   1.0159   3.4450   0.4073   0.0521   0.0220   0.4523
#> 188:    0.4325   1.0288   3.4478   0.4215   0.0657   0.0185   0.4640
#> 189:    0.4640   1.0269   3.4614   0.4195   0.0620   0.0203   0.4738
#> 190:    0.4584   1.0094   3.4703   0.3284   0.0639   0.0201   0.4751
#> 191:    0.4504   1.0322   3.4554   0.3686   0.0596   0.0199   0.4417
#> 192:    0.4375   1.0286   3.4520   0.3424   0.0682   0.0173   0.4629
#> 193:    0.4613   1.0300   3.4533   0.3388   0.0658   0.0194   0.4628
#> 194:    0.4070   1.0239   3.4385   0.4053   0.0638   0.0137   0.4819
#> 195:    0.4061   1.0279   3.4320   0.4020   0.0572   0.0139   0.4906
#> 196:    0.4134   1.0078   3.4434   0.4239   0.0793   0.0145   0.4803
#> 197:    0.4134   1.0191   3.4626   0.3640   0.0683   0.0177   0.5033
#> 198:    0.4838   0.9943   3.4596   0.4695   0.0774   0.0166   0.4694
#> 199:    0.4656   0.9885   3.4531   0.4601   0.0747   0.0167   0.4767
#> 200:    0.4431   0.9989   3.4501   0.4098   0.0791   0.0215   0.5171
#> 201:    0.4232   0.9965   3.4533   0.3900   0.0763   0.0212   0.5069
#> 202:    0.4189   1.0049   3.4498   0.4047   0.0749   0.0212   0.5113
#> 203:    0.4346   1.0043   3.4487   0.4139   0.0752   0.0218   0.5050
#> 204:    0.4403   1.0008   3.4494   0.4216   0.0734   0.0216   0.5040
#> 205:    0.4392   1.0072   3.4484   0.4205   0.0759   0.0220   0.5018
#> 206:    0.4427   1.0137   3.4478   0.4356   0.0754   0.0219   0.5002
#> 207:    0.4390   1.0159   3.4455   0.4369   0.0735   0.0218   0.4999
#> 208:    0.4385   1.0175   3.4459   0.4269   0.0718   0.0218   0.4970
#> 209:    0.4423   1.0164   3.4454   0.4281   0.0708   0.0217   0.4944
#> 210:    0.4462   1.0153   3.4466   0.4284   0.0700   0.0216   0.4915
#> 211:    0.4509   1.0161   3.4478   0.4295   0.0697   0.0213   0.4892
#> 212:    0.4552   1.0149   3.4486   0.4285   0.0691   0.0209   0.4901
#> 213:    0.4557   1.0115   3.4497   0.4241   0.0690   0.0207   0.4922
#> 214:    0.4617   1.0100   3.4513   0.4254   0.0696   0.0205   0.4926
#> 215:    0.4634   1.0094   3.4530   0.4228   0.0701   0.0205   0.4934
#> 216:    0.4687   1.0070   3.4543   0.4281   0.0702   0.0204   0.4945
#> 217:    0.4699   1.0071   3.4554   0.4298   0.0703   0.0205   0.4940
#> 218:    0.4711   1.0074   3.4555   0.4308   0.0703   0.0204   0.4943
#> 219:    0.4703   1.0069   3.4551   0.4301   0.0699   0.0203   0.4938
#> 220:    0.4677   1.0071   3.4546   0.4263   0.0698   0.0203   0.4938
#> 221:    0.4656   1.0071   3.4544   0.4216   0.0696   0.0203   0.4937
#> 222:    0.4652   1.0073   3.4540   0.4226   0.0696   0.0203   0.4939
#> 223:    0.4665   1.0072   3.4543   0.4236   0.0695   0.0203   0.4927
#> 224:    0.4683   1.0075   3.4547   0.4247   0.0696   0.0204   0.4917
#> 225:    0.4679   1.0063   3.4551   0.4231   0.0697   0.0204   0.4917
#> 226:    0.4697   1.0050   3.4554   0.4235   0.0698   0.0205   0.4913
#> 227:    0.4703   1.0041   3.4560   0.4227   0.0702   0.0205   0.4913
#> 228:    0.4695   1.0047   3.4557   0.4211   0.0700   0.0205   0.4912
#> 229:    0.4684   1.0051   3.4552   0.4213   0.0698   0.0207   0.4909
#> 230:    0.4684   1.0053   3.4548   0.4200   0.0698   0.0209   0.4902
#> 231:    0.4687   1.0057   3.4550   0.4177   0.0695   0.0211   0.4895
#> 232:    0.4679   1.0056   3.4551   0.4175   0.0693   0.0211   0.4886
#> 233:    0.4681   1.0045   3.4555   0.4188   0.0690   0.0211   0.4886
#> 234:    0.4684   1.0041   3.4557   0.4175   0.0683   0.0212   0.4887
#> 235:    0.4683   1.0034   3.4557   0.4161   0.0679   0.0213   0.4886
#> 236:    0.4681   1.0036   3.4558   0.4151   0.0676   0.0213   0.4884
#> 237:    0.4686   1.0033   3.4562   0.4155   0.0677   0.0213   0.4885
#> 238:    0.4688   1.0037   3.4562   0.4158   0.0676   0.0213   0.4891
#> 239:    0.4683   1.0043   3.4560   0.4157   0.0676   0.0214   0.4893
#> 240:    0.4681   1.0041   3.4561   0.4163   0.0677   0.0214   0.4888
#> 241:    0.4687   1.0033   3.4562   0.4172   0.0679   0.0214   0.4889
#> 242:    0.4696   1.0028   3.4565   0.4167   0.0681   0.0214   0.4891
#> 243:    0.4695   1.0036   3.4563   0.4158   0.0685   0.0214   0.4888
#> 244:    0.4683   1.0041   3.4564   0.4142   0.0687   0.0212   0.4891
#> 245:    0.4684   1.0039   3.4566   0.4121   0.0687   0.0211   0.4899
#> 246:    0.4690   1.0039   3.4568   0.4115   0.0688   0.0211   0.4900
#> 247:    0.4685   1.0038   3.4570   0.4111   0.0690   0.0211   0.4900
#> 248:    0.4679   1.0039   3.4570   0.4105   0.0690   0.0211   0.4901
#> 249:    0.4665   1.0042   3.4567   0.4090   0.0689   0.0210   0.4903
#> 250:    0.4671   1.0047   3.4567   0.4088   0.0690   0.0211   0.4903
#> 251:    0.4675   1.0050   3.4570   0.4091   0.0692   0.0210   0.4904
#> 252:    0.4681   1.0052   3.4571   0.4114   0.0691   0.0211   0.4902
#> 253:    0.4681   1.0057   3.4571   0.4121   0.0689   0.0211   0.4906
#> 254:    0.4679   1.0063   3.4571   0.4126   0.0688   0.0210   0.4908
#> 255:    0.4675   1.0068   3.4571   0.4117   0.0689   0.0211   0.4908
#> 256:    0.4674   1.0068   3.4571   0.4122   0.0690   0.0211   0.4903
#> 257:    0.4673   1.0063   3.4570   0.4122   0.0688   0.0210   0.4902
#> 258:    0.4670   1.0059   3.4571   0.4126   0.0688   0.0210   0.4899
#> 259:    0.4669   1.0057   3.4570   0.4120   0.0688   0.0210   0.4899
#> 260:    0.4665   1.0060   3.4570   0.4117   0.0688   0.0211   0.4896
#> 261:    0.4665   1.0061   3.4571   0.4103   0.0689   0.0211   0.4895
#> 262:    0.4660   1.0065   3.4570   0.4097   0.0690   0.0211   0.4892
#> 263:    0.4656   1.0067   3.4569   0.4088   0.0690   0.0211   0.4893
#> 264:    0.4656   1.0070   3.4568   0.4090   0.0692   0.0211   0.4896
#> 265:    0.4659   1.0069   3.4569   0.4095   0.0696   0.0210   0.4898
#> 266:    0.4660   1.0067   3.4570   0.4111   0.0695   0.0210   0.4898
#> 267:    0.4663   1.0069   3.4570   0.4139   0.0696   0.0209   0.4899
#> 268:    0.4667   1.0068   3.4570   0.4154   0.0696   0.0208   0.4900
#> 269:    0.4673   1.0065   3.4571   0.4158   0.0696   0.0208   0.4899
#> 270:    0.4680   1.0066   3.4570   0.4171   0.0696   0.0208   0.4894
#> 271:    0.4685   1.0067   3.4571   0.4176   0.0695   0.0208   0.4890
#> 272:    0.4688   1.0064   3.4571   0.4175   0.0696   0.0207   0.4892
#> 273:    0.4685   1.0064   3.4571   0.4168   0.0698   0.0207   0.4890
#> 274:    0.4682   1.0069   3.4571   0.4153   0.0700   0.0206   0.4890
#> 275:    0.4683   1.0070   3.4571   0.4145   0.0702   0.0206   0.4887
#> 276:    0.4679   1.0072   3.4570   0.4146   0.0703   0.0205   0.4887
#> 277:    0.4674   1.0074   3.4572   0.4139   0.0704   0.0204   0.4891
#> 278:    0.4670   1.0073   3.4571   0.4145   0.0704   0.0204   0.4896
#> 279:    0.4668   1.0071   3.4571   0.4153   0.0704   0.0203   0.4900
#> 280:    0.4668   1.0071   3.4573   0.4154   0.0705   0.0203   0.4898
#> 281:    0.4670   1.0074   3.4573   0.4160   0.0706   0.0202   0.4896
#> 282:    0.4672   1.0075   3.4572   0.4174   0.0707   0.0202   0.4894
#> 283:    0.4670   1.0077   3.4572   0.4184   0.0706   0.0201   0.4899
#> 284:    0.4667   1.0078   3.4571   0.4194   0.0706   0.0200   0.4895
#> 285:    0.4664   1.0080   3.4569   0.4195   0.0707   0.0199   0.4892
#> 286:    0.4661   1.0082   3.4568   0.4189   0.0707   0.0199   0.4889
#> 287:    0.4658   1.0086   3.4566   0.4189   0.0709   0.0198   0.4889
#> 288:    0.4655   1.0087   3.4566   0.4179   0.0709   0.0198   0.4890
#> 289:    0.4653   1.0086   3.4565   0.4167   0.0709   0.0198   0.4891
#> 290:    0.4654   1.0086   3.4565   0.4163   0.0710   0.0198   0.4892
#> 291:    0.4652   1.0090   3.4564   0.4162   0.0711   0.0198   0.4890
#> 292:    0.4648   1.0093   3.4563   0.4155   0.0713   0.0198   0.4889
#> 293:    0.4645   1.0092   3.4561   0.4149   0.0713   0.0198   0.4889
#> 294:    0.4639   1.0095   3.4558   0.4144   0.0713   0.0198   0.4888
#> 295:    0.4635   1.0093   3.4557   0.4141   0.0713   0.0198   0.4885
#> 296:    0.4632   1.0094   3.4556   0.4133   0.0714   0.0198   0.4884
#> 297:    0.4632   1.0095   3.4556   0.4134   0.0714   0.0199   0.4883
#> 298:    0.4633   1.0093   3.4556   0.4131   0.0715   0.0198   0.4881
#> 299:    0.4634   1.0092   3.4557   0.4134   0.0716   0.0198   0.4880
#> 300:    0.4633   1.0091   3.4557   0.4136   0.0715   0.0197   0.4881
#> 301:    0.4634   1.0090   3.4556   0.4132   0.0715   0.0197   0.4883
#> 302:    0.4638   1.0092   3.4556   0.4130   0.0716   0.0197   0.4880
#> 303:    0.4640   1.0092   3.4557   0.4129   0.0716   0.0197   0.4877
#> 304:    0.4641   1.0093   3.4556   0.4137   0.0716   0.0197   0.4876
#> 305:    0.4640   1.0092   3.4556   0.4142   0.0717   0.0197   0.4875
#> 306:    0.4639   1.0091   3.4556   0.4140   0.0717   0.0197   0.4875
#> 307:    0.4637   1.0091   3.4554   0.4131   0.0718   0.0197   0.4877
#> 308:    0.4636   1.0091   3.4553   0.4134   0.0718   0.0197   0.4879
#> 309:    0.4634   1.0090   3.4553   0.4135   0.0719   0.0196   0.4881
#> 310:    0.4634   1.0090   3.4551   0.4136   0.0719   0.0196   0.4886
#> 311:    0.4632   1.0089   3.4551   0.4135   0.0721   0.0196   0.4887
#> 312:    0.4630   1.0089   3.4550   0.4139   0.0722   0.0195   0.4887
#> 313:    0.4625   1.0088   3.4548   0.4140   0.0721   0.0195   0.4889
#> 314:    0.4622   1.0087   3.4548   0.4140   0.0721   0.0195   0.4890
#> 315:    0.4620   1.0087   3.4548   0.4140   0.0721   0.0195   0.4888
#> 316:    0.4619   1.0090   3.4547   0.4144   0.0720   0.0194   0.4888
#> 317:    0.4618   1.0091   3.4547   0.4137   0.0721   0.0194   0.4887
#> 318:    0.4618   1.0092   3.4547   0.4132   0.0721   0.0194   0.4887
#> 319:    0.4617   1.0092   3.4547   0.4125   0.0721   0.0194   0.4889
#> 320:    0.4613   1.0091   3.4546   0.4124   0.0721   0.0193   0.4890
#> 321:    0.4608   1.0093   3.4544   0.4129   0.0720   0.0193   0.4889
#> 322:    0.4607   1.0093   3.4543   0.4134   0.0720   0.0193   0.4888
#> 323:    0.4606   1.0093   3.4542   0.4135   0.0721   0.0193   0.4886
#> 324:    0.4604   1.0096   3.4541   0.4139   0.0721   0.0193   0.4884
#> 325:    0.4603   1.0097   3.4540   0.4142   0.0722   0.0193   0.4882
#> 326:    0.4600   1.0098   3.4538   0.4142   0.0722   0.0193   0.4880
#> 327:    0.4602   1.0099   3.4537   0.4149   0.0722   0.0193   0.4879
#> 328:    0.4602   1.0101   3.4536   0.4155   0.0721   0.0193   0.4878
#> 329:    0.4605   1.0102   3.4536   0.4159   0.0721   0.0193   0.4878
#> 330:    0.4602   1.0103   3.4535   0.4159   0.0720   0.0193   0.4878
#> 331:    0.4600   1.0104   3.4535   0.4157   0.0721   0.0193   0.4876
#> 332:    0.4601   1.0105   3.4536   0.4164   0.0721   0.0193   0.4877
#> 333:    0.4600   1.0106   3.4534   0.4172   0.0721   0.0192   0.4876
#> 334:    0.4596   1.0105   3.4533   0.4174   0.0722   0.0193   0.4874
#> 335:    0.4594   1.0107   3.4531   0.4182   0.0722   0.0192   0.4873
#> 336:    0.4587   1.0111   3.4528   0.4185   0.0723   0.0192   0.4874
#> 337:    0.4584   1.0116   3.4526   0.4192   0.0725   0.0193   0.4874
#> 338:    0.4582   1.0118   3.4525   0.4197   0.0726   0.0193   0.4874
#> 339:    0.4579   1.0120   3.4524   0.4200   0.0726   0.0193   0.4873
#> 340:    0.4577   1.0119   3.4523   0.4202   0.0727   0.0193   0.4872
#> 341:    0.4577   1.0120   3.4523   0.4204   0.0728   0.0194   0.4872
#> 342:    0.4574   1.0119   3.4522   0.4200   0.0728   0.0194   0.4873
#> 343:    0.4571   1.0119   3.4521   0.4196   0.0728   0.0193   0.4871
#> 344:    0.4570   1.0119   3.4520   0.4195   0.0728   0.0193   0.4870
#> 345:    0.4568   1.0119   3.4520   0.4193   0.0727   0.0193   0.4870
#> 346:    0.4565   1.0120   3.4519   0.4191   0.0726   0.0193   0.4868
#> 347:    0.4563   1.0120   3.4518   0.4192   0.0726   0.0193   0.4868
#> 348:    0.4560   1.0122   3.4516   0.4193   0.0725   0.0193   0.4869
#> 349:    0.4558   1.0122   3.4516   0.4195   0.0725   0.0193   0.4869
#> 350:    0.4555   1.0122   3.4514   0.4197   0.0725   0.0193   0.4869
#> 351:    0.4552   1.0124   3.4512   0.4198   0.0725   0.0192   0.4870
#> 352:    0.4548   1.0125   3.4512   0.4198   0.0725   0.0193   0.4873
#> 353:    0.4549   1.0127   3.4512   0.4200   0.0726   0.0192   0.4874
#> 354:    0.4549   1.0126   3.4511   0.4202   0.0726   0.0192   0.4874
#> 355:    0.4550   1.0126   3.4511   0.4202   0.0726   0.0192   0.4873
#> 356:    0.4547   1.0124   3.4511   0.4198   0.0727   0.0192   0.4874
#> 357:    0.4548   1.0125   3.4511   0.4196   0.0727   0.0192   0.4875
#> 358:    0.4549   1.0126   3.4512   0.4195   0.0727   0.0191   0.4874
#> 359:    0.4548   1.0127   3.4512   0.4192   0.0728   0.0191   0.4874
#> 360:    0.4545   1.0127   3.4511   0.4192   0.0728   0.0191   0.4873
#> 361:    0.4543   1.0128   3.4510   0.4188   0.0727   0.0190   0.4872
#> 362:    0.4542   1.0128   3.4509   0.4186   0.0727   0.0190   0.4869
#> 363:    0.4541   1.0128   3.4508   0.4186   0.0726   0.0190   0.4866
#> 364:    0.4543   1.0129   3.4508   0.4189   0.0725   0.0191   0.4865
#> 365:    0.4541   1.0129   3.4508   0.4190   0.0724   0.0191   0.4865
#> 366:    0.4538   1.0130   3.4507   0.4185   0.0724   0.0191   0.4866
#> 367:    0.4537   1.0129   3.4508   0.4181   0.0724   0.0191   0.4867
#> 368:    0.4538   1.0129   3.4508   0.4181   0.0723   0.0191   0.4865
#> 369:    0.4540   1.0129   3.4507   0.4184   0.0723   0.0191   0.4865
#> 370:    0.4539   1.0130   3.4507   0.4181   0.0723   0.0191   0.4867
#> 371:    0.4537   1.0132   3.4506   0.4177   0.0723   0.0191   0.4865
#> 372:    0.4534   1.0134   3.4504   0.4174   0.0723   0.0191   0.4865
#> 373:    0.4531   1.0137   3.4503   0.4173   0.0723   0.0191   0.4864
#> 374:    0.4528   1.0138   3.4502   0.4171   0.0723   0.0191   0.4862
#> 375:    0.4525   1.0138   3.4502   0.4169   0.0723   0.0191   0.4862
#> 376:    0.4525   1.0140   3.4502   0.4167   0.0723   0.0191   0.4860
#> 377:    0.4525   1.0140   3.4502   0.4164   0.0724   0.0191   0.4859
#> 378:    0.4523   1.0140   3.4502   0.4164   0.0724   0.0191   0.4860
#> 379:    0.4521   1.0139   3.4501   0.4166   0.0724   0.0190   0.4859
#> 380:    0.4520   1.0139   3.4500   0.4166   0.0724   0.0190   0.4858
#> 381:    0.4518   1.0139   3.4500   0.4166   0.0724   0.0190   0.4857
#> 382:    0.4516   1.0138   3.4500   0.4166   0.0723   0.0190   0.4857
#> 383:    0.4516   1.0137   3.4500   0.4164   0.0723   0.0190   0.4855
#> 384:    0.4514   1.0136   3.4501   0.4159   0.0722   0.0191   0.4855
#> 385:    0.4515   1.0136   3.4501   0.4160   0.0721   0.0191   0.4855
#> 386:    0.4516   1.0137   3.4501   0.4162   0.0722   0.0191   0.4854
#> 387:    0.4518   1.0137   3.4502   0.4160   0.0721   0.0191   0.4854
#> 388:    0.4520   1.0138   3.4503   0.4163   0.0721   0.0192   0.4852
#> 389:    0.4520   1.0136   3.4503   0.4163   0.0721   0.0192   0.4851
#> 390:    0.4518   1.0137   3.4503   0.4164   0.0721   0.0192   0.4850
#> 391:    0.4519   1.0137   3.4503   0.4166   0.0720   0.0192   0.4849
#> 392:    0.4520   1.0137   3.4503   0.4170   0.0720   0.0192   0.4850
#> 393:    0.4520   1.0136   3.4502   0.4174   0.0720   0.0192   0.4850
#> 394:    0.4521   1.0136   3.4501   0.4175   0.0719   0.0192   0.4850
#> 395:    0.4522   1.0137   3.4501   0.4173   0.0718   0.0192   0.4850
#> 396:    0.4523   1.0137   3.4501   0.4174   0.0718   0.0192   0.4850
#> 397:    0.4524   1.0136   3.4502   0.4174   0.0718   0.0192   0.4850
#> 398:    0.4528   1.0135   3.4503   0.4174   0.0718   0.0192   0.4850
#> 399:    0.4531   1.0135   3.4505   0.4174   0.0718   0.0192   0.4850
#> 400:    0.4531   1.0135   3.4506   0.4172   0.0717   0.0192   0.4849
#> 401:    0.4531   1.0135   3.4506   0.4169   0.0716   0.0192   0.4851
#> 402:    0.4531   1.0135   3.4506   0.4167   0.0716   0.0193   0.4852
#> 403:    0.4534   1.0135   3.4506   0.4166   0.0716   0.0193   0.4853
#> 404:    0.4536   1.0135   3.4506   0.4166   0.0717   0.0192   0.4852
#> 405:    0.4536   1.0135   3.4507   0.4164   0.0717   0.0192   0.4852
#> 406:    0.4536   1.0135   3.4507   0.4162   0.0717   0.0192   0.4852
#> 407:    0.4536   1.0135   3.4507   0.4162   0.0717   0.0193   0.4851
#> 408:    0.4538   1.0136   3.4508   0.4161   0.0718   0.0193   0.4851
#> 409:    0.4537   1.0137   3.4507   0.4159   0.0717   0.0193   0.4851
#> 410:    0.4536   1.0137   3.4506   0.4160   0.0717   0.0193   0.4851
#> 411:    0.4535   1.0136   3.4505   0.4158   0.0717   0.0193   0.4851
#> 412:    0.4534   1.0138   3.4505   0.4154   0.0717   0.0193   0.4852
#> 413:    0.4533   1.0139   3.4504   0.4150   0.0716   0.0193   0.4852
#> 414:    0.4531   1.0140   3.4504   0.4152   0.0716   0.0193   0.4851
#> 415:    0.4531   1.0140   3.4504   0.4152   0.0715   0.0194   0.4850
#> 416:    0.4531   1.0141   3.4504   0.4151   0.0715   0.0194   0.4850
#> 417:    0.4532   1.0141   3.4504   0.4153   0.0715   0.0194   0.4848
#> 418:    0.4531   1.0141   3.4504   0.4151   0.0715   0.0194   0.4847
#> 419:    0.4532   1.0142   3.4504   0.4152   0.0715   0.0194   0.4846
#> 420:    0.4534   1.0141   3.4505   0.4151   0.0715   0.0194   0.4844
#> 421:    0.4533   1.0141   3.4505   0.4149   0.0716   0.0194   0.4843
#> 422:    0.4534   1.0140   3.4505   0.4150   0.0715   0.0193   0.4843
#> 423:    0.4534   1.0139   3.4505   0.4149   0.0716   0.0193   0.4842
#> 424:    0.4532   1.0139   3.4504   0.4149   0.0716   0.0193   0.4842
#> 425:    0.4532   1.0140   3.4503   0.4151   0.0716   0.0193   0.4840
#> 426:    0.4532   1.0141   3.4503   0.4152   0.0716   0.0194   0.4840
#> 427:    0.4530   1.0142   3.4502   0.4153   0.0716   0.0193   0.4839
#> 428:    0.4528   1.0143   3.4502   0.4152   0.0716   0.0194   0.4838
#> 429:    0.4528   1.0143   3.4502   0.4152   0.0716   0.0194   0.4838
#> 430:    0.4526   1.0144   3.4502   0.4150   0.0716   0.0194   0.4839
#> 431:    0.4525   1.0144   3.4502   0.4147   0.0716   0.0193   0.4839
#> 432:    0.4524   1.0145   3.4502   0.4145   0.0717   0.0193   0.4838
#> 433:    0.4524   1.0145   3.4502   0.4145   0.0717   0.0193   0.4839
#> 434:    0.4524   1.0144   3.4502   0.4146   0.0718   0.0193   0.4838
#> 435:    0.4524   1.0144   3.4502   0.4145   0.0718   0.0193   0.4838
#> 436:    0.4524   1.0144   3.4502   0.4146   0.0718   0.0193   0.4838
#> 437:    0.4525   1.0144   3.4502   0.4147   0.0718   0.0193   0.4839
#> 438:    0.4524   1.0145   3.4502   0.4146   0.0719   0.0193   0.4838
#> 439:    0.4525   1.0146   3.4502   0.4148   0.0719   0.0193   0.4837
#> 440:    0.4524   1.0146   3.4502   0.4146   0.0719   0.0192   0.4837
#> 441:    0.4526   1.0146   3.4502   0.4147   0.0718   0.0192   0.4836
#> 442:    0.4525   1.0147   3.4502   0.4146   0.0718   0.0192   0.4835
#> 443:    0.4526   1.0148   3.4502   0.4148   0.0718   0.0192   0.4835
#> 444:    0.4525   1.0149   3.4501   0.4149   0.0719   0.0192   0.4835
#> 445:    0.4525   1.0149   3.4500   0.4152   0.0719   0.0191   0.4835
#> 446:    0.4523   1.0150   3.4499   0.4153   0.0720   0.0191   0.4835
#> 447:    0.4524   1.0150   3.4499   0.4153   0.0720   0.0191   0.4835
#> 448:    0.4523   1.0150   3.4500   0.4149   0.0719   0.0191   0.4835
#> 449:    0.4523   1.0151   3.4501   0.4145   0.0719   0.0191   0.4835
#> 450:    0.4524   1.0150   3.4502   0.4141   0.0718   0.0191   0.4835
#> 451:    0.4525   1.0150   3.4502   0.4141   0.0718   0.0192   0.4835
#> 452:    0.4527   1.0150   3.4503   0.4141   0.0717   0.0192   0.4834
#> 453:    0.4527   1.0150   3.4502   0.4138   0.0717   0.0192   0.4833
#> 454:    0.4526   1.0151   3.4502   0.4137   0.0717   0.0191   0.4832
#> 455:    0.4527   1.0152   3.4502   0.4139   0.0716   0.0191   0.4831
#> 456:    0.4527   1.0152   3.4502   0.4141   0.0716   0.0191   0.4830
#> 457:    0.4527   1.0151   3.4502   0.4143   0.0716   0.0191   0.4829
#> 458:    0.4528   1.0152   3.4502   0.4145   0.0715   0.0192   0.4828
#> 459:    0.4526   1.0153   3.4502   0.4144   0.0715   0.0191   0.4828
#> 460:    0.4524   1.0153   3.4501   0.4142   0.0714   0.0192   0.4828
#> 461:    0.4523   1.0154   3.4501   0.4140   0.0714   0.0192   0.4829
#> 462:    0.4521   1.0156   3.4501   0.4139   0.0713   0.0192   0.4828
#> 463:    0.4522   1.0156   3.4501   0.4137   0.0713   0.0192   0.4829
#> 464:    0.4522   1.0157   3.4500   0.4136   0.0713   0.0191   0.4829
#> 465:    0.4522   1.0156   3.4501   0.4134   0.0713   0.0191   0.4828
#> 466:    0.4522   1.0156   3.4500   0.4131   0.0713   0.0191   0.4828
#> 467:    0.4520   1.0156   3.4501   0.4128   0.0713   0.0191   0.4828
#> 468:    0.4520   1.0156   3.4501   0.4126   0.0713   0.0191   0.4828
#> 469:    0.4521   1.0158   3.4501   0.4128   0.0713   0.0191   0.4828
#> 470:    0.4520   1.0158   3.4501   0.4129   0.0714   0.0191   0.4828
#> 471:    0.4520   1.0159   3.4501   0.4127   0.0715   0.0191   0.4830
#> 472:    0.4520   1.0159   3.4501   0.4127   0.0715   0.0191   0.4831
#> 473:    0.4519   1.0159   3.4501   0.4125   0.0715   0.0190   0.4831
#> 474:    0.4519   1.0159   3.4501   0.4124   0.0714   0.0190   0.4832
#> 475:    0.4520   1.0160   3.4501   0.4123   0.0714   0.0190   0.4832
#> 476:    0.4521   1.0160   3.4501   0.4124   0.0714   0.0190   0.4832
#> 477:    0.4520   1.0159   3.4501   0.4123   0.0714   0.0190   0.4832
#> 478:    0.4520   1.0159   3.4502   0.4122   0.0714   0.0190   0.4832
#> 479:    0.4521   1.0159   3.4501   0.4124   0.0714   0.0190   0.4833
#> 480:    0.4520   1.0159   3.4501   0.4122   0.0714   0.0190   0.4834
#> 481:    0.4520   1.0160   3.4501   0.4119   0.0714   0.0190   0.4835
#> 482:    0.4520   1.0160   3.4501   0.4119   0.0714   0.0190   0.4835
#> 483:    0.4520   1.0160   3.4501   0.4121   0.0714   0.0190   0.4835
#> 484:    0.4520   1.0159   3.4501   0.4123   0.0714   0.0190   0.4834
#> 485:    0.4521   1.0160   3.4501   0.4125   0.0713   0.0190   0.4834
#> 486:    0.4521   1.0160   3.4502   0.4123   0.0714   0.0190   0.4833
#> 487:    0.4523   1.0160   3.4502   0.4123   0.0714   0.0190   0.4833
#> 488:    0.4523   1.0160   3.4502   0.4121   0.0714   0.0190   0.4833
#> 489:    0.4526   1.0160   3.4503   0.4121   0.0714   0.0190   0.4832
#> 490:    0.4528   1.0160   3.4503   0.4124   0.0714   0.0190   0.4832
#> 491:    0.4529   1.0160   3.4504   0.4122   0.0714   0.0190   0.4831
#> 492:    0.4529   1.0161   3.4505   0.4120   0.0714   0.0190   0.4832
#> 493:    0.4529   1.0160   3.4505   0.4118   0.0714   0.0190   0.4832
#> 494:    0.4530   1.0160   3.4505   0.4116   0.0714   0.0190   0.4832
#> 495:    0.4531   1.0160   3.4505   0.4116   0.0713   0.0190   0.4831
#> 496:    0.4532   1.0160   3.4505   0.4116   0.0714   0.0191   0.4831
#> 497:    0.4532   1.0160   3.4505   0.4118   0.0713   0.0191   0.4830
#> 498:    0.4532   1.0160   3.4505   0.4119   0.0713   0.0191   0.4830
#> 499:    0.4533   1.0160   3.4505   0.4121   0.0713   0.0191   0.4828
#> 500:    0.4534   1.0160   3.4504   0.4122   0.0713   0.0191   0.4828
#> Calculating covariance matrix
#> 
#> Warning: Bounds are ignored in SAEM

print(f)
#> ── nlmixr SAEM(Solved); OBJF not calculated fit ────────────────────────────────
#> 
#>  Gaussian/Laplacian Likelihoods: AIC() or $objf etc. 
#>  FOCEi CWRES & Likelihoods: addCwres() 
#> 
#> ── Time (sec $time): ───────────────────────────────────────────────────────────
#> 
#>          saem    .time covariance    other
#> elapsed 0.925 0.288553      0.008 0.445447
#> 
#> ── Population Parameters ($parFixed or $parFixedDf): ───────────────────────────
#> 
#>        Parameter  Est.     SE %RSE Back-transformed(95%CI) BSV(CV%) Shrink(SD)%
#> tka       Log Ka 0.453  0.195 43.1       1.57 (1.07, 2.31)     71.4    -0.445% 
#> tcl       Log Cl  1.02 0.0843 8.29       2.76 (2.34, 3.26)     27.2      3.88% 
#> tv         log V  3.45 0.0467 1.35       31.5 (28.8, 34.5)     13.9      10.2% 
#> add.sd           0.695                               0.695                     
#>  
#>   Covariance Type ($covMethod): linFim
#>   Fixed parameter correlations in $cor
#>   No correlations in between subject variability (BSV) matrix
#>   Full BSV covariance ($omega) or correlation ($omegaR; diagonals=SDs) 
#>   Distribution stats (mean/skewness/kurtosis/p-value) available in $shrink 

# Now add the tables

f <- addTable(f)
#> Calculating residuals/tables
#> done

print(f)
#> ── nlmixr SAEM(Solved); OBJF not calculated fit ────────────────────────────────
#> 
#>  Gaussian/Laplacian Likelihoods: AIC() or $objf etc. 
#>  FOCEi CWRES & Likelihoods: addCwres() 
#> 
#> ── Time (sec $time): ───────────────────────────────────────────────────────────
#> 
#>          saem    .time covariance    other table
#> elapsed 0.925 0.288553      0.008 0.445447  0.01
#> 
#> ── Population Parameters ($parFixed or $parFixedDf): ───────────────────────────
#> 
#>        Parameter  Est.     SE %RSE Back-transformed(95%CI) BSV(CV%) Shrink(SD)%
#> tka       Log Ka 0.453  0.195 43.1       1.57 (1.07, 2.31)     71.4    -0.445% 
#> tcl       Log Cl  1.02 0.0843 8.29       2.76 (2.34, 3.26)     27.2      3.88% 
#> tv         log V  3.45 0.0467 1.35       31.5 (28.8, 34.5)     13.9      10.2% 
#> add.sd           0.695                               0.695                     
#>  
#>   Covariance Type ($covMethod): linFim
#>   Fixed parameter correlations in $cor
#>   No correlations in between subject variability (BSV) matrix
#>   Full BSV covariance ($omega) or correlation ($omegaR; diagonals=SDs) 
#>   Distribution stats (mean/skewness/kurtosis/p-value) available in $shrink 
#> 
#> ── Fit Data (object is a modified tibble): ─────────────────────────────────────
#> # A tibble: 132 × 16
#>   ID     TIME    DV  PRED    RES IPRED   IRES  IWRES eta.ka eta.cl   eta.v    ka
#>   <fct> <dbl> <dbl> <dbl>  <dbl> <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl> <dbl>
#> 1 1      0     0.74  0     0.74   0     0.74   1.07   0.107 -0.485 -0.0809  1.75
#> 2 1      0.25  2.84  3.26 -0.424  3.87 -1.03  -1.49   0.107 -0.485 -0.0809  1.75
#> 3 1      0.57  6.57  5.84  0.726  6.82 -0.250 -0.360  0.107 -0.485 -0.0809  1.75
#> # … with 129 more rows, and 4 more variables: cl <dbl>, v <dbl>, tad <dbl>,
#> #   dosenum <dbl>

# }