Change a nlmixr fit object to a huxtable

This is a thin layer that differs from huxreg to make it easier to produce reasonable huxtables with nlmixr fit objects.

asHux.nlmixrFitCore(
  ...,
  error_format = "({std.error})",
  error_style = c("stderr", "ci", "statistic", "pvalue"),
  error_pos = c("below", "same", "right"),
  number_format = "%.3f",
  align = ".",
  pad_decimal = ".",
  ci_level = NULL,
  tidy_args = NULL,
  stars = c(`***` = 0.001, `**` = 0.01, `*` = 0.05),
  bold_signif = NULL,
  borders = 0.4,
  outer_borders = 0.8,
  note = if (is.null(stars)) NULL else "{stars}.",
  statistics = c(N = "nobs", `Objective Function` = "OBJF", "logLik", "AIC"),
  coefs = NULL,
  omit_coefs = NULL,
  na_omit = c("all", "any", "none")
)

Arguments

...
Models, or a single list of models. Names will be used as column headings.

error_format
How to display uncertainty in estimates. See below.

error_style
Deprecated. One or more of "stderr", "ci" (confidence interval), "statistic" or "pvalue".

error_pos
Display uncertainty "below", to the "right" of, or in the "same" cell as estimates.

number_format
Format for numbering. See number_format() for details.

align
Alignment for table cells. Set to a single character to align on this character.

pad_decimal
Deprecated in favour of align.

ci_level
Confidence level for intervals. Set to NULL to not calculate confidence intervals.

tidy_args
List of arguments to pass to broom::tidy(). You can also pass a list of lists; if so, the nth element will be used for the nth column.

stars
Levels for p value stars. Names of stars are symbols to use. Set to NULL to not show stars.

bold_signif
Where p values are below this number, cells will be displayed in bold. Use NULL to turn off this behaviour.

borders
Thickness of inner horizontal borders. Set to 0 for no borders.

outer_borders
Thickness of outer (top and bottom) horizontal borders. Set to 0 for no borders.

note
Footnote for bottom cell, which spans all columns. {stars} will be replaced by a note about significance stars. Set to NULL for no footnote.

statistics
A vector of summary statistics to display. Set to NULL to show all available statistics. To change display names, name the statistics vector: c("Displayed title" = "statistic_name", ...)

coefs
A vector of coefficients to display. Overrules omit_coefs. To change display names, name the coef vector: c("Displayed title" = "coefficient_name", ...)

omit_coefs
Omit these coefficients.

na_omit
How NAs are handled in converting to huxtable

Details

Drops NA values from tables, so (NA) is not shown for estimates without a standard error.

Filters out any blank rows.

Removes R.squared from the statistics and replaces with the objective function value.

Adjust broom separators for an easier to read format.