Bootstrap data

Bootstrap data by sampling the same number of subjects from the original dataset by sampling with replacement.

bootdata(dat)

Arguments

dat
model data to be bootstrapped

Value

Bootstrapped data

Examples

specs <- list(fixed=lKA+lCL+lV~1, random = pdDiag(lKA+lCL~1), start=c(lKA=0.5, lCL=-3.2, lV=-1))
set.seed(99); nboot = 5;

cat("generating", nboot, "bootstrap samples...\n")

#> generating 5 bootstrap samples...

cmat <- matrix(NA, nboot, 3)
for (i in 1:nboot)
{
        #print(i)
        bd <- bootdata(theo_md)
        fit <- nlme_lin_cmpt(bd, par_model=specs, ncmt=1)
        cmat[i,] = fit$coefficients$fixed
}
dimnames(cmat)[[2]] <- names(fit$coefficients$fixed)
print(head(cmat))

#>             lKA       lCL       lV
#> [1,] 0.13698361 1.0644287 3.373275
#> [2,] 0.39809058 1.0499153 3.420109
#> [3,] 0.02314097 0.9161608 3.337754
#> [4,] 0.08522340 1.0965773 3.348626
#> [5,] 0.15967969 1.0309578 3.350310