Configure an SAEM model

Configure an SAEM model by generating an input list to the SAEM model function

configsaem(
  model,
  data,
  inits,
  mcmc = list(niter = c(200, 300), nmc = 3, nu = c(2, 2, 2)),
  ODEopt = list(atol = 1e-06, rtol = 1e-04, method = "lsoda", transitAbs = FALSE,
    maxeval = 1e+05),
  distribution = c("normal", "poisson", "binomial"),
  addProp = c("combined2", "combined1"),
  seed = 99,
  fixed = NULL,
  DEBUG = 0,
  tol = 1e-04,
  itmax = 100L,
  type = c("nelder-mead", "newuoa"),
  lambdaRange = 3,
  powRange = 10,
  odeRecalcFactor = 10^(0.5),
  maxOdeRecalc = 5L
)

Arguments

model

a compiled saem model by gen_saem_user_fn()

data

input data

inits

initial values

mcmc

a list of various mcmc options

ODEopt

optional ODE solving options

distribution

one of c("normal","poisson","binomial")

addProp

one of "combined1" and "combined2"; These are the two forms of additive+proportional errors supported by monolix/nonmem:

combined1: transform(y)=transform(f)+(a+b*f^c)*eps

combined2: transform(y)=transform(f)+(a^2+b^2*f^(2c))*eps

seed

seed for random number generator

fixed

a character vector of fixed effect only parameters (no random effects attached) to be fixed

DEBUG

Integer determining if debugging is enabled

tol

This is the tolerance for the regression models used for complex residual errors (ie add+prop etc)

itmax

This is the maximum number of iterations for the regression models used for complex residual errors. The number of iterations is itmax*number of parameters

type

indicates the type of optimization for the residuals; Can be one of c("nelder-mead", "newuoa")

lambdaRange

This indicates the range that Box-Cox and Yeo-Johnson parameters are constrained to be; The default is 3 indicating the range (-3,3)

powRange

This indicates the range that powers can take for residual errors; By default this is 10 indicating the range is c(1/10, 10) or c(0.1,10)

odeRecalcFactor

The factor to increase the rtol/atol with bad ODE solving.

maxOdeRecalc

Maximum number of times to reduce the ODE tolerances and try to resolve the system if there was a bad ODE solve.

Value

Returns a list neede for the saem fit procedure

Details

Fit a generalized nonlinear mixed-effect model by he Stochastic Approximation Expectation-Maximization (SAEM) algorithm

Author

Wenping Wang & Matthew Fidler

Examples

# \donttest{

# In this ODE system we simply specify the ODEs

ode <- "d/dt(depot) =-KA*depot;
        d/dt(centr) = KA*depot - KE*centr;"
m1 <- RxODE(ode)
#>  

# In this ode System, we also specify the concentration as C2 = centr/V

ode <- "C2 = centr/V;
      d/dt(depot) =-KA*depot;
      d/dt(centr) = KA*depot - KE*centr;"
m2 = RxODE(ode)
#>  

PKpars <- function() {
  CL <- exp(lCL)
  V <- exp(lV)
  KA <- exp(lKA)
  KE <- CL / V
}

PRED <- function() centr / V
PRED2 <- function() C2

saem_fit <- gen_saem_user_fn(model = m1, PKpars, pred = PRED)
#>  
#> Warning: this was not generated from nlmixr, loading into symengine
#> → creating SAEM model...
#> → pruning branches (`if`/`else`)...
#> ✔ done
#> → loading into symengine environment...
#> ✔ done

# Can also use PRED2
saem_fit <- gen_saem_user_fn(model=m2, PKpars, pred=PRED2)
#>  
#> Warning: this was not generated from nlmixr, loading into symengine
#> → creating SAEM model...
#> → pruning branches (`if`/`else`)...
#> ✔ done
#> → loading into symengine environment...
#> ✔ done

# You can also use the nlmixr UI to run this model and call the lower level functions

one.compartment <- function() {
ini({
  tka <- 0.45 # Log Ka
  tcl <- 1 # Log Cl
  tv <- 3.45    # Log V
  eta.ka ~ 0.6
  eta.cl ~ 0.3
  eta.v ~ 0.1
  add.sd <- 0.7
  wt.est <- 0.0
})
model({
  ka <- exp(tka + eta.ka)
  cl <- exp(tcl + eta.cl)
  v <- exp(tv + eta.v + wt.est * WT)
  d/dt(depot) = -ka * depot
  d/dt(center) = ka * depot - cl / v * center
  cp = center / v
  cp ~ add(add.sd)
})
}
fit  <- nlmixr(one.compartment, theo_sd, "saem")
#> ℹ parameter labels from comments will be replaced by 'label()'
#>  
#> → generate SAEM model
#> ✔ done
#> 1:    0.2288   0.9677   3.0859   0.0056   0.5700   0.2850   0.0950   4.5245
#> 2:    0.3511   0.9707   3.2955   0.0034   0.5415   0.2707   0.0902   1.8178
#> 3:    0.5238   0.9353   3.2970   0.0027   0.5144   0.2572   0.0857   0.9123
#> 4:    0.4607   0.9651   3.2319   0.0036   0.4887   0.2444   0.0815   0.7016
#> 5:    0.4891   0.9815   3.0947   0.0054   0.4643   0.2321   0.0774   0.6063
#> 6:    0.5061   0.9684   3.0698   0.0058   0.4411   0.2205   0.0735   0.5068
#> 7:    0.5023   0.9883   2.9614   0.0074   0.4414   0.2095   0.0698   0.5062
#> 8:    0.4468   0.9980   2.9028   0.0080   0.4193   0.1990   0.0663   0.5032
#> 9:    0.4153   0.9773   2.9886   0.0067   0.3983   0.1891   0.0630   0.5190
#> 10:    0.4430   0.9780   2.7978   0.0094   0.4423   0.1796   0.0599   0.4963
#> 11:    0.4818   0.9891   2.7850   0.0097   0.4328   0.1706   0.0569   0.4746
#> 12:    0.4665   1.0258   2.8853   0.0082   0.4111   0.1621   0.0540   0.4743
#> 13:    0.4232   1.0326   2.8661   0.0082   0.3906   0.1540   0.0513   0.5009
#> 14:    0.4645   1.0229   2.7932   0.0095   0.4726   0.1463   0.0488   0.4809
#> 15:    0.5062   1.0085   2.8418   0.0090   0.4489   0.1390   0.0463   0.4652
#> 16:    0.5111   1.0073   2.8482   0.0091   0.4265   0.1320   0.0440   0.4645
#> 17:    0.5158   1.0003   2.8472   0.0091   0.4344   0.1254   0.0418   0.4466
#> 18:    0.5092   0.9948   2.8562   0.0089   0.4127   0.1192   0.0397   0.4793
#> 19:    0.4526   0.9930   2.8934   0.0081   0.3974   0.1132   0.0377   0.4742
#> 20:    0.4593   1.0190   2.9236   0.0076   0.4383   0.1075   0.0358   0.4910
#> 21:    0.4705   0.9978   2.9321   0.0077   0.4420   0.1022   0.0341   0.4647
#> 22:    0.4778   1.0054   2.9304   0.0077   0.4458   0.0971   0.0324   0.4556
#> 23:    0.5338   1.0089   2.9851   0.0070   0.5067   0.0922   0.0307   0.4845
#> 24:    0.4738   1.0230   2.9628   0.0071   0.4814   0.0876   0.0292   0.4673
#> 25:    0.4489   1.0101   2.8713   0.0083   0.4573   0.0915   0.0277   0.4668
#> 26:    0.4073   1.0336   2.8700   0.0080   0.4344   0.0869   0.0264   0.4755
#> 27:    0.3808   1.0525   2.8990   0.0076   0.4127   0.0825   0.0250   0.4971
#> 28:    0.4344   1.0355   2.9030   0.0078   0.3980   0.0784   0.0238   0.4465
#> 29:    0.4449   1.0333   2.9591   0.0068   0.3781   0.0745   0.0226   0.4823
#> 30:    0.4594   1.0127   2.9503   0.0072   0.3956   0.0708   0.0215   0.4810
#> 31:    0.4110   1.0045   2.9568   0.0071   0.3758   0.0724   0.0204   0.4825
#> 32:    0.4334   0.9960   2.9692   0.0071   0.4048   0.0688   0.0194   0.4654
#> 33:    0.4543   1.0036   2.9884   0.0068   0.4051   0.0654   0.0184   0.4750
#> 34:    0.4781   1.0037   2.9265   0.0076   0.3849   0.0700   0.0175   0.4857
#> 35:    0.4588   1.0011   2.9969   0.0066   0.3656   0.0716   0.0166   0.4759
#> 36:    0.4455   1.0307   3.0652   0.0055   0.3474   0.0773   0.0158   0.4665
#> 37:    0.4349   1.0312   2.9285   0.0076   0.3354   0.0839   0.0150   0.4916
#> 38:    0.4869   1.0158   2.9028   0.0078   0.4736   0.0797   0.0142   0.4837
#> 39:    0.4968   0.9977   2.8434   0.0090   0.4499   0.0815   0.0140   0.4578
#> 40:    0.5032   0.9914   2.7436   0.0106   0.4274   0.0774   0.0134   0.4765
#> 41:    0.4678   0.9660   2.8057   0.0096   0.4060   0.0735   0.0132   0.4977
#> 42:    0.4708   0.9883   2.8399   0.0090   0.4092   0.0701   0.0126   0.4429
#> 43:    0.4380   1.0116   2.8828   0.0081   0.3887   0.0768   0.0128   0.4403
#> 44:    0.4164   1.0246   2.8060   0.0091   0.4065   0.0730   0.0126   0.4442
#> 45:    0.3910   1.0598   2.7412   0.0098   0.3862   0.0781   0.0119   0.4632
#> 46:    0.3975   1.0336   2.9049   0.0076   0.3669   0.0865   0.0113   0.4583
#> 47:    0.4134   1.0289   2.9616   0.0067   0.3818   0.0858   0.0108   0.4664
#> 48:    0.4256   1.0193   2.8668   0.0081   0.4218   0.0815   0.0106   0.4875
#> 49:    0.4285   1.0266   2.9383   0.0073   0.4007   0.0853   0.0101   0.4973
#> 50:    0.4991   0.9878   2.9406   0.0073   0.4039   0.0929   0.0096   0.5091
#> 51:    0.4606   0.9888   2.9466   0.0072   0.4161   0.0917   0.0094   0.4993
#> 52:    0.5027   1.0133   3.0064   0.0065   0.4647   0.0871   0.0089   0.4759
#> 53:    0.4773   1.0112   2.9368   0.0075   0.4500   0.0827   0.0085   0.4720
#> 54:    0.4237   0.9982   2.9494   0.0072   0.4275   0.0830   0.0080   0.4867
#> 55:    0.4425   1.0135   2.9529   0.0071   0.4061   0.0789   0.0077   0.5019
#> 56:    0.4366   0.9921   2.9077   0.0077   0.3858   0.0749   0.0086   0.4981
#> 57:    0.4427   0.9931   2.8733   0.0083   0.3991   0.0725   0.0086   0.5088
#> 58:    0.4913   0.9949   2.9670   0.0069   0.4044   0.0791   0.0081   0.4903
#> 59:    0.4631   0.9812   3.0400   0.0058   0.4732   0.1079   0.0077   0.5167
#> 60:    0.4629   0.9885   3.0546   0.0057   0.4875   0.1025   0.0099   0.5172
#> 61:    0.5446   0.9957   3.0156   0.0064   0.5543   0.0974   0.0116   0.5103
#> 62:    0.4900   1.0253   2.9865   0.0069   0.5266   0.1013   0.0130   0.5280
#> 63:    0.4927   0.9962   3.0938   0.0053   0.5003   0.1023   0.0152   0.5047
#> 64:    0.4871   1.0171   3.0618   0.0057   0.4753   0.0972   0.0144   0.4929
#> 65:    0.4838   1.0020   3.0185   0.0062   0.4644   0.0923   0.0137   0.4628
#> 66:    0.4921   1.0017   3.0340   0.0062   0.4412   0.0877   0.0135   0.4587
#> 67:    0.4716   0.9783   3.0422   0.0060   0.4796   0.0833   0.0151   0.4604
#> 68:    0.4366   0.9940   3.0199   0.0064   0.4556   0.0791   0.0143   0.5108
#> 69:    0.4385   0.9982   2.9486   0.0072   0.4328   0.0752   0.0136   0.5002
#> 70:    0.4221   1.0189   2.9937   0.0066   0.4860   0.0757   0.0129   0.5009
#> 71:    0.4016   1.0347   2.9678   0.0065   0.4617   0.0758   0.0142   0.5264
#> 72:    0.4151   0.9912   3.0228   0.0060   0.4386   0.0824   0.0135   0.5032
#> 73:    0.4749   1.0136   3.0081   0.0066   0.4167   0.0783   0.0128   0.5045
#> 74:    0.4607   0.9816   2.9735   0.0071   0.3959   0.0743   0.0121   0.4821
#> 75:    0.4948   0.9853   2.9088   0.0083   0.4011   0.0833   0.0122   0.4848
#> 76:    0.5378   0.9899   2.9698   0.0076   0.3811   0.0792   0.0122   0.5201
#> 77:    0.5506   0.9519   2.8874   0.0086   0.3806   0.0752   0.0131   0.5408
#> 78:    0.4911   0.9855   2.7746   0.0101   0.4157   0.0795   0.0144   0.5161
#> 79:    0.5034   1.0018   2.7222   0.0109   0.4490   0.0755   0.0137   0.5172
#> 80:    0.5105   0.9842   2.7954   0.0096   0.4817   0.0755   0.0130   0.5032
#> 81:    0.5110   0.9991   2.8877   0.0083   0.4607   0.0717   0.0141   0.5045
#> 82:    0.4897   0.9758   2.8211   0.0094   0.4761   0.0682   0.0163   0.4973
#> 83:    0.4645   0.9945   2.8349   0.0091   0.4523   0.0685   0.0167   0.4678
#> 84:    0.4845   0.9784   2.9136   0.0079   0.4509   0.0661   0.0180   0.4708
#> 85:    0.4779   1.0076   2.9204   0.0079   0.4445   0.0642   0.0171   0.4917
#> 86:    0.4861   0.9979   2.8459   0.0090   0.4222   0.0753   0.0169   0.4880
#> 87:    0.5066   0.9855   2.9493   0.0075   0.4011   0.0743   0.0160   0.4859
#> 88:    0.4139   0.9689   2.8814   0.0084   0.3964   0.0729   0.0152   0.5056
#> 89:    0.4880   0.9880   3.0100   0.0064   0.3976   0.0715   0.0146   0.4985
#> 90:    0.4508   1.0008   2.9872   0.0066   0.4187   0.0699   0.0154   0.5140
#> 91:    0.4260   1.0124   2.9353   0.0070   0.4915   0.0721   0.0170   0.4828
#> 92:    0.4507   1.0310   2.9768   0.0065   0.5508   0.0821   0.0186   0.4849
#> 93:    0.4156   1.0215   2.9663   0.0067   0.5232   0.0818   0.0176   0.4967
#> 94:    0.4253   1.0309   2.9672   0.0067   0.4971   0.0927   0.0168   0.4940
#> 95:    0.4293   1.0402   2.9557   0.0069   0.5670   0.0880   0.0159   0.5130
#> 96:    0.4091   1.0215   2.7608   0.0095   0.6011   0.0836   0.0179   0.4914
#> 97:    0.4148   1.0283   2.7167   0.0103   0.5710   0.0795   0.0182   0.4873
#> 98:    0.4568   1.0388   2.7550   0.0099   0.5425   0.0755   0.0173   0.4683
#> 99:    0.4516   1.0081   2.8131   0.0090   0.5154   0.0717   0.0164   0.4685
#> 100:    0.4161   1.0146   2.7195   0.0104   0.4896   0.0681   0.0156   0.4648
#> 101:    0.4124   0.9984   2.7309   0.0100   0.4862   0.0774   0.0148   0.4674
#> 102:    0.4114   1.0021   2.7699   0.0094   0.4965   0.0735   0.0141   0.4836
#> 103:    0.4631   1.0080   2.7506   0.0099   0.5232   0.0699   0.0144   0.4663
#> 104:    0.4380   0.9986   2.8772   0.0083   0.5338   0.0694   0.0141   0.5044
#> 105:    0.4544   1.0045   2.8444   0.0085   0.5273   0.0803   0.0134   0.5151
#> 106:    0.4545   0.9904   2.8583   0.0086   0.5009   0.0840   0.0128   0.4922
#> 107:    0.4684   1.0089   2.8140   0.0091   0.5415   0.0892   0.0121   0.4703
#> 108:    0.4306   0.9987   2.9950   0.0065   0.5144   0.0847   0.0118   0.4906
#> 109:    0.4180   1.0095   2.9333   0.0075   0.4887   0.0968   0.0114   0.4718
#> 110:    0.4760   1.0181   2.9466   0.0071   0.4643   0.0919   0.0117   0.4844
#> 111:    0.4700   1.0164   2.9324   0.0075   0.4411   0.0976   0.0113   0.4680
#> 112:    0.4558   1.0036   2.9026   0.0077   0.4766   0.0927   0.0108   0.4614
#> 113:    0.4388   1.0328   2.8671   0.0083   0.5079   0.0881   0.0102   0.4698
#> 114:    0.4567   1.0134   2.9205   0.0075   0.4825   0.0971   0.0098   0.4603
#> 115:    0.4423   1.0116   2.9967   0.0064   0.4583   0.1005   0.0095   0.4718
#> 116:    0.4732   1.0196   2.9908   0.0066   0.4465   0.0955   0.0093   0.4558
#> 117:    0.4275   1.0510   2.9419   0.0072   0.4242   0.0907   0.0089   0.4711
#> 118:    0.4356   1.0201   2.8569   0.0084   0.4142   0.0894   0.0090   0.4577
#> 119:    0.4511   1.0283   2.8167   0.0090   0.4283   0.0931   0.0112   0.4782
#> 120:    0.4049   1.0045   2.8063   0.0092   0.4234   0.0885   0.0107   0.4742
#> 121:    0.4341   1.0186   2.8541   0.0084   0.4063   0.0840   0.0111   0.4768
#> 122:    0.4254   0.9997   2.7958   0.0094   0.4185   0.0798   0.0141   0.4655
#> 123:    0.4609   1.0209   2.8832   0.0081   0.4395   0.0758   0.0141   0.4529
#> 124:    0.4608   1.0518   2.8655   0.0082   0.4666   0.0720   0.0134   0.4620
#> 125:    0.4488   1.0380   2.9991   0.0062   0.4433   0.0787   0.0127   0.4632
#> 126:    0.4332   1.0575   2.8804   0.0078   0.4494   0.0855   0.0121   0.4689
#> 127:    0.4617   1.0350   2.8689   0.0080   0.4269   0.0942   0.0115   0.4848
#> 128:    0.3910   1.0242   2.8530   0.0083   0.4511   0.0895   0.0133   0.4976
#> 129:    0.4243   1.0408   2.8930   0.0078   0.4872   0.0891   0.0127   0.4864
#> 130:    0.4753   1.0640   2.8764   0.0080   0.5144   0.0846   0.0120   0.4768
#> 131:    0.4224   1.0240   2.9354   0.0070   0.4886   0.0816   0.0114   0.4668
#> 132:    0.3935   1.0262   2.9957   0.0061   0.4642   0.0886   0.0108   0.5029
#> 133:    0.4031   1.0150   3.0069   0.0060   0.4410   0.0842   0.0103   0.5101
#> 134:    0.4223   0.9989   3.0363   0.0059   0.4189   0.0800   0.0131   0.4907
#> 135:    0.4447   1.0099   3.0033   0.0065   0.3980   0.0760   0.0125   0.4672
#> 136:    0.4646   1.0001   2.9817   0.0068   0.3854   0.0722   0.0118   0.4583
#> 137:    0.4921   0.9963   3.0924   0.0054   0.4607   0.0686   0.0134   0.4840
#> 138:    0.4692   1.0106   3.0587   0.0058   0.4377   0.0651   0.0131   0.4779
#> 139:    0.4420   1.0132   3.0325   0.0062   0.4158   0.0675   0.0140   0.4725
#> 140:    0.4556   1.0254   2.9416   0.0075   0.4146   0.0690   0.0133   0.4819
#> 141:    0.4540   0.9972   2.9206   0.0078   0.4090   0.0714   0.0127   0.4595
#> 142:    0.5174   0.9817   2.9641   0.0073   0.4369   0.0819   0.0120   0.4643
#> 143:    0.5016   1.0166   2.8976   0.0082   0.4396   0.0778   0.0124   0.4634
#> 144:    0.5336   1.0114   2.9333   0.0078   0.4176   0.0740   0.0118   0.4726
#> 145:    0.4565   1.0076   2.8688   0.0083   0.3968   0.0802   0.0116   0.4578
#> 146:    0.4438   1.0191   2.9896   0.0067   0.4003   0.0762   0.0125   0.4758
#> 147:    0.4326   0.9820   2.8972   0.0079   0.3803   0.0729   0.0119   0.4618
#> 148:    0.4267   1.0131   2.8369   0.0088   0.3648   0.0693   0.0113   0.4653
#> 149:    0.4456   1.0167   2.9018   0.0079   0.3466   0.0727   0.0121   0.4622
#> 150:    0.4858   0.9967   2.8976   0.0080   0.3817   0.0715   0.0115   0.4682
#> 151:    0.4110   1.0269   2.8744   0.0082   0.3701   0.0679   0.0117   0.4809
#> 152:    0.4328   1.0276   2.8772   0.0082   0.3669   0.0755   0.0112   0.4811
#> 153:    0.4322   1.0284   2.8966   0.0079   0.3996   0.0801   0.0138   0.4670
#> 154:    0.4750   1.0416   2.9668   0.0070   0.4104   0.0841   0.0138   0.4819
#> 155:    0.4535   0.9959   2.9690   0.0069   0.4584   0.0740   0.0171   0.4693
#> 156:    0.4958   1.0126   3.0649   0.0057   0.4047   0.0805   0.0188   0.4825
#> 157:    0.4477   0.9991   3.0897   0.0054   0.3324   0.0704   0.0200   0.5065
#> 158:    0.5050   1.0234   3.1133   0.0051   0.3720   0.0596   0.0229   0.4760
#> 159:    0.4442   1.0203   3.0917   0.0054   0.3527   0.0514   0.0198   0.4889
#> 160:    0.4549   0.9931   3.0166   0.0064   0.4168   0.0570   0.0158   0.4968
#> 161:    0.4608   1.0038   3.0785   0.0055   0.3741   0.0534   0.0188   0.4789
#> 162:    0.4568   1.0266   2.9913   0.0066   0.4522   0.0527   0.0190   0.4742
#> 163:    0.4874   1.0062   3.0151   0.0062   0.4967   0.0417   0.0170   0.4633
#> 164:    0.4707   1.0040   2.8950   0.0080   0.4550   0.0453   0.0198   0.4505
#> 165:    0.4828   1.0009   2.8603   0.0086   0.4469   0.0505   0.0182   0.4531
#> 166:    0.5274   1.0169   2.8670   0.0086   0.4745   0.0496   0.0178   0.4634
#> 167:    0.4898   1.0285   2.9447   0.0075   0.4278   0.0585   0.0218   0.4708
#> 168:    0.4873   1.0282   2.9624   0.0074   0.3973   0.0578   0.0207   0.4528
#> 169:    0.4812   1.0384   2.9701   0.0070   0.3871   0.0667   0.0179   0.4842
#> 170:    0.4844   1.0288   2.9239   0.0076   0.4184   0.0556   0.0184   0.4802
#> 171:    0.4471   1.0276   2.8019   0.0093   0.4141   0.0578   0.0188   0.4546
#> 172:    0.4315   1.0385   2.8005   0.0093   0.4145   0.0571   0.0204   0.4507
#> 173:    0.4244   1.0388   2.6534   0.0112   0.3821   0.0610   0.0226   0.4576
#> 174:    0.4490   1.0203   2.7964   0.0094   0.3334   0.0550   0.0223   0.4525
#> 175:    0.4491   1.0067   2.7608   0.0099   0.3644   0.0637   0.0239   0.4566
#> 176:    0.4285   1.0386   2.8126   0.0093   0.3820   0.0631   0.0199   0.4672
#> 177:    0.4596   1.0290   2.6672   0.0114   0.4707   0.0672   0.0222   0.4599
#> 178:    0.4589   1.0065   2.9180   0.0079   0.3906   0.0752   0.0170   0.4490
#> 179:    0.4890   1.0179   2.9000   0.0080   0.3831   0.0773   0.0150   0.4694
#> 180:    0.4382   1.0205   2.9316   0.0075   0.3146   0.0757   0.0144   0.5138
#> 181:    0.4384   1.0189   2.9433   0.0071   0.3520   0.0731   0.0141   0.4903
#> 182:    0.3792   1.0020   2.8568   0.0082   0.3414   0.0916   0.0118   0.5128
#> 183:    0.4133   1.0272   2.8032   0.0090   0.3523   0.0814   0.0118   0.5067
#> 184:    0.3963   1.0338   2.8254   0.0086   0.4331   0.0696   0.0114   0.4769
#> 185:    0.3783   1.0473   2.6707   0.0108   0.4167   0.0649   0.0124   0.4843
#> 186:    0.4024   1.0594   2.7015   0.0104   0.4177   0.0698   0.0128   0.4901
#> 187:    0.4299   1.0275   2.7684   0.0097   0.4662   0.0519   0.0139   0.4750
#> 188:    0.4289   1.0439   2.8010   0.0091   0.5849   0.0656   0.0138   0.4645
#> 189:    0.4313   1.0321   2.7656   0.0097   0.5111   0.0670   0.0127   0.4678
#> 190:    0.4391   1.0272   2.7202   0.0104   0.4190   0.0535   0.0137   0.4502
#> 191:    0.4100   1.0340   2.7200   0.0101   0.4611   0.0647   0.0137   0.4510
#> 192:    0.4006   1.0357   2.7894   0.0092   0.4084   0.0707   0.0116   0.4706
#> 193:    0.4149   1.0503   2.9093   0.0075   0.4382   0.0688   0.0135   0.4404
#> 194:    0.4167   1.0333   2.9343   0.0071   0.4483   0.0717   0.0096   0.4892
#> 195:    0.3717   1.0380   2.9064   0.0073   0.4141   0.0609   0.0095   0.4875
#> 196:    0.3941   1.0387   2.8818   0.0078   0.4471   0.0775   0.0100   0.4823
#> 197:    0.4041   1.0437   2.8901   0.0080   0.4481   0.0665   0.0129   0.5150
#> 198:    0.4718   1.0247   2.9158   0.0075   0.6110   0.0663   0.0155   0.4890
#> 199:    0.4223   1.0229   2.8925   0.0078   0.4937   0.0727   0.0136   0.4724
#> 200:    0.4234   1.0167   2.8196   0.0089   0.4497   0.0770   0.0145   0.5080
#> 201:    0.4196   1.0146   2.8090   0.0090   0.4285   0.0765   0.0148   0.4862
#> 202:    0.4221   1.0182   2.8035   0.0091   0.4430   0.0730   0.0158   0.4848
#> 203:    0.4311   1.0161   2.7955   0.0092   0.4506   0.0737   0.0166   0.4807
#> 204:    0.4343   1.0121   2.8241   0.0089   0.4525   0.0726   0.0161   0.4817
#> 205:    0.4315   1.0157   2.8386   0.0087   0.4450   0.0744   0.0165   0.4818
#> 206:    0.4339   1.0202   2.8425   0.0086   0.4498   0.0749   0.0163   0.4861
#> 207:    0.4304   1.0213   2.8367   0.0087   0.4450   0.0746   0.0161   0.4830
#> 208:    0.4304   1.0201   2.8441   0.0086   0.4351   0.0733   0.0163   0.4814
#> 209:    0.4358   1.0169   2.8511   0.0085   0.4397   0.0728   0.0163   0.4798
#> 210:    0.4398   1.0141   2.8553   0.0085   0.4401   0.0719   0.0161   0.4777
#> 211:    0.4463   1.0130   2.8558   0.0085   0.4403   0.0719   0.0159   0.4760
#> 212:    0.4530   1.0112   2.8588   0.0085   0.4411   0.0712   0.0156   0.4756
#> 213:    0.4568   1.0080   2.8646   0.0084   0.4404   0.0711   0.0155   0.4780
#> 214:    0.4622   1.0065   2.8773   0.0082   0.4414   0.0719   0.0154   0.4782
#> 215:    0.4638   1.0068   2.8915   0.0081   0.4392   0.0722   0.0156   0.4777
#> 216:    0.4697   1.0057   2.9001   0.0079   0.4454   0.0722   0.0154   0.4800
#> 217:    0.4722   1.0061   2.9040   0.0079   0.4476   0.0720   0.0154   0.4790
#> 218:    0.4743   1.0064   2.9118   0.0078   0.4511   0.0715   0.0155   0.4790
#> 219:    0.4754   1.0059   2.9234   0.0076   0.4527   0.0709   0.0156   0.4799
#> 220:    0.4738   1.0063   2.9283   0.0075   0.4499   0.0706   0.0158   0.4806
#> 221:    0.4727   1.0066   2.9310   0.0075   0.4451   0.0701   0.0159   0.4802
#> 222:    0.4730   1.0072   2.9332   0.0075   0.4459   0.0702   0.0160   0.4798
#> 223:    0.4751   1.0076   2.9349   0.0075   0.4458   0.0698   0.0161   0.4788
#> 224:    0.4777   1.0081   2.9325   0.0075   0.4480   0.0699   0.0162   0.4789
#> 225:    0.4771   1.0068   2.9309   0.0075   0.4474   0.0694   0.0163   0.4794
#> 226:    0.4779   1.0056   2.9263   0.0076   0.4482   0.0690   0.0164   0.4788
#> 227:    0.4773   1.0059   2.9236   0.0076   0.4473   0.0686   0.0164   0.4785
#> 228:    0.4759   1.0069   2.9224   0.0076   0.4456   0.0680   0.0164   0.4781
#> 229:    0.4747   1.0069   2.9185   0.0077   0.4456   0.0679   0.0165   0.4773
#> 230:    0.4741   1.0072   2.9148   0.0077   0.4438   0.0684   0.0165   0.4768
#> 231:    0.4733   1.0082   2.9121   0.0078   0.4412   0.0684   0.0165   0.4765
#> 232:    0.4721   1.0088   2.9114   0.0078   0.4408   0.0685   0.0164   0.4758
#> 233:    0.4707   1.0086   2.9138   0.0078   0.4426   0.0682   0.0165   0.4758
#> 234:    0.4703   1.0088   2.9113   0.0078   0.4413   0.0679   0.0165   0.4761
#> 235:    0.4700   1.0089   2.9091   0.0078   0.4419   0.0677   0.0165   0.4761
#> 236:    0.4691   1.0096   2.9077   0.0078   0.4410   0.0677   0.0166   0.4761
#> 237:    0.4687   1.0096   2.9068   0.0078   0.4405   0.0678   0.0167   0.4766
#> 238:    0.4683   1.0098   2.9069   0.0078   0.4408   0.0677   0.0167   0.4766
#> 239:    0.4678   1.0102   2.9051   0.0079   0.4420   0.0676   0.0167   0.4763
#> 240:    0.4679   1.0100   2.9038   0.0079   0.4406   0.0677   0.0168   0.4759
#> 241:    0.4688   1.0091   2.9044   0.0079   0.4414   0.0679   0.0168   0.4759
#> 242:    0.4690   1.0084   2.9050   0.0079   0.4399   0.0683   0.0167   0.4762
#> 243:    0.4690   1.0088   2.9050   0.0079   0.4380   0.0688   0.0167   0.4765
#> 244:    0.4681   1.0087   2.9065   0.0079   0.4355   0.0691   0.0167   0.4765
#> 245:    0.4684   1.0084   2.9075   0.0079   0.4330   0.0692   0.0166   0.4772
#> 246:    0.4693   1.0085   2.9071   0.0079   0.4324   0.0694   0.0166   0.4775
#> 247:    0.4699   1.0083   2.9082   0.0079   0.4323   0.0696   0.0166   0.4778
#> 248:    0.4708   1.0083   2.9078   0.0079   0.4328   0.0695   0.0166   0.4775
#> 249:    0.4707   1.0081   2.9090   0.0078   0.4325   0.0695   0.0166   0.4774
#> 250:    0.4712   1.0083   2.9082   0.0079   0.4331   0.0694   0.0167   0.4775
#> 251:    0.4714   1.0087   2.9081   0.0079   0.4352   0.0697   0.0166   0.4775
#> 252:    0.4711   1.0088   2.9086   0.0079   0.4369   0.0696   0.0166   0.4774
#> 253:    0.4705   1.0090   2.9102   0.0078   0.4378   0.0693   0.0167   0.4779
#> 254:    0.4701   1.0094   2.9093   0.0078   0.4380   0.0692   0.0166   0.4776
#> 255:    0.4693   1.0097   2.9071   0.0079   0.4371   0.0690   0.0166   0.4775
#> 256:    0.4696   1.0098   2.9058   0.0079   0.4369   0.0690   0.0167   0.4773
#> 257:    0.4693   1.0093   2.9063   0.0079   0.4363   0.0687   0.0166   0.4774
#> 258:    0.4686   1.0092   2.9042   0.0079   0.4365   0.0687   0.0166   0.4772
#> 259:    0.4679   1.0091   2.9029   0.0079   0.4362   0.0689   0.0165   0.4771
#> 260:    0.4673   1.0095   2.9031   0.0079   0.4359   0.0688   0.0165   0.4769
#> 261:    0.4671   1.0098   2.9037   0.0079   0.4350   0.0687   0.0165   0.4773
#> 262:    0.4669   1.0101   2.9044   0.0079   0.4346   0.0688   0.0165   0.4774
#> 263:    0.4664   1.0106   2.9037   0.0079   0.4338   0.0687   0.0166   0.4777
#> 264:    0.4659   1.0111   2.9039   0.0079   0.4343   0.0687   0.0166   0.4783
#> 265:    0.4658   1.0113   2.9046   0.0079   0.4349   0.0688   0.0166   0.4785
#> 266:    0.4660   1.0112   2.9043   0.0079   0.4364   0.0689   0.0166   0.4784
#> 267:    0.4662   1.0111   2.9043   0.0079   0.4383   0.0689   0.0165   0.4783
#> 268:    0.4663   1.0109   2.9044   0.0079   0.4392   0.0688   0.0165   0.4785
#> 269:    0.4665   1.0108   2.9035   0.0079   0.4391   0.0689   0.0165   0.4786
#> 270:    0.4671   1.0108   2.9040   0.0079   0.4401   0.0691   0.0164   0.4785
#> 271:    0.4677   1.0105   2.9037   0.0079   0.4408   0.0692   0.0164   0.4786
#> 272:    0.4681   1.0102   2.9039   0.0079   0.4408   0.0693   0.0163   0.4792
#> 273:    0.4682   1.0102   2.9040   0.0079   0.4406   0.0697   0.0163   0.4793
#> 274:    0.4686   1.0101   2.9045   0.0079   0.4397   0.0698   0.0163   0.4791
#> 275:    0.4689   1.0102   2.9056   0.0079   0.4393   0.0699   0.0163   0.4789
#> 276:    0.4693   1.0102   2.9068   0.0079   0.4398   0.0699   0.0163   0.4788
#> 277:    0.4694   1.0102   2.9074   0.0079   0.4394   0.0699   0.0162   0.4788
#> 278:    0.4693   1.0101   2.9076   0.0079   0.4389   0.0699   0.0162   0.4788
#> 279:    0.4696   1.0099   2.9078   0.0079   0.4394   0.0698   0.0162   0.4790
#> 280:    0.4696   1.0097   2.9071   0.0079   0.4392   0.0699   0.0162   0.4791
#> 281:    0.4695   1.0097   2.9068   0.0079   0.4395   0.0699   0.0162   0.4787
#> 282:    0.4697   1.0100   2.9068   0.0079   0.4407   0.0700   0.0161   0.4785
#> 283:    0.4699   1.0101   2.9077   0.0079   0.4420   0.0700   0.0160   0.4789
#> 284:    0.4696   1.0101   2.9080   0.0079   0.4431   0.0700   0.0160   0.4788
#> 285:    0.4693   1.0101   2.9076   0.0079   0.4428   0.0700   0.0160   0.4790
#> 286:    0.4691   1.0100   2.9072   0.0079   0.4427   0.0700   0.0160   0.4789
#> 287:    0.4687   1.0101   2.9082   0.0079   0.4425   0.0701   0.0160   0.4790
#> 288:    0.4684   1.0102   2.9064   0.0079   0.4419   0.0700   0.0160   0.4792
#> 289:    0.4682   1.0100   2.9060   0.0079   0.4414   0.0699   0.0160   0.4792
#> 290:    0.4686   1.0100   2.9061   0.0079   0.4412   0.0699   0.0160   0.4792
#> 291:    0.4691   1.0101   2.9061   0.0079   0.4419   0.0699   0.0160   0.4790
#> 292:    0.4693   1.0101   2.9059   0.0079   0.4427   0.0700   0.0160   0.4789
#> 293:    0.4695   1.0100   2.9061   0.0079   0.4434   0.0700   0.0160   0.4791
#> 294:    0.4696   1.0100   2.9066   0.0079   0.4440   0.0700   0.0159   0.4790
#> 295:    0.4697   1.0097   2.9069   0.0079   0.4448   0.0700   0.0159   0.4788
#> 296:    0.4695   1.0099   2.9070   0.0079   0.4440   0.0700   0.0159   0.4789
#> 297:    0.4695   1.0101   2.9076   0.0079   0.4438   0.0700   0.0160   0.4789
#> 298:    0.4695   1.0101   2.9079   0.0079   0.4435   0.0699   0.0159   0.4788
#> 299:    0.4694   1.0102   2.9079   0.0079   0.4436   0.0699   0.0159   0.4787
#> 300:    0.4691   1.0100   2.9075   0.0079   0.4434   0.0698   0.0158   0.4788
#> 301:    0.4689   1.0099   2.9073   0.0079   0.4428   0.0698   0.0158   0.4792
#> 302:    0.4690   1.0099   2.9068   0.0079   0.4427   0.0698   0.0158   0.4790
#> 303:    0.4690   1.0099   2.9063   0.0079   0.4426   0.0697   0.0158   0.4787
#> 304:    0.4693   1.0100   2.9048   0.0079   0.4437   0.0696   0.0158   0.4787
#> 305:    0.4691   1.0099   2.9033   0.0079   0.4440   0.0696   0.0158   0.4787
#> 306:    0.4687   1.0098   2.9016   0.0080   0.4439   0.0697   0.0158   0.4787
#> 307:    0.4683   1.0097   2.9000   0.0080   0.4431   0.0697   0.0158   0.4790
#> 308:    0.4678   1.0098   2.8990   0.0080   0.4430   0.0697   0.0158   0.4792
#> 309:    0.4675   1.0098   2.8987   0.0080   0.4423   0.0697   0.0158   0.4796
#> 310:    0.4680   1.0099   2.8984   0.0080   0.4425   0.0698   0.0158   0.4799
#> 311:    0.4681   1.0099   2.8973   0.0080   0.4427   0.0701   0.0157   0.4799
#> 312:    0.4681   1.0099   2.8954   0.0080   0.4425   0.0703   0.0157   0.4799
#> 313:    0.4678   1.0096   2.8948   0.0080   0.4423   0.0704   0.0156   0.4803
#> 314:    0.4675   1.0095   2.8943   0.0081   0.4420   0.0703   0.0156   0.4803
#> 315:    0.4674   1.0096   2.8938   0.0081   0.4422   0.0703   0.0156   0.4802
#> 316:    0.4674   1.0097   2.8937   0.0081   0.4424   0.0703   0.0155   0.4801
#> 317:    0.4675   1.0098   2.8938   0.0081   0.4418   0.0703   0.0155   0.4802
#> 318:    0.4676   1.0098   2.8938   0.0081   0.4411   0.0702   0.0155   0.4801
#> 319:    0.4675   1.0098   2.8930   0.0081   0.4404   0.0702   0.0155   0.4802
#> 320:    0.4672   1.0096   2.8925   0.0081   0.4410   0.0702   0.0155   0.4803
#> 321:    0.4667   1.0097   2.8911   0.0081   0.4410   0.0700   0.0155   0.4804
#> 322:    0.4664   1.0098   2.8907   0.0081   0.4416   0.0699   0.0155   0.4804
#> 323:    0.4662   1.0098   2.8907   0.0081   0.4418   0.0699   0.0155   0.4801
#> 324:    0.4659   1.0099   2.8908   0.0081   0.4419   0.0699   0.0154   0.4798
#> 325:    0.4658   1.0102   2.8903   0.0081   0.4419   0.0698   0.0154   0.4796
#> 326:    0.4653   1.0103   2.8902   0.0081   0.4417   0.0698   0.0154   0.4795
#> 327:    0.4654   1.0104   2.8897   0.0081   0.4425   0.0698   0.0155   0.4793
#> 328:    0.4655   1.0106   2.8900   0.0081   0.4431   0.0697   0.0155   0.4793
#> 329:    0.4654   1.0106   2.8897   0.0081   0.4431   0.0697   0.0154   0.4793
#> 330:    0.4651   1.0107   2.8895   0.0081   0.4430   0.0697   0.0154   0.4795
#> 331:    0.4649   1.0109   2.8894   0.0081   0.4429   0.0697   0.0154   0.4793
#> 332:    0.4651   1.0109   2.8895   0.0081   0.4432   0.0698   0.0154   0.4794
#> 333:    0.4651   1.0110   2.8891   0.0081   0.4435   0.0698   0.0154   0.4794
#> 334:    0.4648   1.0110   2.8882   0.0081   0.4432   0.0698   0.0154   0.4795
#> 335:    0.4645   1.0109   2.8882   0.0081   0.4433   0.0698   0.0154   0.4795
#> 336:    0.4638   1.0112   2.8876   0.0081   0.4432   0.0699   0.0154   0.4796
#> 337:    0.4635   1.0115   2.8871   0.0081   0.4433   0.0700   0.0154   0.4797
#> 338:    0.4635   1.0115   2.8860   0.0081   0.4435   0.0701   0.0154   0.4796
#> 339:    0.4633   1.0116   2.8859   0.0081   0.4434   0.0700   0.0154   0.4795
#> 340:    0.4632   1.0116   2.8856   0.0082   0.4436   0.0701   0.0154   0.4795
#> 341:    0.4631   1.0117   2.8850   0.0082   0.4438   0.0702   0.0154   0.4795
#> 342:    0.4627   1.0116   2.8844   0.0082   0.4436   0.0702   0.0154   0.4795
#> 343:    0.4623   1.0116   2.8839   0.0082   0.4431   0.0703   0.0154   0.4796
#> 344:    0.4623   1.0116   2.8837   0.0082   0.4431   0.0702   0.0154   0.4796
#> 345:    0.4621   1.0115   2.8832   0.0082   0.4430   0.0702   0.0153   0.4797
#> 346:    0.4619   1.0117   2.8831   0.0082   0.4428   0.0701   0.0153   0.4797
#> 347:    0.4615   1.0118   2.8832   0.0082   0.4425   0.0702   0.0153   0.4797
#> 348:    0.4611   1.0119   2.8828   0.0082   0.4424   0.0702   0.0153   0.4799
#> 349:    0.4605   1.0120   2.8828   0.0082   0.4424   0.0702   0.0153   0.4799
#> 350:    0.4602   1.0120   2.8825   0.0082   0.4428   0.0702   0.0153   0.4799
#> 351:    0.4601   1.0122   2.8822   0.0082   0.4429   0.0702   0.0153   0.4799
#> 352:    0.4597   1.0122   2.8823   0.0082   0.4429   0.0702   0.0154   0.4802
#> 353:    0.4596   1.0124   2.8816   0.0082   0.4428   0.0702   0.0154   0.4803
#> 354:    0.4594   1.0125   2.8818   0.0082   0.4429   0.0702   0.0154   0.4804
#> 355:    0.4594   1.0126   2.8826   0.0082   0.4428   0.0702   0.0154   0.4804
#> 356:    0.4592   1.0125   2.8824   0.0082   0.4427   0.0702   0.0154   0.4806
#> 357:    0.4595   1.0127   2.8831   0.0082   0.4428   0.0701   0.0153   0.4805
#> 358:    0.4593   1.0127   2.8836   0.0082   0.4426   0.0701   0.0153   0.4806
#> 359:    0.4591   1.0129   2.8842   0.0082   0.4424   0.0701   0.0153   0.4805
#> 360:    0.4590   1.0129   2.8846   0.0081   0.4423   0.0701   0.0153   0.4804
#> 361:    0.4588   1.0130   2.8851   0.0081   0.4420   0.0701   0.0153   0.4804
#> 362:    0.4588   1.0131   2.8851   0.0081   0.4417   0.0701   0.0153   0.4802
#> 363:    0.4587   1.0130   2.8849   0.0081   0.4417   0.0700   0.0153   0.4798
#> 364:    0.4589   1.0129   2.8850   0.0081   0.4422   0.0699   0.0153   0.4798
#> 365:    0.4588   1.0128   2.8842   0.0081   0.4422   0.0699   0.0153   0.4798
#> 366:    0.4587   1.0128   2.8838   0.0082   0.4417   0.0699   0.0153   0.4801
#> 367:    0.4586   1.0128   2.8836   0.0082   0.4414   0.0699   0.0153   0.4801
#> 368:    0.4588   1.0128   2.8830   0.0082   0.4414   0.0699   0.0153   0.4799
#> 369:    0.4588   1.0128   2.8830   0.0082   0.4418   0.0699   0.0153   0.4799
#> 370:    0.4587   1.0128   2.8825   0.0082   0.4417   0.0699   0.0153   0.4800
#> 371:    0.4585   1.0130   2.8820   0.0082   0.4413   0.0699   0.0153   0.4800
#> 372:    0.4580   1.0132   2.8812   0.0082   0.4410   0.0699   0.0153   0.4800
#> 373:    0.4574   1.0135   2.8805   0.0082   0.4407   0.0700   0.0152   0.4801
#> 374:    0.4569   1.0137   2.8797   0.0082   0.4407   0.0700   0.0152   0.4802
#> 375:    0.4566   1.0138   2.8792   0.0082   0.4406   0.0700   0.0152   0.4803
#> 376:    0.4563   1.0139   2.8787   0.0082   0.4403   0.0701   0.0152   0.4802
#> 377:    0.4561   1.0140   2.8783   0.0082   0.4401   0.0701   0.0152   0.4801
#> 378:    0.4557   1.0140   2.8788   0.0082   0.4399   0.0701   0.0152   0.4802
#> 379:    0.4556   1.0139   2.8791   0.0082   0.4398   0.0701   0.0152   0.4802
#> 380:    0.4554   1.0139   2.8790   0.0082   0.4395   0.0702   0.0152   0.4801
#> 381:    0.4551   1.0140   2.8789   0.0082   0.4393   0.0703   0.0152   0.4801
#> 382:    0.4550   1.0140   2.8793   0.0082   0.4390   0.0703   0.0151   0.4802
#> 383:    0.4550   1.0140   2.8793   0.0082   0.4387   0.0702   0.0152   0.4800
#> 384:    0.4548   1.0140   2.8790   0.0082   0.4383   0.0702   0.0152   0.4800
#> 385:    0.4549   1.0139   2.8784   0.0082   0.4383   0.0702   0.0152   0.4799
#> 386:    0.4551   1.0140   2.8777   0.0082   0.4385   0.0702   0.0152   0.4798
#> 387:    0.4554   1.0140   2.8772   0.0082   0.4385   0.0701   0.0152   0.4798
#> 388:    0.4556   1.0140   2.8764   0.0082   0.4387   0.0702   0.0152   0.4797
#> 389:    0.4556   1.0138   2.8762   0.0082   0.4387   0.0702   0.0152   0.4797
#> 390:    0.4555   1.0137   2.8765   0.0082   0.4388   0.0702   0.0152   0.4797
#> 391:    0.4557   1.0136   2.8770   0.0082   0.4389   0.0703   0.0152   0.4796
#> 392:    0.4558   1.0136   2.8771   0.0082   0.4393   0.0702   0.0152   0.4797
#> 393:    0.4560   1.0135   2.8775   0.0082   0.4399   0.0702   0.0152   0.4798
#> 394:    0.4561   1.0134   2.8779   0.0082   0.4403   0.0702   0.0152   0.4799
#> 395:    0.4562   1.0134   2.8782   0.0082   0.4402   0.0701   0.0152   0.4799
#> 396:    0.4563   1.0134   2.8785   0.0082   0.4404   0.0701   0.0152   0.4799
#> 397:    0.4565   1.0133   2.8785   0.0082   0.4405   0.0702   0.0152   0.4799
#> 398:    0.4567   1.0132   2.8783   0.0082   0.4403   0.0702   0.0152   0.4800
#> 399:    0.4567   1.0132   2.8779   0.0082   0.4403   0.0703   0.0152   0.4800
#> 400:    0.4567   1.0133   2.8773   0.0082   0.4403   0.0703   0.0152   0.4800
#> 401:    0.4564   1.0133   2.8761   0.0083   0.4402   0.0703   0.0152   0.4802
#> 402:    0.4564   1.0134   2.8752   0.0083   0.4399   0.0703   0.0152   0.4801
#> 403:    0.4564   1.0135   2.8749   0.0083   0.4400   0.0703   0.0152   0.4800
#> 404:    0.4563   1.0135   2.8743   0.0083   0.4398   0.0703   0.0152   0.4799
#> 405:    0.4561   1.0136   2.8739   0.0083   0.4400   0.0703   0.0152   0.4799
#> 406:    0.4560   1.0136   2.8737   0.0083   0.4399   0.0703   0.0152   0.4800
#> 407:    0.4559   1.0136   2.8738   0.0083   0.4399   0.0704   0.0152   0.4799
#> 408:    0.4561   1.0136   2.8740   0.0083   0.4399   0.0704   0.0152   0.4800
#> 409:    0.4560   1.0137   2.8744   0.0083   0.4395   0.0704   0.0152   0.4799
#> 410:    0.4560   1.0137   2.8748   0.0083   0.4395   0.0704   0.0152   0.4800
#> 411:    0.4562   1.0136   2.8751   0.0083   0.4397   0.0704   0.0152   0.4800
#> 412:    0.4560   1.0138   2.8758   0.0083   0.4391   0.0704   0.0152   0.4800
#> 413:    0.4558   1.0138   2.8760   0.0082   0.4385   0.0704   0.0152   0.4801
#> 414:    0.4556   1.0139   2.8758   0.0083   0.4387   0.0704   0.0152   0.4800
#> 415:    0.4558   1.0139   2.8757   0.0083   0.4391   0.0704   0.0152   0.4799
#> 416:    0.4558   1.0139   2.8753   0.0083   0.4392   0.0704   0.0152   0.4799
#> 417:    0.4559   1.0139   2.8750   0.0083   0.4394   0.0704   0.0152   0.4798
#> 418:    0.4558   1.0138   2.8748   0.0083   0.4393   0.0704   0.0152   0.4796
#> 419:    0.4559   1.0139   2.8746   0.0083   0.4390   0.0704   0.0152   0.4796
#> 420:    0.4559   1.0139   2.8748   0.0083   0.4388   0.0705   0.0151   0.4794
#> 421:    0.4557   1.0139   2.8747   0.0083   0.4385   0.0704   0.0151   0.4794
#> 422:    0.4558   1.0139   2.8748   0.0083   0.4387   0.0704   0.0151   0.4793
#> 423:    0.4557   1.0139   2.8749   0.0083   0.4387   0.0704   0.0151   0.4792
#> 424:    0.4555   1.0140   2.8747   0.0083   0.4387   0.0705   0.0151   0.4791
#> 425:    0.4555   1.0140   2.8745   0.0083   0.4388   0.0705   0.0151   0.4790
#> 426:    0.4555   1.0141   2.8744   0.0083   0.4389   0.0705   0.0151   0.4789
#> 427:    0.4554   1.0142   2.8748   0.0083   0.4391   0.0705   0.0151   0.4788
#> 428:    0.4554   1.0142   2.8753   0.0083   0.4390   0.0705   0.0151   0.4789
#> 429:    0.4555   1.0142   2.8753   0.0083   0.4390   0.0705   0.0151   0.4788
#> 430:    0.4554   1.0143   2.8758   0.0083   0.4390   0.0705   0.0151   0.4788
#> 431:    0.4554   1.0143   2.8762   0.0082   0.4387   0.0705   0.0151   0.4787
#> 432:    0.4555   1.0144   2.8767   0.0082   0.4387   0.0705   0.0151   0.4785
#> 433:    0.4555   1.0145   2.8765   0.0082   0.4389   0.0706   0.0152   0.4784
#> 434:    0.4555   1.0144   2.8766   0.0082   0.4389   0.0706   0.0152   0.4783
#> 435:    0.4555   1.0144   2.8768   0.0082   0.4388   0.0706   0.0152   0.4782
#> 436:    0.4555   1.0144   2.8777   0.0082   0.4386   0.0706   0.0152   0.4783
#> 437:    0.4555   1.0144   2.8781   0.0082   0.4384   0.0706   0.0152   0.4783
#> 438:    0.4555   1.0145   2.8779   0.0082   0.4383   0.0706   0.0151   0.4782
#> 439:    0.4556   1.0146   2.8780   0.0082   0.4384   0.0706   0.0151   0.4781
#> 440:    0.4556   1.0147   2.8782   0.0082   0.4382   0.0706   0.0151   0.4781
#> 441:    0.4557   1.0146   2.8784   0.0082   0.4383   0.0705   0.0151   0.4781
#> 442:    0.4558   1.0146   2.8784   0.0082   0.4383   0.0705   0.0151   0.4779
#> 443:    0.4559   1.0147   2.8785   0.0082   0.4385   0.0705   0.0151   0.4778
#> 444:    0.4558   1.0148   2.8790   0.0082   0.4385   0.0706   0.0151   0.4777
#> 445:    0.4558   1.0149   2.8788   0.0082   0.4389   0.0706   0.0151   0.4777
#> 446:    0.4557   1.0149   2.8788   0.0082   0.4392   0.0706   0.0150   0.4777
#> 447:    0.4559   1.0149   2.8789   0.0082   0.4393   0.0706   0.0150   0.4776
#> 448:    0.4560   1.0148   2.8793   0.0082   0.4391   0.0706   0.0150   0.4776
#> 449:    0.4562   1.0148   2.8795   0.0082   0.4387   0.0706   0.0150   0.4776
#> 450:    0.4563   1.0147   2.8796   0.0082   0.4385   0.0705   0.0150   0.4777
#> 451:    0.4566   1.0146   2.8795   0.0082   0.4388   0.0705   0.0151   0.4777
#> 452:    0.4567   1.0146   2.8798   0.0082   0.4390   0.0705   0.0150   0.4777
#> 453:    0.4567   1.0146   2.8799   0.0082   0.4390   0.0705   0.0150   0.4776
#> 454:    0.4567   1.0147   2.8801   0.0082   0.4389   0.0705   0.0150   0.4775
#> 455:    0.4567   1.0147   2.8801   0.0082   0.4391   0.0705   0.0150   0.4775
#> 456:    0.4566   1.0147   2.8801   0.0082   0.4391   0.0705   0.0150   0.4775
#> 457:    0.4564   1.0147   2.8801   0.0082   0.4392   0.0704   0.0150   0.4775
#> 458:    0.4564   1.0148   2.8803   0.0082   0.4393   0.0704   0.0150   0.4773
#> 459:    0.4562   1.0150   2.8804   0.0082   0.4392   0.0704   0.0150   0.4773
#> 460:    0.4560   1.0150   2.8808   0.0082   0.4390   0.0704   0.0150   0.4773
#> 461:    0.4560   1.0151   2.8813   0.0082   0.4388   0.0703   0.0150   0.4774
#> 462:    0.4557   1.0152   2.8815   0.0082   0.4386   0.0703   0.0150   0.4774
#> 463:    0.4557   1.0153   2.8819   0.0082   0.4384   0.0703   0.0150   0.4774
#> 464:    0.4557   1.0153   2.8824   0.0082   0.4384   0.0703   0.0150   0.4774
#> 465:    0.4557   1.0152   2.8827   0.0082   0.4382   0.0703   0.0150   0.4774
#> 466:    0.4557   1.0152   2.8829   0.0082   0.4379   0.0703   0.0150   0.4775
#> 467:    0.4555   1.0151   2.8833   0.0081   0.4376   0.0703   0.0150   0.4774
#> 468:    0.4556   1.0151   2.8837   0.0081   0.4374   0.0703   0.0150   0.4774
#> 469:    0.4557   1.0152   2.8842   0.0081   0.4374   0.0704   0.0149   0.4773
#> 470:    0.4557   1.0152   2.8843   0.0081   0.4374   0.0704   0.0149   0.4774
#> 471:    0.4557   1.0152   2.8844   0.0081   0.4374   0.0705   0.0149   0.4775
#> 472:    0.4557   1.0151   2.8842   0.0081   0.4374   0.0705   0.0149   0.4776
#> 473:    0.4557   1.0150   2.8845   0.0081   0.4372   0.0705   0.0149   0.4776
#> 474:    0.4557   1.0149   2.8844   0.0081   0.4370   0.0705   0.0149   0.4777
#> 475:    0.4558   1.0151   2.8848   0.0081   0.4368   0.0705   0.0149   0.4777
#> 476:    0.4558   1.0151   2.8847   0.0081   0.4367   0.0705   0.0149   0.4777
#> 477:    0.4557   1.0151   2.8845   0.0081   0.4365   0.0705   0.0149   0.4777
#> 478:    0.4555   1.0150   2.8840   0.0081   0.4363   0.0705   0.0149   0.4777
#> 479:    0.4556   1.0150   2.8838   0.0081   0.4362   0.0705   0.0149   0.4777
#> 480:    0.4555   1.0151   2.8841   0.0081   0.4359   0.0706   0.0149   0.4778
#> 481:    0.4553   1.0151   2.8842   0.0081   0.4356   0.0706   0.0149   0.4779
#> 482:    0.4553   1.0152   2.8842   0.0081   0.4355   0.0705   0.0149   0.4779
#> 483:    0.4552   1.0152   2.8844   0.0081   0.4357   0.0705   0.0149   0.4779
#> 484:    0.4550   1.0151   2.8844   0.0081   0.4359   0.0705   0.0149   0.4779
#> 485:    0.4551   1.0152   2.8844   0.0081   0.4361   0.0705   0.0149   0.4779
#> 486:    0.4550   1.0153   2.8844   0.0081   0.4360   0.0705   0.0149   0.4779
#> 487:    0.4552   1.0153   2.8848   0.0081   0.4359   0.0705   0.0149   0.4779
#> 488:    0.4553   1.0154   2.8851   0.0081   0.4358   0.0706   0.0149   0.4778
#> 489:    0.4554   1.0153   2.8851   0.0081   0.4358   0.0705   0.0149   0.4778
#> 490:    0.4555   1.0153   2.8852   0.0081   0.4358   0.0705   0.0149   0.4777
#> 491:    0.4554   1.0153   2.8851   0.0081   0.4357   0.0705   0.0149   0.4775
#> 492:    0.4554   1.0154   2.8850   0.0081   0.4356   0.0704   0.0149   0.4775
#> 493:    0.4553   1.0154   2.8850   0.0081   0.4354   0.0704   0.0149   0.4775
#> 494:    0.4553   1.0154   2.8848   0.0081   0.4353   0.0704   0.0149   0.4775
#> 495:    0.4554   1.0154   2.8847   0.0081   0.4352   0.0703   0.0149   0.4775
#> 496:    0.4555   1.0155   2.8845   0.0081   0.4351   0.0703   0.0149   0.4775
#> 497:    0.4555   1.0154   2.8843   0.0081   0.4352   0.0702   0.0149   0.4775
#> 498:    0.4555   1.0154   2.8840   0.0081   0.4351   0.0702   0.0149   0.4775
#> 499:    0.4556   1.0155   2.8838   0.0081   0.4354   0.0702   0.0149   0.4774
#> 500:    0.4557   1.0155   2.8839   0.0081   0.4356   0.0702   0.0148   0.4773
#> Calculating covariance matrix
#> 
#> → creating full model...
#> → pruning branches (`if`/`else`)...
#> ✔ done
#> → loading into symengine environment...
#> ✔ done
#> → compiling EBE model...
#>  
#> ✔ done
#> Needed Covariates:
#> [1] "WT"
#> Calculating residuals/tables
#> done
fit
#> ── nlmixr SAEM(ODE); OBJF not calculated fit ───────────────────────────────────
#> 
#>  Gaussian/Laplacian Likelihoods: AIC() or $objf etc. 
#>  FOCEi CWRES & Likelihoods: addCwres() 
#> 
#> ── Time (sec $time): ───────────────────────────────────────────────────────────
#> 
#>          saem    setup table covariance    other
#> elapsed 6.506 2.010524 0.016      0.012 0.668476
#> 
#> ── Population Parameters ($parFixed or $parFixedDf): ───────────────────────────
#> 
#>        Parameter    Est.      SE %RSE     Back-transformed(95%CI) BSV(CV%)
#> tka       Log Ka   0.456   0.201   44           1.58 (1.06, 2.34)     73.9
#> tcl       Log Cl    1.02  0.0837 8.24           2.76 (2.34, 3.25)     27.0
#> tv         Log V    2.88   0.316   11           17.9 (9.62, 33.2)     12.2
#> add.sd             0.691                                    0.691         
#> wt.est           0.00814 0.00449 55.2 0.00814 (-0.000659, 0.0169)         
#>        Shrink(SD)%
#> tka       -0.309% 
#> tcl         3.49% 
#> tv          14.2% 
#> add.sd            
#> wt.est            
#>  
#>   Covariance Type ($covMethod): linFim
#>   Some strong fixed parameter correlations exist ($cor) :
#>        cor:tcl,tka     cor:tv,tka cor:wt.est,tka     cor:tv,tcl cor:wt.est,tcl 
#>       -0.0542          0.0174       -0.000404         -0.0261         0.00832   
#>  cor:wt.est,tv 
#>        -0.991  
#>  
#> 
#>   No correlations in between subject variability (BSV) matrix
#>   Full BSV covariance ($omega) or correlation ($omegaR; diagonals=SDs) 
#>   Distribution stats (mean/skewness/kurtosis/p-value) available in $shrink 
#> 
#> ── Fit Data (object is a modified tibble): ─────────────────────────────────────
#> # A tibble: 132 × 20
#>   ID     TIME    DV  PRED    RES IPRED   IRES  IWRES eta.ka eta.cl  eta.v    cp
#>   <fct> <dbl> <dbl> <dbl>  <dbl> <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl> <dbl>
#> 1 1      0     0.74  0     0.74   0     0.74   1.07   0.120 -0.498 -0.150  0   
#> 2 1      0.25  2.84  3.02 -0.179  3.87 -1.03  -1.50   0.120 -0.498 -0.150  3.87
#> 3 1      0.57  6.57  5.41  1.16   6.80 -0.229 -0.332  0.120 -0.498 -0.150  6.80
#> # … with 129 more rows, and 8 more variables: depot <dbl>, center <dbl>,
#> #   ka <dbl>, cl <dbl>, v <dbl>, tad <dbl>, dosenum <dbl>, WT <dbl>

# }