Augmented Prediction for nlmixr fit

Augmented Prediction for nlmixr fit

nlmixrAugPred(
  object,
  ...,
  covsInterpolation = c("locf", "linear", "nocb", "midpoint"),
  primary = NULL,
  minimum = NULL,
  maximum = NULL,
  length.out = 51L
)

# S3 method for nlmixrFitData
augPred(
  object,
  primary = NULL,
  minimum = NULL,
  maximum = NULL,
  length.out = 51,
  ...
)

Arguments

object

Nlmixr fit object

...

some methods for the generic may require additional arguments.

covsInterpolation

specifies the interpolation method for time-varying covariates. When solving ODEs it often samples times outside the sampling time specified in events. When this happens, the time varying covariates are interpolated. Currently this can be:

"linear" interpolation, which interpolates the covariate by solving the line between the observed covariates and extrapolating the new covariate value.

"constant" -- Last observation carried forward (the default).

"NOCB" -- Next Observation Carried Backward. This is the same method that NONMEM uses.

"midpoint" Last observation carried forward to midpoint; Next observation carried backward to midpoint.

primary

an optional one-sided formula specifying the primary covariate to be used to generate the augmented predictions. By default, if a covariate can be extracted from the data used to generate object (using getCovariate), it will be used as primary.

minimum

an optional lower limit for the primary covariate. Defaults to min(primary).

maximum

an optional upper limit for the primary covariate. Defaults to max(primary).

length.out

an optional integer with the number of primary covariate values at which to evaluate the predictions. Defaults to 51.

Value

Stacked data.frame with observations, individual/population predictions.

Author

Matthew L. Fidler