Fit a nlmixr model

Fit a nlmixr model

nlmixr_fit(
  uif,
  data,
  est = NULL,
  control = list(),
  ...,
  sum.prod = FALSE,
  table = tableControl(),
  keep = NULL,
  drop = NULL,
  save = NULL,
  envir = parent.frame()
)

Arguments

uif
Parsed nlmixr model (by nlmixr(mod.fn)).

data
Dataset to estimate. Needs to be RxODE compatible (see https://nlmixrdevelopment.github.io/RxODE/articles/RxODE-event-types.html for detailed dataset requirements).

est
Estimation method

control
Estimation control options. They could be nlmeControl, saemControl or foceiControl

...
Parameters passed to estimation method.

sum.prod
Take the RxODE model and use more precise products/sums. Increases solving accuracy and solving time.

table
A list controlling the table options (i.e. CWRES, NPDE etc). See tableControl.

keep
Columns to keep from either the input dataset. For the input dataset, if any records are added to the data LOCF (Last Observation Carried forward) imputation is performed.

drop
Columns to drop from the output

save
This option determines if the fit will be saved to be reloaded if already run. If NULL, get the option from options("nlmixr.save");

envir
Environment that nlmixr is evaluated in.

Value

nlmixr fit object

Author

Matthew L. Fidler