Create a run summary word document

Create a run summary word document

nmDocx(
  x,
  docxOut = NULL,
  docxTemplate = NULL,
  plot = TRUE,
  titleStyle = getOption("nlmixr.docx.title", "Title"),
  subtitleStyle = getOption("nlmixr.docx.subtitle", "Subtitle"),
  normalStyle = getOption("nlmixr.docx.normal", "Normal"),
  headerStyle = getOption("nlmixr.docx.heading1", "Heading 1"),
  centeredStyle = getOption("nlmixr.docx.centered", "centered"),
  preformattedStyle = getOption("nlmixr.docx.preformatted", "HTML Preformatted"),
  width = getOption("nlmixr.docx.width", 69),
  save = FALSE
)

nmSave(x, ..., save = TRUE)

Arguments

x

nlmixr fit object.

docxOut

Output file for run information document. If not specified it is the name of R object where the fit is located with the -YEAR-MONTH-DAY.docx appended. If it is NULL the document is not saved, but the officer object is returned.

docxTemplate

This is the document template. If not specified it defaults to option("nlmixr.docx.template"). If option("nlmixr.docx.template") is not specified it uses the included nlmixr document template. When docxTemplate=NULL it uses the officer blank document.

plot

Boolean indicating if the default goodness of fit plots are added to the document. By default TRUE

titleStyle

This is the word style name for the nlmixr title; Usually this is nlmixr version (R object). Defaults to option("nlmixr.docx.title") or Title

subtitleStyle

This is the word style for the subtitle which is nlmixr model name and date. Defaults to option("nlmixr.docx.subtitle") or Subtitle

normalStyle

This is the word style for normal text. Defaults to option("nlmixr.docx.normal") or Normal

headerStyle

This is the word style for heading text. Defaults to option("nlmixr.docx.heading1") or Heading 1

centeredStyle

This is the word style for centered text which is used for the figures. Defaults to option("nlmixr.docx.centered") or centered

preformattedStyle

This is the preformatted text style for R output lines. Defaults to option("nlmixr.docx.preformatted") or HTML Preformatted

width

Is an integer representing the number of characters your preformatted style supports. By default this is option("nlmixr.docx.width") or 69

save

Should the docx be saved in a zip file with the R rds data object for the fit? By default this is FALSE with nmDocx and TRUE with nmSave

...

when using `nmSave` these arguments are passed to `nmDocx`

Value

An officer docx object

Author

Matthew Fidler

Examples

# \donttest{
 library(nlmixr)
pheno <- function() {
    # Pheno with covariance
  ini({
    tcl <- log(0.008) # typical value of clearance
    tv <-  log(0.6)   # typical value of volume
    ## var(eta.cl)
    eta.cl + eta.v ~ c(1,
                       0.01, 1) ## cov(eta.cl, eta.v), var(eta.v)
                      # interindividual variability on clearance and volume
    add.err <- 0.1    # residual variability
  })
  model({
    cl <- exp(tcl + eta.cl) # individual value of clearance
    v <- exp(tv + eta.v)    # individual value of volume
    ke <- cl / v            # elimination rate constant
    d/dt(A1) = - ke * A1    # model differential equation
    cp = A1 / v             # concentration in plasma
    cp ~ add(add.err)       # define error model
  })
}

fit.s <- nlmixr(pheno, pheno_sd, "saem")
#> ℹ parameter labels from comments will be replaced by 'label()'
#>  
#> → generate SAEM model
#> ✔ done
#> 1:    -5.0359    0.2663    0.9500    0.9500   57.8992
#> 2:    -5.1889    0.3926    0.9025    0.9025   41.7667
#> 3:    -5.3101    0.4225    0.8574    0.8574   29.8681
#> 4:    -5.3731    0.4329    0.8145    0.8145   22.5301
#> 5:    -5.5189    0.4195    0.7738    0.7738   16.9880
#> 6:    -5.5092    0.4104    0.7351    0.7351   15.0575
#> 7:    -5.4898    0.4010    0.6983    0.6983   14.3425
#> 8:    -5.4838    0.4203    0.6634    0.6634   12.5614
#> 9:    -5.4854    0.4203    0.6302    0.6302   11.2022
#> 10:   -5.4476   0.4042   0.5987   0.5987   9.0113
#> 11:   -5.4404   0.4132   0.5688   0.5688   9.8449
#> 12:   -5.3835   0.4174   0.5404   0.5404   8.9020
#> 13:   -5.3784   0.3866   0.5133   0.5133   7.9389
#> 14:   -5.3393   0.3826   0.4877   0.4877   8.0370
#> 15:   -5.3492   0.3853   0.4633   0.4633   8.7886
#> 16:   -5.3128   0.3848   0.4401   0.4401   9.0697
#> 17:   -5.2671   0.3847   0.4181   0.4181   8.3329
#> 18:   -5.3051   0.3743   0.3972   0.3972   8.0292
#> 19:   -5.2309   0.3648   0.3774   0.3774   7.7241
#> 20:   -5.1893   0.3488   0.3585   0.3585   7.4189
#> 21:   -5.1836   0.3726   0.3406   0.3406   7.2053
#> 22:   -5.1668   0.3643   0.3235   0.3235   6.7400
#> 23:   -5.2086   0.3620   0.3074   0.3074   7.3308
#> 24:   -5.2072   0.3605   0.2958   0.2920   7.4623
#> 25:   -5.2475   0.3717   0.2877   0.2774   7.6228
#> 26:   -5.2588   0.3727   0.2904   0.2635   7.2070
#> 27:   -5.2242   0.3830   0.2958   0.2503   6.9180
#> 28:   -5.1947   0.3749   0.2810   0.2378   7.2450
#> 29:   -5.2412   0.3756   0.2670   0.2259   7.9364
#> 30:   -5.2152   0.3761   0.2536   0.2146   8.2754
#> 31:   -5.1819   0.3731   0.2415   0.2039   7.7887
#> 32:   -5.1790   0.3725   0.2682   0.1937   7.4837
#> 33:   -5.1551   0.3560   0.2548   0.1958   8.5285
#> 34:   -5.1690   0.3785   0.2421   0.1879   8.4316
#> 35:   -5.1804   0.3712   0.2300   0.1989   8.4451
#> 36:   -5.1959   0.3795   0.2185   0.2091   8.1348
#> 37:   -5.2145   0.3692   0.2076   0.1986   8.5251
#> 38:   -5.1880   0.3620   0.1999   0.2054   8.6286
#> 39:   -5.1780   0.3612   0.1899   0.2020   8.5912
#> 40:   -5.1894   0.3743   0.2070   0.2064   9.4070
#> 41:   -5.1662   0.3768   0.1966   0.1985   9.3148
#> 42:   -5.1527   0.3484   0.1917   0.1981   8.9524
#> 43:   -5.1422   0.3694   0.1821   0.2039   8.7593
#> 44:   -5.1418   0.3633   0.1730   0.2161   8.7221
#> 45:   -5.1672   0.3681   0.1799   0.2114   8.3544
#> 46:   -5.1805   0.3630   0.1875   0.2044   8.1314
#> 47:   -5.1624   0.3642   0.2159   0.1971   8.2985
#> 48:   -5.1677   0.3649   0.2211   0.1967   9.6814
#> 49:   -5.1622   0.3643   0.2213   0.1869   9.4949
#> 50:   -5.1646   0.3719   0.2294   0.1922   8.6191
#> 51:   -5.1513   0.3667   0.2278   0.1923   8.8473
#> 52:   -5.1565   0.3617   0.2336   0.1940   8.7060
#> 53:   -5.1787   0.3834   0.2494   0.1930   9.2807
#> 54:   -5.1976   0.3729   0.3016   0.1897   8.7931
#> 55:   -5.2162   0.3784   0.2865   0.1935   7.7340
#> 56:   -5.2002   0.3752   0.2722   0.2035   7.6944
#> 57:   -5.1623   0.3644   0.2681   0.1933   8.2134
#> 58:   -5.1542   0.3758   0.2547   0.1836   7.9217
#> 59:   -5.1916   0.3715   0.2599   0.2001   7.4921
#> 60:   -5.1811   0.3643   0.2559   0.1920   7.6357
#> 61:   -5.2024   0.3782   0.3188   0.1870   8.1094
#> 62:   -5.1633   0.3684   0.3029   0.1910   8.2306
#> 63:   -5.2389   0.3693   0.2878   0.1814   8.4394
#> 64:   -5.2283   0.3704   0.2734   0.1894   8.3236
#> 65:   -5.2330   0.3764   0.2870   0.2014   7.6091
#> 66:   -5.1767   0.3539   0.2727   0.1913   7.7561
#> 67:   -5.1660   0.3644   0.2591   0.1946   7.2901
#> 68:   -5.1912   0.3687   0.2461   0.1924   6.8679
#> 69:   -5.1872   0.3623   0.2416   0.1867   6.8828
#> 70:   -5.1607   0.3758   0.2456   0.1909   7.4610
#> 71:   -5.1610   0.3683   0.2449   0.2091   7.9744
#> 72:   -5.1758   0.3787   0.2347   0.1986   7.3759
#> 73:   -5.1766   0.3693   0.2230   0.1938   7.9417
#> 74:   -5.2362   0.3811   0.2151   0.2052   8.2163
#> 75:   -5.2189   0.3710   0.2129   0.2013   8.5755
#> 76:   -5.2045   0.3683   0.2022   0.2111   8.4615
#> 77:   -5.2072   0.3683   0.1921   0.2015   8.2897
#> 78:   -5.2114   0.3766   0.1935   0.2039   7.7556
#> 79:   -5.2124   0.3751   0.1916   0.2032   8.0111
#> 80:   -5.2396   0.3719   0.1820   0.2260   7.5432
#> 81:   -5.2389   0.3773   0.1766   0.2201   7.7797
#> 82:   -5.1866   0.3691   0.1687   0.2215   8.0363
#> 83:   -5.1858   0.3576   0.1603   0.2129   8.5190
#> 84:   -5.1487   0.3499   0.1611   0.2214   8.3299
#> 85:   -5.1543   0.3622   0.1585   0.2104   8.7578
#> 86:   -5.1824   0.3674   0.1541   0.1998   8.5612
#> 87:   -5.1439   0.3570   0.1613   0.2024   8.4641
#> 88:   -5.1443   0.3588   0.1620   0.2086   9.0916
#> 89:   -5.1699   0.3636   0.1668   0.1982   9.0053
#> 90:   -5.1817   0.3665   0.1814   0.2117   8.8737
#> 91:   -5.1593   0.3737   0.1838   0.2105   8.6212
#> 92:   -5.1906   0.3525   0.1766   0.2014   8.0170
#> 93:   -5.1874   0.3756   0.2142   0.2112   7.9887
#> 94:   -5.1986   0.3603   0.2192   0.2111   8.1652
#> 95:   -5.1819   0.3838   0.2239   0.2005   9.2053
#> 96:   -5.1624   0.3625   0.2127   0.2271   8.3498
#> 97:   -5.1369   0.3648   0.2021   0.2158   8.2901
#> 98:   -5.1505   0.3598   0.1920   0.2107   8.1688
#> 99:   -5.1660   0.3725   0.2012   0.2002   7.9863
#> 100:   -5.2007   0.3717   0.2158   0.1981   7.1822
#> 101:   -5.1943   0.3696   0.2057   0.2098   8.1001
#> 102:   -5.1562   0.3591   0.1954   0.2214   8.6574
#> 103:   -5.1252   0.3616   0.1857   0.2104   7.8677
#> 104:   -5.1353   0.3552   0.2073   0.2082   7.6256
#> 105:   -5.1297   0.3531   0.2232   0.2000   7.8383
#> 106:   -5.1834   0.3665   0.2303   0.2025   7.4169
#> 107:   -5.1646   0.3574   0.2310   0.1981   7.0770
#> 108:   -5.1234   0.3718   0.2291   0.2097   7.1829
#> 109:   -5.1501   0.3686   0.2465   0.1992   7.1778
#> 110:   -5.1777   0.3742   0.2413   0.1892   7.4347
#> 111:   -5.1693   0.3690   0.2418   0.1894   7.4731
#> 112:   -5.1979   0.3623   0.2392   0.2060   7.2989
#> 113:   -5.1863   0.3641   0.2378   0.1957   7.4889
#> 114:   -5.1774   0.3695   0.2546   0.1958   7.6738
#> 115:   -5.2043   0.3823   0.2529   0.2026   7.7646
#> 116:   -5.1539   0.3533   0.2414   0.2165   7.8500
#> 117:   -5.1477   0.3651   0.2293   0.2160   8.2157
#> 118:   -5.1489   0.3632   0.2178   0.2083   8.2006
#> 119:   -5.1454   0.3773   0.2086   0.1979   7.7287
#> 120:   -5.1785   0.3587   0.2364   0.1880   7.8211
#> 121:   -5.2043   0.3683   0.2246   0.1976   7.6115
#> 122:   -5.1648   0.3539   0.2240   0.1985   7.8621
#> 123:   -5.1518   0.3624   0.2446   0.1907   7.7152
#> 124:   -5.1391   0.3528   0.2655   0.1871   7.5348
#> 125:   -5.1501   0.3662   0.2522   0.1954   7.6286
#> 126:   -5.1652   0.3596   0.2485   0.1863   7.4144
#> 127:   -5.1609   0.3612   0.2553   0.1803   7.5299
#> 128:   -5.2113   0.3700   0.2426   0.1922   7.2630
#> 129:   -5.1993   0.3712   0.2365   0.1839   7.5161
#> 130:   -5.1960   0.3749   0.2314   0.1884   7.4384
#> 131:   -5.1796   0.3669   0.2371   0.2029   7.6925
#> 132:   -5.1663   0.3623   0.2610   0.2066   7.8105
#> 133:   -5.1622   0.3648   0.2480   0.2046   6.7284
#> 134:   -5.1993   0.3638   0.2356   0.2087   7.4467
#> 135:   -5.2133   0.3851   0.2238   0.2005   7.4995
#> 136:   -5.2066   0.3977   0.2126   0.2042   7.8724
#> 137:   -5.1792   0.3705   0.2069   0.1996   7.4733
#> 138:   -5.1560   0.3591   0.2209   0.2038   7.4087
#> 139:   -5.1415   0.3677   0.2099   0.1994   8.0960
#> 140:   -5.1327   0.3677   0.2262   0.2007   7.7773
#> 141:   -5.1385   0.3537   0.2286   0.1987   7.9027
#> 142:   -5.1682   0.3602   0.2191   0.1981   8.0387
#> 143:   -5.1239   0.3500   0.2337   0.1957   7.8361
#> 144:   -5.1246   0.3450   0.2579   0.1859   8.0069
#> 145:   -5.1044   0.3561   0.2450   0.1962   7.6904
#> 146:   -5.0630   0.3422   0.2327   0.2015   7.4451
#> 147:   -5.0919   0.3633   0.2329   0.1914   7.4583
#> 148:   -5.1172   0.3514   0.2212   0.1856   7.6795
#> 149:   -5.1110   0.3676   0.2102   0.2064   7.6409
#> 150:   -5.1329   0.3601   0.2023   0.2077   7.4696
#> 151:   -5.1565   0.3658   0.1921   0.2051   7.3003
#> 152:   -5.1531   0.3574   0.1860   0.2075   7.6912
#> 153:   -5.1368   0.3598   0.1992   0.2042   7.0819
#> 154:   -5.1197   0.3576   0.2125   0.1981   7.4606
#> 155:   -5.0971   0.3401   0.2141   0.1955   7.6490
#> 156:   -5.0993   0.3455   0.2125   0.1741   7.5645
#> 157:   -5.0863   0.3614   0.2182   0.1783   7.0782
#> 158:   -5.0887   0.3495   0.2314   0.1826   7.5388
#> 159:   -5.0783   0.3541   0.2244   0.1792   7.3034
#> 160:   -5.0690   0.3617   0.2322   0.1791   7.0222
#> 161:   -5.0420   0.3536   0.2103   0.1733   7.6565
#> 162:   -5.0536   0.3558   0.2393   0.1757   7.5488
#> 163:   -5.0332   0.3611   0.2256   0.1879   6.9114
#> 164:   -5.0432   0.3481   0.2335   0.1786   6.7526
#> 165:   -5.0435   0.3599   0.2271   0.1632   6.8728
#> 166:   -5.0419   0.3608   0.2469   0.1687   7.3423
#> 167:   -5.0442   0.3506   0.2255   0.1617   6.8157
#> 168:   -5.0297   0.3465   0.2315   0.1720   7.1303
#> 169:   -5.0318   0.3551   0.2259   0.1649   7.6775
#> 170:   -5.0332   0.3582   0.2416   0.1552   7.4090
#> 171:   -5.0230   0.3620   0.2410   0.1591   6.9659
#> 172:   -5.0301   0.3491   0.2651   0.1635   6.7796
#> 173:   -5.0273   0.3547   0.2645   0.1665   6.7502
#> 174:   -5.0154   0.3548   0.2674   0.1538   7.0588
#> 175:   -5.0047   0.3461   0.2630   0.1487   7.1087
#> 176:   -5.0060   0.3461   0.2748   0.1526   7.1074
#> 177:   -5.0056   0.3470   0.2775   0.1485   7.6389
#> 178:   -5.0244   0.3350   0.2860   0.1484   7.6941
#> 179:   -5.0223   0.3386   0.2933   0.1549   7.7922
#> 180:   -5.0102   0.3462   0.2873   0.1492   7.1890
#> 181:   -5.0073   0.3575   0.2806   0.1442   7.9706
#> 182:   -5.0152   0.3577   0.2805   0.1421   7.8807
#> 183:   -5.0077   0.3586   0.2979   0.1490   7.3785
#> 184:   -5.0122   0.3535   0.2704   0.1473   7.9537
#> 185:   -5.0083   0.3562   0.2719   0.1552   8.2670
#> 186:   -4.9963   0.3612   0.2887   0.1669   8.1549
#> 187:   -4.9958   0.3501   0.2527   0.1467   8.2338
#> 188:   -5.0026   0.3483   0.2576   0.1549   8.3997
#> 189:   -4.9937   0.3556   0.2666   0.1536   8.1778
#> 190:   -4.9911   0.3640   0.2538   0.1506   8.7438
#> 191:   -5.0084   0.3572   0.2577   0.1508   8.7269
#> 192:   -5.0037   0.3610   0.2630   0.1492   8.6486
#> 193:   -5.0301   0.3479   0.2604   0.1491   8.3267
#> 194:   -5.0106   0.3644   0.2520   0.1503   8.3641
#> 195:   -5.0221   0.3536   0.2463   0.1423   8.3655
#> 196:   -5.0220   0.3564   0.2572   0.1450   8.1480
#> 197:   -5.0171   0.3582   0.2539   0.1427   7.9381
#> 198:   -5.0200   0.3427   0.2485   0.1376   7.8716
#> 199:   -5.0277   0.3406   0.2546   0.1448   7.7382
#> 200:   -5.0182   0.3441   0.2495   0.1455   7.9370
#> 201:   -5.0180   0.3445   0.2514   0.1447   7.8544
#> 202:   -5.0201   0.3443   0.2566   0.1474   7.8523
#> 203:   -5.0192   0.3464   0.2594   0.1476   7.8785
#> 204:   -5.0173   0.3482   0.2597   0.1481   7.8679
#> 205:   -5.0156   0.3497   0.2579   0.1472   7.8203
#> 206:   -5.0142   0.3498   0.2573   0.1467   7.8232
#> 207:   -5.0151   0.3489   0.2576   0.1466   7.8045
#> 208:   -5.0157   0.3485   0.2566   0.1465   7.7664
#> 209:   -5.0157   0.3487   0.2557   0.1464   7.7955
#> 210:   -5.0158   0.3491   0.2552   0.1460   7.7491
#> 211:   -5.0158   0.3488   0.2557   0.1457   7.7528
#> 212:   -5.0157   0.3486   0.2556   0.1458   7.7560
#> 213:   -5.0155   0.3491   0.2559   0.1456   7.7830
#> 214:   -5.0152   0.3494   0.2557   0.1459   7.8280
#> 215:   -5.0152   0.3496   0.2557   0.1458   7.8562
#> 216:   -5.0158   0.3493   0.2561   0.1459   7.8671
#> 217:   -5.0162   0.3488   0.2556   0.1458   7.8829
#> 218:   -5.0161   0.3488   0.2554   0.1454   7.8879
#> 219:   -5.0166   0.3485   0.2551   0.1455   7.8698
#> 220:   -5.0162   0.3488   0.2554   0.1458   7.8821
#> 221:   -5.0158   0.3491   0.2556   0.1460   7.8931
#> 222:   -5.0156   0.3493   0.2556   0.1461   7.8937
#> 223:   -5.0148   0.3499   0.2566   0.1467   7.9028
#> 224:   -5.0140   0.3500   0.2572   0.1471   7.9160
#> 225:   -5.0139   0.3501   0.2579   0.1474   7.9400
#> 226:   -5.0138   0.3501   0.2583   0.1480   7.9405
#> 227:   -5.0136   0.3504   0.2590   0.1483   7.9564
#> 228:   -5.0136   0.3502   0.2593   0.1485   7.9369
#> 229:   -5.0138   0.3500   0.2591   0.1484   7.9169
#> 230:   -5.0142   0.3496   0.2588   0.1481   7.9359
#> 231:   -5.0142   0.3492   0.2585   0.1480   7.9489
#> 232:   -5.0142   0.3491   0.2585   0.1481   7.9415
#> 233:   -5.0143   0.3489   0.2585   0.1483   7.9402
#> 234:   -5.0145   0.3488   0.2588   0.1484   7.9279
#> 235:   -5.0140   0.3492   0.2587   0.1485   7.9340
#> 236:   -5.0138   0.3494   0.2589   0.1486   7.9225
#> 237:   -5.0137   0.3494   0.2590   0.1486   7.9169
#> 238:   -5.0137   0.3495   0.2593   0.1487   7.9184
#> 239:   -5.0136   0.3496   0.2593   0.1487   7.9260
#> 240:   -5.0140   0.3495   0.2597   0.1488   7.9389
#> 241:   -5.0140   0.3497   0.2601   0.1490   7.9397
#> 242:   -5.0137   0.3499   0.2602   0.1491   7.9389
#> 243:   -5.0138   0.3499   0.2602   0.1491   7.9372
#> 244:   -5.0139   0.3499   0.2601   0.1491   7.9387
#> 245:   -5.0137   0.3500   0.2602   0.1492   7.9309
#> 246:   -5.0136   0.3500   0.2603   0.1494   7.9216
#> 247:   -5.0138   0.3499   0.2605   0.1494   7.9088
#> 248:   -5.0140   0.3499   0.2605   0.1493   7.8965
#> 249:   -5.0140   0.3498   0.2607   0.1494   7.8929
#> 250:   -5.0138   0.3499   0.2609   0.1496   7.8945
#> 251:   -5.0140   0.3497   0.2611   0.1497   7.8972
#> 252:   -5.0139   0.3498   0.2612   0.1497   7.9028
#> 253:   -5.0140   0.3498   0.2613   0.1498   7.9044
#> 254:   -5.0142   0.3497   0.2614   0.1498   7.9049
#> 255:   -5.0144   0.3495   0.2614   0.1498   7.9028
#> 256:   -5.0146   0.3494   0.2616   0.1500   7.9041
#> 257:   -5.0149   0.3493   0.2618   0.1501   7.9012
#> 258:   -5.0147   0.3494   0.2617   0.1501   7.9116
#> 259:   -5.0145   0.3496   0.2617   0.1501   7.9154
#> 260:   -5.0145   0.3497   0.2618   0.1500   7.9143
#> 261:   -5.0145   0.3497   0.2619   0.1501   7.9225
#> 262:   -5.0145   0.3497   0.2620   0.1501   7.9246
#> 263:   -5.0146   0.3497   0.2621   0.1500   7.9210
#> 264:   -5.0144   0.3498   0.2619   0.1500   7.9223
#> 265:   -5.0144   0.3498   0.2617   0.1498   7.9233
#> 266:   -5.0144   0.3498   0.2616   0.1498   7.9211
#> 267:   -5.0143   0.3499   0.2614   0.1497   7.9189
#> 268:   -5.0143   0.3499   0.2615   0.1498   7.9227
#> 269:   -5.0143   0.3498   0.2614   0.1498   7.9203
#> 270:   -5.0143   0.3498   0.2612   0.1497   7.9153
#> 271:   -5.0142   0.3498   0.2611   0.1496   7.9235
#> 272:   -5.0140   0.3500   0.2609   0.1496   7.9260
#> 273:   -5.0140   0.3500   0.2608   0.1496   7.9233
#> 274:   -5.0138   0.3502   0.2609   0.1496   7.9235
#> 275:   -5.0137   0.3502   0.2608   0.1496   7.9292
#> 276:   -5.0138   0.3501   0.2608   0.1496   7.9281
#> 277:   -5.0140   0.3500   0.2607   0.1495   7.9296
#> 278:   -5.0140   0.3500   0.2607   0.1495   7.9326
#> 279:   -5.0140   0.3500   0.2608   0.1495   7.9396
#> 280:   -5.0140   0.3500   0.2609   0.1496   7.9454
#> 281:   -5.0139   0.3500   0.2609   0.1496   7.9521
#> 282:   -5.0139   0.3501   0.2609   0.1496   7.9582
#> 283:   -5.0137   0.3502   0.2608   0.1496   7.9597
#> 284:   -5.0136   0.3503   0.2608   0.1497   7.9604
#> 285:   -5.0135   0.3503   0.2607   0.1497   7.9625
#> 286:   -5.0134   0.3504   0.2607   0.1496   7.9654
#> 287:   -5.0132   0.3506   0.2606   0.1496   7.9624
#> 288:   -5.0132   0.3506   0.2606   0.1496   7.9614
#> 289:   -5.0132   0.3506   0.2607   0.1497   7.9600
#> 290:   -5.0132   0.3506   0.2607   0.1496   7.9554
#> 291:   -5.0132   0.3508   0.2607   0.1496   7.9612
#> 292:   -5.0134   0.3507   0.2607   0.1496   7.9645
#> 293:   -5.0133   0.3508   0.2608   0.1497   7.9692
#> 294:   -5.0132   0.3508   0.2609   0.1498   7.9733
#> 295:   -5.0130   0.3508   0.2611   0.1499   7.9708
#> 296:   -5.0127   0.3510   0.2611   0.1499   7.9714
#> 297:   -5.0125   0.3511   0.2613   0.1500   7.9750
#> 298:   -5.0122   0.3512   0.2614   0.1500   7.9755
#> 299:   -5.0121   0.3513   0.2615   0.1499   7.9721
#> 300:   -5.0121   0.3513   0.2617   0.1500   7.9683
#> 301:   -5.0120   0.3514   0.2616   0.1500   7.9703
#> 302:   -5.0119   0.3515   0.2616   0.1499   7.9715
#> 303:   -5.0117   0.3516   0.2615   0.1499   7.9705
#> 304:   -5.0117   0.3516   0.2615   0.1499   7.9705
#> 305:   -5.0118   0.3515   0.2615   0.1500   7.9710
#> 306:   -5.0117   0.3516   0.2615   0.1501   7.9739
#> 307:   -5.0116   0.3516   0.2615   0.1500   7.9785
#> 308:   -5.0116   0.3516   0.2616   0.1500   7.9781
#> 309:   -5.0115   0.3516   0.2616   0.1500   7.9849
#> 310:   -5.0116   0.3515   0.2615   0.1499   7.9925
#> 311:   -5.0117   0.3515   0.2614   0.1499   7.9960
#> 312:   -5.0117   0.3515   0.2614   0.1499   7.9945
#> 313:   -5.0115   0.3515   0.2614   0.1499   7.9965
#> 314:   -5.0115   0.3515   0.2615   0.1500   8.0009
#> 315:   -5.0114   0.3515   0.2616   0.1500   7.9996
#> 316:   -5.0114   0.3515   0.2616   0.1501   7.9995
#> 317:   -5.0113   0.3516   0.2618   0.1501   8.0030
#> 318:   -5.0113   0.3516   0.2618   0.1501   8.0007
#> 319:   -5.0114   0.3514   0.2618   0.1501   8.0001
#> 320:   -5.0115   0.3514   0.2617   0.1501   7.9996
#> 321:   -5.0115   0.3514   0.2616   0.1501   7.9993
#> 322:   -5.0116   0.3513   0.2616   0.1500   8.0007
#> 323:   -5.0117   0.3512   0.2617   0.1500   8.0027
#> 324:   -5.0117   0.3512   0.2617   0.1500   8.0028
#> 325:   -5.0116   0.3512   0.2617   0.1500   8.0099
#> 326:   -5.0116   0.3512   0.2616   0.1499   8.0115
#> 327:   -5.0116   0.3512   0.2616   0.1499   8.0118
#> 328:   -5.0116   0.3512   0.2615   0.1499   8.0143
#> 329:   -5.0115   0.3513   0.2614   0.1499   8.0187
#> 330:   -5.0115   0.3514   0.2614   0.1498   8.0205
#> 331:   -5.0115   0.3513   0.2613   0.1498   8.0192
#> 332:   -5.0115   0.3513   0.2611   0.1497   8.0214
#> 333:   -5.0114   0.3512   0.2611   0.1496   8.0203
#> 334:   -5.0114   0.3513   0.2610   0.1496   8.0235
#> 335:   -5.0114   0.3512   0.2611   0.1497   8.0232
#> 336:   -5.0115   0.3512   0.2610   0.1496   8.0177
#> 337:   -5.0116   0.3511   0.2609   0.1496   8.0147
#> 338:   -5.0115   0.3512   0.2610   0.1497   8.0147
#> 339:   -5.0114   0.3512   0.2610   0.1497   8.0174
#> 340:   -5.0114   0.3512   0.2610   0.1497   8.0183
#> 341:   -5.0113   0.3512   0.2611   0.1497   8.0224
#> 342:   -5.0114   0.3512   0.2612   0.1498   8.0241
#> 343:   -5.0114   0.3512   0.2612   0.1498   8.0215
#> 344:   -5.0114   0.3512   0.2613   0.1499   8.0189
#> 345:   -5.0115   0.3511   0.2613   0.1499   8.0168
#> 346:   -5.0114   0.3511   0.2613   0.1500   8.0204
#> 347:   -5.0115   0.3511   0.2613   0.1499   8.0225
#> 348:   -5.0114   0.3512   0.2614   0.1500   8.0218
#> 349:   -5.0113   0.3512   0.2614   0.1500   8.0202
#> 350:   -5.0114   0.3511   0.2615   0.1501   8.0193
#> 351:   -5.0115   0.3511   0.2616   0.1501   8.0176
#> 352:   -5.0115   0.3511   0.2616   0.1501   8.0170
#> 353:   -5.0116   0.3510   0.2615   0.1501   8.0141
#> 354:   -5.0117   0.3510   0.2616   0.1501   8.0195
#> 355:   -5.0117   0.3509   0.2615   0.1501   8.0240
#> 356:   -5.0118   0.3509   0.2615   0.1501   8.0227
#> 357:   -5.0117   0.3509   0.2614   0.1501   8.0207
#> 358:   -5.0118   0.3509   0.2614   0.1501   8.0196
#> 359:   -5.0119   0.3509   0.2615   0.1501   8.0175
#> 360:   -5.0120   0.3508   0.2615   0.1501   8.0161
#> 361:   -5.0121   0.3508   0.2616   0.1501   8.0194
#> 362:   -5.0121   0.3508   0.2615   0.1501   8.0189
#> 363:   -5.0121   0.3508   0.2616   0.1501   8.0135
#> 364:   -5.0121   0.3507   0.2616   0.1502   8.0094
#> 365:   -5.0121   0.3507   0.2616   0.1502   8.0080
#> 366:   -5.0121   0.3508   0.2616   0.1502   8.0070
#> 367:   -5.0120   0.3508   0.2615   0.1502   8.0056
#> 368:   -5.0120   0.3508   0.2616   0.1502   8.0054
#> 369:   -5.0119   0.3509   0.2616   0.1503   8.0063
#> 370:   -5.0119   0.3508   0.2616   0.1502   8.0051
#> 371:   -5.0120   0.3507   0.2615   0.1502   8.0037
#> 372:   -5.0121   0.3507   0.2614   0.1501   8.0028
#> 373:   -5.0121   0.3507   0.2613   0.1501   8.0013
#> 374:   -5.0120   0.3507   0.2613   0.1501   7.9996
#> 375:   -5.0120   0.3507   0.2612   0.1501   8.0012
#> 376:   -5.0120   0.3507   0.2611   0.1500   8.0052
#> 377:   -5.0120   0.3507   0.2611   0.1500   8.0067
#> 378:   -5.0119   0.3508   0.2611   0.1500   8.0080
#> 379:   -5.0120   0.3507   0.2610   0.1500   8.0108
#> 380:   -5.0119   0.3508   0.2609   0.1500   8.0158
#> 381:   -5.0119   0.3507   0.2609   0.1499   8.0174
#> 382:   -5.0119   0.3507   0.2609   0.1499   8.0197
#> 383:   -5.0119   0.3508   0.2609   0.1499   8.0198
#> 384:   -5.0118   0.3508   0.2609   0.1499   8.0189
#> 385:   -5.0118   0.3508   0.2609   0.1500   8.0199
#> 386:   -5.0118   0.3508   0.2610   0.1500   8.0183
#> 387:   -5.0119   0.3507   0.2611   0.1500   8.0196
#> 388:   -5.0118   0.3507   0.2612   0.1501   8.0194
#> 389:   -5.0118   0.3507   0.2612   0.1501   8.0184
#> 390:   -5.0118   0.3507   0.2613   0.1502   8.0200
#> 391:   -5.0117   0.3508   0.2613   0.1502   8.0226
#> 392:   -5.0117   0.3508   0.2613   0.1502   8.0229
#> 393:   -5.0117   0.3508   0.2613   0.1502   8.0251
#> 394:   -5.0116   0.3508   0.2613   0.1502   8.0253
#> 395:   -5.0117   0.3509   0.2614   0.1501   8.0244
#> 396:   -5.0117   0.3508   0.2614   0.1501   8.0219
#> 397:   -5.0117   0.3508   0.2614   0.1501   8.0214
#> 398:   -5.0117   0.3508   0.2614   0.1502   8.0217
#> 399:   -5.0118   0.3508   0.2615   0.1502   8.0181
#> 400:   -5.0118   0.3509   0.2615   0.1502   8.0184
#> 401:   -5.0117   0.3509   0.2615   0.1502   8.0204
#> 402:   -5.0118   0.3509   0.2615   0.1502   8.0216
#> 403:   -5.0117   0.3509   0.2615   0.1502   8.0213
#> 404:   -5.0116   0.3509   0.2614   0.1501   8.0247
#> 405:   -5.0115   0.3510   0.2615   0.1501   8.0270
#> 406:   -5.0115   0.3510   0.2615   0.1501   8.0295
#> 407:   -5.0114   0.3511   0.2614   0.1501   8.0282
#> 408:   -5.0114   0.3511   0.2614   0.1501   8.0267
#> 409:   -5.0115   0.3510   0.2613   0.1501   8.0238
#> 410:   -5.0115   0.3510   0.2614   0.1501   8.0244
#> 411:   -5.0115   0.3510   0.2614   0.1502   8.0251
#> 412:   -5.0115   0.3510   0.2615   0.1502   8.0243
#> 413:   -5.0115   0.3509   0.2615   0.1503   8.0254
#> 414:   -5.0114   0.3510   0.2616   0.1503   8.0268
#> 415:   -5.0114   0.3509   0.2616   0.1503   8.0278
#> 416:   -5.0113   0.3510   0.2616   0.1503   8.0291
#> 417:   -5.0112   0.3511   0.2616   0.1503   8.0306
#> 418:   -5.0112   0.3511   0.2616   0.1503   8.0308
#> 419:   -5.0111   0.3511   0.2616   0.1503   8.0324
#> 420:   -5.0111   0.3511   0.2616   0.1504   8.0340
#> 421:   -5.0110   0.3511   0.2616   0.1504   8.0346
#> 422:   -5.0111   0.3511   0.2616   0.1504   8.0338
#> 423:   -5.0111   0.3512   0.2616   0.1504   8.0337
#> 424:   -5.0111   0.3512   0.2615   0.1504   8.0331
#> 425:   -5.0110   0.3512   0.2616   0.1503   8.0331
#> 426:   -5.0110   0.3513   0.2616   0.1504   8.0313
#> 427:   -5.0110   0.3513   0.2616   0.1504   8.0332
#> 428:   -5.0110   0.3513   0.2616   0.1504   8.0358
#> 429:   -5.0111   0.3512   0.2616   0.1504   8.0359
#> 430:   -5.0111   0.3512   0.2616   0.1504   8.0361
#> 431:   -5.0110   0.3512   0.2616   0.1504   8.0371
#> 432:   -5.0110   0.3513   0.2616   0.1505   8.0390
#> 433:   -5.0110   0.3512   0.2617   0.1505   8.0373
#> 434:   -5.0111   0.3512   0.2617   0.1505   8.0356
#> 435:   -5.0112   0.3512   0.2617   0.1505   8.0351
#> 436:   -5.0111   0.3512   0.2617   0.1505   8.0341
#> 437:   -5.0111   0.3511   0.2617   0.1505   8.0327
#> 438:   -5.0111   0.3512   0.2617   0.1505   8.0332
#> 439:   -5.0112   0.3512   0.2617   0.1505   8.0316
#> 440:   -5.0111   0.3512   0.2617   0.1505   8.0292
#> 441:   -5.0111   0.3512   0.2618   0.1505   8.0277
#> 442:   -5.0110   0.3512   0.2618   0.1506   8.0290
#> 443:   -5.0110   0.3513   0.2618   0.1506   8.0311
#> 444:   -5.0111   0.3513   0.2618   0.1505   8.0310
#> 445:   -5.0111   0.3513   0.2618   0.1505   8.0301
#> 446:   -5.0110   0.3513   0.2618   0.1506   8.0285
#> 447:   -5.0110   0.3513   0.2619   0.1506   8.0275
#> 448:   -5.0109   0.3514   0.2619   0.1506   8.0274
#> 449:   -5.0109   0.3514   0.2618   0.1506   8.0273
#> 450:   -5.0109   0.3514   0.2618   0.1505   8.0294
#> 451:   -5.0108   0.3514   0.2617   0.1505   8.0296
#> 452:   -5.0109   0.3514   0.2617   0.1505   8.0284
#> 453:   -5.0109   0.3514   0.2616   0.1505   8.0289
#> 454:   -5.0109   0.3514   0.2616   0.1504   8.0297
#> 455:   -5.0110   0.3514   0.2615   0.1504   8.0294
#> 456:   -5.0109   0.3514   0.2615   0.1503   8.0288
#> 457:   -5.0110   0.3513   0.2614   0.1503   8.0272
#> 458:   -5.0110   0.3514   0.2614   0.1503   8.0280
#> 459:   -5.0110   0.3513   0.2614   0.1503   8.0288
#> 460:   -5.0110   0.3513   0.2614   0.1503   8.0277
#> 461:   -5.0111   0.3513   0.2614   0.1503   8.0247
#> 462:   -5.0111   0.3513   0.2614   0.1503   8.0245
#> 463:   -5.0111   0.3513   0.2613   0.1503   8.0228
#> 464:   -5.0111   0.3513   0.2613   0.1503   8.0235
#> 465:   -5.0111   0.3513   0.2613   0.1503   8.0219
#> 466:   -5.0111   0.3513   0.2613   0.1503   8.0212
#> 467:   -5.0111   0.3513   0.2614   0.1503   8.0232
#> 468:   -5.0110   0.3514   0.2613   0.1503   8.0237
#> 469:   -5.0111   0.3513   0.2613   0.1503   8.0245
#> 470:   -5.0110   0.3513   0.2614   0.1503   8.0265
#> 471:   -5.0110   0.3513   0.2614   0.1503   8.0291
#> 472:   -5.0110   0.3514   0.2614   0.1504   8.0308
#> 473:   -5.0110   0.3514   0.2614   0.1504   8.0326
#> 474:   -5.0110   0.3514   0.2614   0.1503   8.0329
#> 475:   -5.0110   0.3514   0.2614   0.1504   8.0334
#> 476:   -5.0110   0.3514   0.2615   0.1504   8.0352
#> 477:   -5.0110   0.3513   0.2615   0.1504   8.0356
#> 478:   -5.0110   0.3513   0.2615   0.1504   8.0346
#> 479:   -5.0111   0.3513   0.2615   0.1504   8.0339
#> 480:   -5.0111   0.3513   0.2616   0.1505   8.0344
#> 481:   -5.0111   0.3513   0.2616   0.1505   8.0341
#> 482:   -5.0111   0.3513   0.2616   0.1505   8.0340
#> 483:   -5.0111   0.3513   0.2616   0.1505   8.0350
#> 484:   -5.0111   0.3512   0.2615   0.1505   8.0347
#> 485:   -5.0111   0.3512   0.2615   0.1505   8.0336
#> 486:   -5.0111   0.3512   0.2616   0.1505   8.0310
#> 487:   -5.0111   0.3512   0.2615   0.1505   8.0304
#> 488:   -5.0111   0.3512   0.2615   0.1505   8.0325
#> 489:   -5.0112   0.3512   0.2616   0.1505   8.0322
#> 490:   -5.0111   0.3512   0.2616   0.1505   8.0327
#> 491:   -5.0111   0.3512   0.2617   0.1505   8.0325
#> 492:   -5.0111   0.3512   0.2617   0.1505   8.0328
#> 493:   -5.0111   0.3512   0.2617   0.1505   8.0347
#> 494:   -5.0111   0.3512   0.2616   0.1505   8.0353
#> 495:   -5.0111   0.3512   0.2616   0.1505   8.0360
#> 496:   -5.0111   0.3512   0.2616   0.1505   8.0375
#> 497:   -5.0111   0.3512   0.2616   0.1505   8.0390
#> 498:   -5.0111   0.3512   0.2616   0.1505   8.0389
#> 499:   -5.0111   0.3512   0.2616   0.1505   8.0401
#> 500:   -5.0111   0.3512   0.2615   0.1505   8.0418
#> Calculating covariance matrix
#> 
#> → creating full model...
#> → pruning branches (`if`/`else`)...
#> ✔ done
#> → loading into symengine environment...
#> ✔ done
#> → compiling EBE model...
#>  
#> ✔ done
#> Calculating residuals/tables
#> done

## Save output information into a word document
RxODE::.rxWithWd(tempdir(), # Put document in temporary directory
  nmDocx(fit.s)
)
#> Calculating -2LL by Gaussian quadrature (nnodes=3,nsd=1.6)
#> 
#> Warning: Unrecognized statistics: r.squared
#> Try setting `statistics` explicitly in the call to `huxreg()`
#> Error in read_xml.raw(charToRaw(enc2utf8(x)), "UTF-8", ..., as_html = as_html,     options = options): PCDATA invalid Char value 27 [9]

# }