Add table information to nlmixr fit object without tables

Add table information to nlmixr fit object without tables

addTable(
  object,
  updateObject = FALSE,
  data = object$dataSav,
  thetaEtaParameters = .foceiThetaEtaParameters(object),
  table = tableControl(),
  keep = NULL,
  drop = NULL,
  envir = parent.frame(1)
)

Arguments

object
nlmixr family of objects

updateObject
Update the object (default FALSE)

data
Saved data from

thetaEtaParameters
Intenral theta/eta parameters

table
a `tableControl()` list of options

keep
Character Vector of items to keep

drop
Character Vector of items to drop or NULL

envir
ENvironment to search for updating

Value

Fit with table information attached

Author

Matthew Fidler

Examples

# \donttest{

one.cmt <- function() {
  ini({
    ## You may label each parameter with a comment
    tka <- 0.45 # Log Ka
    tcl <- log(c(0, 2.7, 100)) # Log Cl
    ## This works with interactive models
    ## You may also label the preceding line with label("label text")
    tv <- 3.45; label("log V")
    ## the label("Label name") works with all models
    eta.ka ~ 0.6
    eta.cl ~ 0.3
    eta.v ~ 0.1
    add.sd <- 0.7
  })
  model({
    ka <- exp(tka + eta.ka)
    cl <- exp(tcl + eta.cl)
    v <- exp(tv + eta.v)
    linCmt() ~ add(add.sd)
  })
}

# run without tables step
f <- nlmixr(one.cmt, theo_sd, "saem", control=list(calcTables=FALSE))

#> ℹ parameter labels from comments will be replaced by 'label()'

#>  

#> 1:    0.1675   0.9587   3.4041   0.5700   0.2850   0.0950   5.4059
#> 2:    0.2395   0.9502   3.4687   0.5415   0.2707   0.0902   1.6722
#> 3:    0.4719   1.0190   3.4676   0.5144   0.2572   0.0857   1.0763
#> 4:    0.4872   1.0342   3.4673   0.4887   0.2444   0.0815   0.7795
#> 5:    0.4815   1.0271   3.4602   0.4685   0.2321   0.0774   0.6143
#> 6:    0.4811   0.9859   3.4657   0.4710   0.2205   0.0735   0.5919
#> 7:    0.4536   0.9806   3.4533   0.5364   0.2095   0.0698   0.5101
#> 8:    0.4212   0.9762   3.4546   0.5096   0.1990   0.0663   0.5289
#> 9:    0.4140   0.9708   3.4612   0.4841   0.1891   0.0630   0.5504
#> 10:    0.5108   1.0095   3.4592   0.4599   0.1796   0.0599   0.5489
#> 11:    0.5269   1.0055   3.4538   0.4904   0.1706   0.0569   0.5140
#> 12:    0.4669   1.0255   3.4571   0.4659   0.1621   0.0540   0.5176
#> 13:    0.4623   1.0226   3.4403   0.4734   0.1540   0.0513   0.5082
#> 14:    0.4842   1.0068   3.4406   0.4881   0.1463   0.0488   0.5144
#> 15:    0.4682   1.0181   3.4470   0.5050   0.1390   0.0463   0.4889
#> 16:    0.4441   1.0080   3.4359   0.4798   0.1320   0.0440   0.5018
#> 17:    0.4256   1.0253   3.4438   0.4558   0.1254   0.0418   0.4933
#> 18:    0.4374   1.0333   3.4481   0.4330   0.1192   0.0397   0.4825
#> 19:    0.4499   0.9956   3.4509   0.4114   0.1132   0.0377   0.5003
#> 20:    0.4125   1.0072   3.4375   0.3908   0.1075   0.0358   0.4953
#> 21:    0.4522   1.0316   3.4449   0.4756   0.1022   0.0341   0.4889
#> 22:    0.4979   1.0253   3.4567   0.4519   0.0971   0.0324   0.4771
#> 23:    0.4801   1.0172   3.4574   0.4427   0.0922   0.0307   0.4513
#> 24:    0.4742   1.0103   3.4533   0.4219   0.0876   0.0292   0.4622
#> 25:    0.4811   0.9926   3.4527   0.4284   0.0832   0.0277   0.4727
#> 26:    0.4502   1.0371   3.4301   0.4449   0.0792   0.0264   0.4584
#> 27:    0.4310   1.0297   3.4364   0.4824   0.0868   0.0250   0.4555
#> 28:    0.4805   1.0639   3.4503   0.5191   0.0824   0.0238   0.4670
#> 29:    0.4745   1.0252   3.4488   0.4932   0.0783   0.0249   0.4686
#> 30:    0.4517   1.0279   3.4430   0.4685   0.0762   0.0236   0.4882
#> 31:    0.4219   1.0297   3.4423   0.4451   0.0812   0.0224   0.4644
#> 32:    0.3745   1.0358   3.4362   0.4228   0.0772   0.0213   0.5047
#> 33:    0.4785   1.0386   3.4545   0.4659   0.0733   0.0226   0.4831
#> 34:    0.4766   1.0120   3.4592   0.4426   0.0696   0.0214   0.4836
#> 35:    0.4673   1.0058   3.4595   0.4205   0.0662   0.0204   0.4869
#> 36:    0.4585   1.0222   3.4547   0.3995   0.0665   0.0224   0.4707
#> 37:    0.4622   1.0200   3.4497   0.4537   0.0656   0.0249   0.4745
#> 38:    0.4814   1.0122   3.4459   0.4822   0.0771   0.0236   0.4747
#> 39:    0.4977   1.0486   3.4444   0.5120   0.0757   0.0225   0.4694
#> 40:    0.4677   1.0446   3.4461   0.4891   0.0719   0.0213   0.4884
#> 41:    0.4717   1.0364   3.4464   0.4646   0.0712   0.0203   0.4712
#> 42:    0.4263   1.0301   3.4484   0.4414   0.0710   0.0235   0.4701
#> 43:    0.4686   1.0017   3.4445   0.4741   0.0783   0.0223   0.4837
#> 44:    0.4615   0.9780   3.4755   0.4504   0.0744   0.0212   0.4861
#> 45:    0.4872   0.9915   3.4730   0.4423   0.0777   0.0201   0.4925
#> 46:    0.4637   1.0116   3.4618   0.4851   0.1086   0.0191   0.4599
#> 47:    0.5257   0.9930   3.4645   0.5623   0.1031   0.0182   0.4732
#> 48:    0.4661   1.0056   3.4547   0.5341   0.0980   0.0172   0.4769
#> 49:    0.4416   0.9987   3.4658   0.5074   0.0931   0.0207   0.4732
#> 50:    0.4457   0.9811   3.4628   0.4821   0.0884   0.0226   0.4890
#> 51:    0.4208   1.0105   3.4543   0.4580   0.0840   0.0225   0.4826
#> 52:    0.4845   1.0282   3.4603   0.4351   0.0798   0.0214   0.4740
#> 53:    0.5210   0.9775   3.4670   0.4133   0.0758   0.0203   0.4693
#> 54:    0.4663   0.9728   3.4671   0.3997   0.0734   0.0203   0.4613
#> 55:    0.4412   0.9856   3.4747   0.3797   0.0698   0.0206   0.4940
#> 56:    0.4764   1.0172   3.4678   0.3609   0.0684   0.0205   0.4733
#> 57:    0.4790   1.0266   3.4578   0.3430   0.0649   0.0219   0.4804
#> 58:    0.4717   0.9826   3.4743   0.3258   0.0617   0.0208   0.4919
#> 59:    0.4709   0.9748   3.4685   0.3487   0.0592   0.0206   0.5175
#> 60:    0.5032   0.9635   3.4796   0.4269   0.0562   0.0195   0.4984
#> 61:    0.4945   0.9877   3.4801   0.4056   0.0591   0.0185   0.4921
#> 62:    0.4869   0.9837   3.4823   0.3853   0.0562   0.0238   0.5109
#> 63:    0.4890   1.0013   3.4745   0.3837   0.0576   0.0226   0.5058
#> 64:    0.4907   0.9869   3.4669   0.4023   0.0561   0.0242   0.4868
#> 65:    0.4301   1.0151   3.4518   0.3861   0.0628   0.0243   0.4595
#> 66:    0.4929   1.0079   3.4520   0.4043   0.0634   0.0282   0.4633
#> 67:    0.4754   1.0287   3.4552   0.4112   0.0602   0.0268   0.4783
#> 68:    0.4444   1.0172   3.4486   0.3906   0.0572   0.0255   0.4962
#> 69:    0.4863   0.9892   3.4664   0.3933   0.0543   0.0254   0.5196
#> 70:    0.4607   0.9756   3.4695   0.4076   0.0516   0.0242   0.5063
#> 71:    0.4398   0.9776   3.4572   0.3872   0.0490   0.0230   0.5046
#> 72:    0.4229   1.0025   3.4505   0.3721   0.0518   0.0246   0.4942
#> 73:    0.4945   1.0053   3.4527   0.4854   0.0565   0.0234   0.4954
#> 74:    0.4394   0.9958   3.4696   0.4612   0.0666   0.0222   0.4619
#> 75:    0.5402   1.0033   3.4750   0.4520   0.0633   0.0211   0.4588
#> 76:    0.4953   1.0041   3.4767   0.4294   0.0753   0.0201   0.4723
#> 77:    0.5253   0.9983   3.4795   0.4079   0.0799   0.0191   0.4892
#> 78:    0.4836   0.9837   3.4634   0.3875   0.0759   0.0181   0.4806
#> 79:    0.4562   0.9985   3.4582   0.3681   0.0813   0.0172   0.4687
#> 80:    0.4603   0.9981   3.4636   0.3622   0.0772   0.0189   0.4529
#> 81:    0.4837   0.9916   3.4625   0.3529   0.0734   0.0206   0.4594
#> 82:    0.4780   1.0169   3.4657   0.4218   0.0785   0.0196   0.4876
#> 83:    0.4566   1.0015   3.4575   0.4007   0.0812   0.0186   0.5087
#> 84:    0.4280   1.0150   3.4402   0.3807   0.0771   0.0177   0.4982
#> 85:    0.4809   1.0299   3.4486   0.3626   0.0896   0.0168   0.5143
#> 86:    0.4600   1.0017   3.4502   0.3583   0.0947   0.0160   0.5015
#> 87:    0.4550   0.9975   3.4611   0.4178   0.0900   0.0161   0.4707
#> 88:    0.4483   1.0054   3.4448   0.4154   0.0874   0.0173   0.4653
#> 89:    0.4122   1.0141   3.4447   0.3947   0.0830   0.0169   0.4665
#> 90:    0.4234   1.0512   3.4297   0.3749   0.0799   0.0160   0.4573
#> 91:    0.4006   1.0182   3.4466   0.4411   0.0800   0.0152   0.4615
#> 92:    0.4354   1.0181   3.4453   0.4191   0.0763   0.0146   0.5104
#> 93:    0.4465   1.0312   3.4445   0.3981   0.0724   0.0169   0.4681
#> 94:    0.4009   1.0283   3.4431   0.3782   0.0688   0.0193   0.4753
#> 95:    0.4253   1.0348   3.4543   0.3627   0.0712   0.0205   0.4991
#> 96:    0.4362   1.0240   3.4325   0.4492   0.0690   0.0195   0.4686
#> 97:    0.4449   1.0226   3.4318   0.4267   0.0897   0.0185   0.4604
#> 98:    0.4535   1.0261   3.4440   0.4054   0.0852   0.0176   0.4596
#> 99:    0.4361   1.0307   3.4352   0.3859   0.0809   0.0181   0.4625
#> 100:    0.4428   1.0394   3.4436   0.3886   0.0866   0.0221   0.4558
#> 101:    0.4437   0.9935   3.4401   0.4292   0.0823   0.0210   0.4920
#> 102:    0.4467   1.0011   3.4497   0.4077   0.0782   0.0201   0.4677
#> 103:    0.4876   0.9860   3.4534   0.4724   0.0773   0.0191   0.4765
#> 104:    0.4183   1.0185   3.4439   0.4488   0.0839   0.0182   0.4674
#> 105:    0.4310   1.0121   3.4404   0.4264   0.0910   0.0173   0.4544
#> 106:    0.4815   1.0261   3.4552   0.4050   0.0865   0.0164   0.4517
#> 107:    0.4478   1.0224   3.4485   0.3915   0.0822   0.0156   0.4480
#> 108:    0.4849   1.0234   3.4520   0.4449   0.0781   0.0164   0.4719
#> 109:    0.4708   1.0262   3.4625   0.4227   0.0742   0.0155   0.4881
#> 110:    0.4361   1.0177   3.4652   0.4015   0.0747   0.0148   0.4828
#> 111:    0.4242   1.0123   3.4623   0.3815   0.0823   0.0145   0.5003
#> 112:    0.4638   1.0074   3.4536   0.3624   0.0806   0.0154   0.4738
#> 113:    0.4494   1.0052   3.4618   0.3605   0.0765   0.0185   0.4988
#> 114:    0.5142   1.0082   3.4773   0.3447   0.0727   0.0247   0.5055
#> 115:    0.5256   1.0081   3.4775   0.3698   0.0691   0.0242   0.5096
#> 116:    0.4846   0.9837   3.4741   0.3513   0.0656   0.0238   0.4916
#> 117:    0.4928   0.9869   3.4658   0.4264   0.0623   0.0226   0.4991
#> 118:    0.5485   1.0047   3.4659   0.4708   0.0601   0.0226   0.5149
#> 119:    0.4855   0.9855   3.4609   0.4472   0.0571   0.0245   0.5234
#> 120:    0.4977   0.9890   3.4810   0.4249   0.0689   0.0233   0.4868
#> 121:    0.4988   0.9864   3.4803   0.4036   0.0655   0.0221   0.5024
#> 122:    0.4619   1.0241   3.4505   0.3834   0.0716   0.0210   0.5085
#> 123:    0.4397   1.0198   3.4410   0.3913   0.0721   0.0205   0.4846
#> 124:    0.4362   1.0147   3.4418   0.4426   0.0685   0.0195   0.4524
#> 125:    0.4747   1.0005   3.4345   0.4373   0.0650   0.0185   0.4602
#> 126:    0.4484   1.0193   3.4392   0.4154   0.0723   0.0176   0.4770
#> 127:    0.4234   1.0376   3.4234   0.4089   0.0686   0.0167   0.4885
#> 128:    0.4266   1.0429   3.4319   0.3926   0.0951   0.0159   0.5080
#> 129:    0.4186   1.0238   3.4269   0.4838   0.0939   0.0176   0.4785
#> 130:    0.4291   1.0160   3.4368   0.4596   0.0892   0.0193   0.4575
#> 131:    0.4518   1.0137   3.4423   0.4366   0.0847   0.0195   0.4534
#> 132:    0.4679   0.9684   3.4607   0.4148   0.0805   0.0192   0.4672
#> 133:    0.5014   0.9833   3.4689   0.3991   0.0765   0.0182   0.4752
#> 134:    0.4860   0.9666   3.4804   0.3791   0.0726   0.0194   0.5035
#> 135:    0.4513   1.0004   3.4735   0.3834   0.0726   0.0184   0.4871
#> 136:    0.4542   1.0105   3.4549   0.4106   0.0769   0.0202   0.4558
#> 137:    0.4655   1.0138   3.4577   0.3900   0.0731   0.0215   0.4608
#> 138:    0.4810   0.9854   3.4574   0.4434   0.0694   0.0264   0.4828
#> 139:    0.4118   1.0218   3.4534   0.4213   0.0762   0.0279   0.4836
#> 140:    0.4510   1.0277   3.4511   0.4451   0.0888   0.0265   0.4981
#> 141:    0.4554   1.0146   3.4555   0.4514   0.0844   0.0259   0.4736
#> 142:    0.4749   1.0205   3.4512   0.4422   0.0802   0.0271   0.4479
#> 143:    0.4588   1.0278   3.4538   0.4438   0.0762   0.0259   0.4494
#> 144:    0.4410   1.0175   3.4430   0.4547   0.0724   0.0246   0.4761
#> 145:    0.4298   1.0060   3.4764   0.4320   0.0687   0.0234   0.5091
#> 146:    0.4647   1.0165   3.4533   0.4302   0.0722   0.0222   0.4973
#> 147:    0.4333   1.0386   3.4402   0.4087   0.0731   0.0211   0.4783
#> 148:    0.4341   1.0263   3.4421   0.4731   0.0834   0.0200   0.4575
#> 149:    0.5010   1.0203   3.4385   0.4495   0.0792   0.0190   0.4785
#> 150:    0.4978   1.0150   3.4651   0.4694   0.0797   0.0181   0.4617
#> 151:    0.4713   1.0062   3.4626   0.4550   0.0801   0.0195   0.4637
#> 152:    0.4665   1.0090   3.4520   0.3401   0.0713   0.0210   0.4786
#> 153:    0.4823   1.0147   3.4542   0.4249   0.0712   0.0190   0.4584
#> 154:    0.4709   0.9941   3.4600   0.3959   0.0771   0.0162   0.4716
#> 155:    0.4733   1.0215   3.4534   0.4202   0.0839   0.0144   0.4769
#> 156:    0.4640   1.0202   3.4494   0.3982   0.0714   0.0178   0.4769
#> 157:    0.4423   1.0216   3.4541   0.4335   0.0650   0.0176   0.4527
#> 158:    0.4591   1.0430   3.4436   0.4344   0.0677   0.0193   0.4803
#> 159:    0.4625   1.0088   3.4526   0.3655   0.0665   0.0172   0.4711
#> 160:    0.4500   0.9884   3.4495   0.3194   0.0564   0.0161   0.4978
#> 161:    0.4970   0.9487   3.4682   0.3584   0.0592   0.0165   0.5070
#> 162:    0.4785   0.9961   3.4706   0.3145   0.0662   0.0180   0.4989
#> 163:    0.5207   0.9990   3.4661   0.4326   0.0602   0.0222   0.4685
#> 164:    0.4918   0.9921   3.4743   0.5098   0.0528   0.0224   0.4760
#> 165:    0.4701   1.0199   3.4704   0.3995   0.0606   0.0229   0.4589
#> 166:    0.4749   0.9960   3.4734   0.3541   0.0643   0.0245   0.4928
#> 167:    0.4695   0.9885   3.4678   0.3757   0.0577   0.0199   0.4876
#> 168:    0.4485   1.0046   3.4520   0.3440   0.0509   0.0181   0.4836
#> 169:    0.4538   1.0155   3.4554   0.3423   0.0551   0.0191   0.4873
#> 170:    0.4649   1.0143   3.4611   0.3701   0.0587   0.0204   0.4751
#> 171:    0.4171   1.0137   3.4344   0.4381   0.0697   0.0141   0.4996
#> 172:    0.4379   1.0099   3.4506   0.3906   0.0689   0.0126   0.4830
#> 173:    0.4252   1.0410   3.4480   0.3538   0.0689   0.0132   0.4890
#> 174:    0.4536   1.0349   3.4385   0.3208   0.0680   0.0156   0.4887
#> 175:    0.4098   1.0207   3.4536   0.2971   0.0742   0.0179   0.5003
#> 176:    0.4141   1.0236   3.4522   0.3145   0.0758   0.0187   0.4843
#> 177:    0.4060   1.0168   3.4403   0.3902   0.0744   0.0180   0.4687
#> 178:    0.4013   1.0315   3.4321   0.3622   0.0743   0.0196   0.5137
#> 179:    0.4311   1.0359   3.4366   0.3193   0.0657   0.0209   0.4859
#> 180:    0.4135   0.9932   3.4488   0.3673   0.0563   0.0188   0.4759
#> 181:    0.4442   1.0029   3.4454   0.3281   0.0586   0.0198   0.4804
#> 182:    0.3903   1.0513   3.4341   0.3645   0.0629   0.0205   0.5063
#> 183:    0.4246   1.0229   3.4525   0.3377   0.0759   0.0200   0.4898
#> 184:    0.4149   1.0175   3.4507   0.3905   0.0689   0.0180   0.5066
#> 185:    0.3999   1.0275   3.4307   0.3428   0.0763   0.0217   0.4982
#> 186:    0.4306   1.0258   3.4358   0.3604   0.0647   0.0203   0.5046
#> 187:    0.4514   1.0111   3.4404   0.4139   0.0725   0.0229   0.4993
#> 188:    0.4389   1.0156   3.4600   0.3539   0.0726   0.0200   0.4850
#> 189:    0.4481   1.0221   3.4489   0.3812   0.0806   0.0186   0.4665
#> 190:    0.4599   0.9954   3.4412   0.4075   0.0651   0.0184   0.4674
#> 191:    0.4575   1.0171   3.4484   0.4020   0.0606   0.0187   0.4783
#> 192:    0.4433   1.0255   3.4377   0.4512   0.0694   0.0181   0.4685
#> 193:    0.3869   1.0341   3.4526   0.3912   0.0768   0.0165   0.4821
#> 194:    0.4273   1.0101   3.4426   0.4306   0.0932   0.0130   0.4754
#> 195:    0.4120   1.0098   3.4381   0.4540   0.0841   0.0129   0.4752
#> 196:    0.4416   1.0110   3.4267   0.5369   0.0790   0.0112   0.4673
#> 197:    0.4255   1.0332   3.4287   0.5670   0.0725   0.0110   0.5039
#> 198:    0.3788   1.0227   3.4260   0.4185   0.0836   0.0100   0.5177
#> 199:    0.4371   1.0365   3.4480   0.4365   0.0862   0.0085   0.5256
#> 200:    0.4271   1.0224   3.4327   0.3783   0.0842   0.0069   0.5312
#> 201:    0.4117   1.0292   3.4288   0.3798   0.0853   0.0075   0.5152
#> 202:    0.4076   1.0315   3.4289   0.3840   0.0877   0.0077   0.5080
#> 203:    0.3990   1.0393   3.4292   0.3926   0.0906   0.0077   0.5138
#> 204:    0.4065   1.0365   3.4314   0.3837   0.0897   0.0078   0.5125
#> 205:    0.4086   1.0346   3.4331   0.3764   0.0905   0.0078   0.5080
#> 206:    0.4137   1.0331   3.4342   0.3710   0.0894   0.0078   0.5052
#> 207:    0.4128   1.0315   3.4332   0.3767   0.0878   0.0081   0.5066
#> 208:    0.4139   1.0302   3.4336   0.3803   0.0860   0.0084   0.5017
#> 209:    0.4123   1.0299   3.4340   0.3778   0.0858   0.0091   0.5012
#> 210:    0.4126   1.0294   3.4350   0.3782   0.0854   0.0095   0.4968
#> 211:    0.4137   1.0286   3.4356   0.3779   0.0850   0.0099   0.4950
#> 212:    0.4154   1.0276   3.4364   0.3728   0.0855   0.0099   0.4953
#> 213:    0.4151   1.0270   3.4368   0.3702   0.0854   0.0101   0.4971
#> 214:    0.4149   1.0270   3.4366   0.3677   0.0847   0.0102   0.4986
#> 215:    0.4150   1.0265   3.4366   0.3688   0.0835   0.0102   0.4990
#> 216:    0.4145   1.0270   3.4363   0.3722   0.0825   0.0103   0.4981
#> 217:    0.4154   1.0272   3.4364   0.3755   0.0828   0.0104   0.4976
#> 218:    0.4167   1.0262   3.4368   0.3804   0.0824   0.0106   0.4974
#> 219:    0.4184   1.0248   3.4369   0.3810   0.0817   0.0107   0.4977
#> 220:    0.4192   1.0241   3.4373   0.3836   0.0817   0.0108   0.4982
#> 221:    0.4205   1.0249   3.4376   0.3845   0.0811   0.0108   0.4965
#> 222:    0.4210   1.0255   3.4379   0.3840   0.0808   0.0109   0.4949
#> 223:    0.4219   1.0254   3.4384   0.3834   0.0805   0.0111   0.4942
#> 224:    0.4239   1.0256   3.4391   0.3858   0.0800   0.0113   0.4934
#> 225:    0.4254   1.0245   3.4396   0.3852   0.0798   0.0114   0.4928
#> 226:    0.4281   1.0240   3.4397   0.3840   0.0799   0.0115   0.4932
#> 227:    0.4277   1.0248   3.4396   0.3821   0.0805   0.0116   0.4928
#> 228:    0.4267   1.0248   3.4391   0.3814   0.0807   0.0115   0.4934
#> 229:    0.4262   1.0248   3.4392   0.3800   0.0806   0.0114   0.4931
#> 230:    0.4263   1.0248   3.4396   0.3794   0.0805   0.0114   0.4931
#> 231:    0.4266   1.0251   3.4398   0.3798   0.0804   0.0115   0.4932
#> 232:    0.4278   1.0258   3.4399   0.3812   0.0802   0.0117   0.4935
#> 233:    0.4288   1.0257   3.4401   0.3811   0.0797   0.0117   0.4943
#> 234:    0.4288   1.0254   3.4402   0.3817   0.0796   0.0118   0.4943
#> 235:    0.4299   1.0255   3.4405   0.3821   0.0800   0.0119   0.4934
#> 236:    0.4321   1.0256   3.4408   0.3829   0.0805   0.0120   0.4926
#> 237:    0.4325   1.0256   3.4410   0.3818   0.0804   0.0120   0.4922
#> 238:    0.4328   1.0260   3.4411   0.3830   0.0804   0.0121   0.4921
#> 239:    0.4328   1.0260   3.4408   0.3839   0.0802   0.0122   0.4921
#> 240:    0.4326   1.0263   3.4405   0.3859   0.0798   0.0123   0.4919
#> 241:    0.4316   1.0271   3.4400   0.3851   0.0796   0.0123   0.4917
#> 242:    0.4313   1.0274   3.4398   0.3865   0.0796   0.0123   0.4914
#> 243:    0.4309   1.0268   3.4399   0.3882   0.0794   0.0124   0.4916
#> 244:    0.4314   1.0268   3.4400   0.3901   0.0793   0.0124   0.4916
#> 245:    0.4310   1.0269   3.4398   0.3901   0.0793   0.0125   0.4914
#> 246:    0.4312   1.0267   3.4401   0.3922   0.0791   0.0126   0.4915
#> 247:    0.4321   1.0267   3.4405   0.3935   0.0788   0.0127   0.4910
#> 248:    0.4334   1.0264   3.4409   0.3946   0.0788   0.0127   0.4904
#> 249:    0.4339   1.0262   3.4412   0.3937   0.0787   0.0127   0.4904
#> 250:    0.4351   1.0258   3.4416   0.3940   0.0786   0.0127   0.4904
#> 251:    0.4366   1.0253   3.4422   0.3935   0.0784   0.0127   0.4907
#> 252:    0.4376   1.0250   3.4425   0.3933   0.0782   0.0128   0.4910
#> 253:    0.4377   1.0251   3.4425   0.3931   0.0781   0.0127   0.4908
#> 254:    0.4383   1.0250   3.4424   0.3939   0.0782   0.0127   0.4909
#> 255:    0.4388   1.0248   3.4425   0.3928   0.0783   0.0127   0.4909
#> 256:    0.4384   1.0244   3.4427   0.3921   0.0787   0.0127   0.4906
#> 257:    0.4391   1.0237   3.4430   0.3929   0.0789   0.0127   0.4900
#> 258:    0.4388   1.0233   3.4433   0.3922   0.0792   0.0127   0.4900
#> 259:    0.4389   1.0226   3.4435   0.3926   0.0794   0.0127   0.4900
#> 260:    0.4387   1.0224   3.4435   0.3927   0.0792   0.0127   0.4900
#> 261:    0.4390   1.0222   3.4436   0.3934   0.0790   0.0128   0.4898
#> 262:    0.4390   1.0223   3.4439   0.3925   0.0788   0.0128   0.4901
#> 263:    0.4388   1.0222   3.4439   0.3937   0.0789   0.0129   0.4901
#> 264:    0.4393   1.0226   3.4441   0.3951   0.0789   0.0129   0.4902
#> 265:    0.4395   1.0222   3.4442   0.3950   0.0791   0.0129   0.4901
#> 266:    0.4392   1.0222   3.4441   0.3951   0.0790   0.0129   0.4898
#> 267:    0.4388   1.0218   3.4440   0.3947   0.0792   0.0130   0.4901
#> 268:    0.4385   1.0219   3.4439   0.3946   0.0792   0.0130   0.4898
#> 269:    0.4387   1.0216   3.4439   0.3943   0.0792   0.0130   0.4896
#> 270:    0.4379   1.0218   3.4439   0.3951   0.0792   0.0130   0.4895
#> 271:    0.4375   1.0216   3.4439   0.3962   0.0791   0.0131   0.4891
#> 272:    0.4375   1.0217   3.4438   0.3968   0.0792   0.0132   0.4888
#> 273:    0.4377   1.0217   3.4440   0.3972   0.0792   0.0132   0.4888
#> 274:    0.4381   1.0213   3.4441   0.3975   0.0792   0.0132   0.4882
#> 275:    0.4381   1.0208   3.4441   0.3971   0.0792   0.0133   0.4881
#> 276:    0.4381   1.0204   3.4441   0.3972   0.0791   0.0134   0.4882
#> 277:    0.4385   1.0208   3.4441   0.3976   0.0790   0.0134   0.4881
#> 278:    0.4387   1.0209   3.4440   0.3979   0.0789   0.0135   0.4877
#> 279:    0.4385   1.0212   3.4440   0.3971   0.0790   0.0135   0.4876
#> 280:    0.4386   1.0212   3.4439   0.3976   0.0791   0.0136   0.4871
#> 281:    0.4383   1.0214   3.4438   0.3978   0.0790   0.0136   0.4872
#> 282:    0.4377   1.0218   3.4437   0.3980   0.0789   0.0136   0.4871
#> 283:    0.4372   1.0219   3.4434   0.3979   0.0789   0.0136   0.4873
#> 284:    0.4368   1.0217   3.4434   0.3986   0.0789   0.0136   0.4872
#> 285:    0.4367   1.0214   3.4434   0.3987   0.0790   0.0136   0.4867
#> 286:    0.4366   1.0211   3.4434   0.3988   0.0789   0.0136   0.4867
#> 287:    0.4372   1.0209   3.4435   0.3988   0.0789   0.0136   0.4867
#> 288:    0.4374   1.0207   3.4436   0.3992   0.0790   0.0136   0.4867
#> 289:    0.4373   1.0206   3.4435   0.3991   0.0790   0.0136   0.4868
#> 290:    0.4367   1.0208   3.4432   0.3991   0.0791   0.0136   0.4868
#> 291:    0.4366   1.0213   3.4431   0.3990   0.0792   0.0136   0.4864
#> 292:    0.4367   1.0217   3.4430   0.3992   0.0791   0.0135   0.4867
#> 293:    0.4365   1.0218   3.4429   0.3990   0.0791   0.0135   0.4867
#> 294:    0.4365   1.0218   3.4428   0.3984   0.0790   0.0135   0.4871
#> 295:    0.4362   1.0219   3.4428   0.3981   0.0790   0.0135   0.4871
#> 296:    0.4364   1.0217   3.4428   0.3984   0.0790   0.0135   0.4872
#> 297:    0.4367   1.0218   3.4429   0.3990   0.0790   0.0136   0.4872
#> 298:    0.4368   1.0218   3.4429   0.3984   0.0790   0.0136   0.4871
#> 299:    0.4370   1.0215   3.4431   0.3979   0.0790   0.0136   0.4869
#> 300:    0.4372   1.0213   3.4431   0.3981   0.0791   0.0136   0.4868
#> 301:    0.4373   1.0212   3.4432   0.3985   0.0790   0.0137   0.4867
#> 302:    0.4372   1.0211   3.4434   0.3985   0.0788   0.0137   0.4866
#> 303:    0.4376   1.0209   3.4436   0.3988   0.0786   0.0137   0.4864
#> 304:    0.4376   1.0208   3.4437   0.3992   0.0785   0.0138   0.4864
#> 305:    0.4375   1.0207   3.4438   0.3991   0.0785   0.0138   0.4863
#> 306:    0.4371   1.0204   3.4439   0.3985   0.0783   0.0138   0.4862
#> 307:    0.4371   1.0203   3.4439   0.3984   0.0782   0.0138   0.4864
#> 308:    0.4373   1.0203   3.4441   0.3982   0.0780   0.0138   0.4865
#> 309:    0.4375   1.0204   3.4442   0.3982   0.0779   0.0138   0.4862
#> 310:    0.4376   1.0205   3.4443   0.3981   0.0777   0.0139   0.4862
#> 311:    0.4376   1.0204   3.4444   0.3981   0.0776   0.0139   0.4861
#> 312:    0.4378   1.0204   3.4445   0.3979   0.0776   0.0139   0.4861
#> 313:    0.4380   1.0204   3.4446   0.3976   0.0775   0.0139   0.4860
#> 314:    0.4383   1.0205   3.4446   0.3980   0.0776   0.0139   0.4858
#> 315:    0.4386   1.0205   3.4446   0.3986   0.0775   0.0139   0.4857
#> 316:    0.4388   1.0205   3.4446   0.3987   0.0775   0.0139   0.4857
#> 317:    0.4389   1.0206   3.4445   0.3989   0.0775   0.0139   0.4857
#> 318:    0.4387   1.0206   3.4443   0.3990   0.0776   0.0139   0.4858
#> 319:    0.4384   1.0206   3.4443   0.3990   0.0776   0.0139   0.4857
#> 320:    0.4383   1.0208   3.4443   0.3993   0.0777   0.0139   0.4857
#> 321:    0.4383   1.0207   3.4442   0.3994   0.0777   0.0139   0.4857
#> 322:    0.4379   1.0207   3.4442   0.3992   0.0777   0.0139   0.4858
#> 323:    0.4380   1.0208   3.4442   0.3993   0.0777   0.0139   0.4860
#> 324:    0.4381   1.0211   3.4442   0.3995   0.0776   0.0139   0.4858
#> 325:    0.4383   1.0213   3.4443   0.3997   0.0775   0.0139   0.4858
#> 326:    0.4382   1.0213   3.4442   0.3994   0.0774   0.0140   0.4857
#> 327:    0.4386   1.0214   3.4442   0.4002   0.0773   0.0140   0.4857
#> 328:    0.4389   1.0214   3.4443   0.4004   0.0774   0.0140   0.4856
#> 329:    0.4391   1.0214   3.4443   0.4006   0.0775   0.0140   0.4856
#> 330:    0.4394   1.0216   3.4442   0.4009   0.0775   0.0140   0.4856
#> 331:    0.4397   1.0213   3.4443   0.4011   0.0773   0.0140   0.4857
#> 332:    0.4397   1.0210   3.4444   0.4008   0.0773   0.0141   0.4858
#> 333:    0.4400   1.0208   3.4445   0.4010   0.0773   0.0141   0.4858
#> 334:    0.4399   1.0206   3.4445   0.4007   0.0773   0.0141   0.4861
#> 335:    0.4401   1.0207   3.4445   0.4005   0.0774   0.0141   0.4861
#> 336:    0.4401   1.0207   3.4445   0.4004   0.0773   0.0141   0.4862
#> 337:    0.4402   1.0207   3.4445   0.4004   0.0772   0.0141   0.4864
#> 338:    0.4402   1.0207   3.4444   0.4009   0.0772   0.0141   0.4865
#> 339:    0.4401   1.0207   3.4444   0.4014   0.0771   0.0141   0.4865
#> 340:    0.4401   1.0208   3.4443   0.4020   0.0771   0.0141   0.4863
#> 341:    0.4402   1.0208   3.4443   0.4022   0.0770   0.0142   0.4863
#> 342:    0.4404   1.0207   3.4442   0.4022   0.0770   0.0142   0.4863
#> 343:    0.4406   1.0206   3.4442   0.4023   0.0770   0.0142   0.4862
#> 344:    0.4405   1.0206   3.4442   0.4028   0.0771   0.0142   0.4863
#> 345:    0.4404   1.0205   3.4441   0.4034   0.0770   0.0142   0.4863
#> 346:    0.4405   1.0206   3.4442   0.4037   0.0770   0.0142   0.4863
#> 347:    0.4406   1.0206   3.4442   0.4043   0.0770   0.0142   0.4863
#> 348:    0.4406   1.0206   3.4442   0.4047   0.0771   0.0142   0.4865
#> 349:    0.4410   1.0205   3.4443   0.4047   0.0770   0.0143   0.4865
#> 350:    0.4413   1.0204   3.4444   0.4046   0.0770   0.0142   0.4866
#> 351:    0.4415   1.0203   3.4444   0.4044   0.0770   0.0142   0.4864
#> 352:    0.4416   1.0203   3.4445   0.4042   0.0770   0.0142   0.4864
#> 353:    0.4417   1.0204   3.4446   0.4038   0.0770   0.0142   0.4863
#> 354:    0.4414   1.0204   3.4446   0.4033   0.0770   0.0142   0.4864
#> 355:    0.4414   1.0204   3.4446   0.4032   0.0770   0.0142   0.4865
#> 356:    0.4414   1.0206   3.4446   0.4030   0.0770   0.0142   0.4865
#> 357:    0.4414   1.0206   3.4446   0.4030   0.0770   0.0143   0.4865
#> 358:    0.4414   1.0207   3.4445   0.4031   0.0771   0.0143   0.4866
#> 359:    0.4415   1.0206   3.4445   0.4029   0.0770   0.0143   0.4866
#> 360:    0.4416   1.0206   3.4445   0.4026   0.0770   0.0143   0.4865
#> 361:    0.4414   1.0205   3.4445   0.4027   0.0769   0.0143   0.4863
#> 362:    0.4415   1.0206   3.4445   0.4028   0.0768   0.0143   0.4861
#> 363:    0.4417   1.0204   3.4446   0.4027   0.0768   0.0143   0.4861
#> 364:    0.4416   1.0202   3.4446   0.4024   0.0767   0.0143   0.4862
#> 365:    0.4415   1.0202   3.4445   0.4024   0.0766   0.0143   0.4862
#> 366:    0.4417   1.0202   3.4445   0.4023   0.0766   0.0144   0.4860
#> 367:    0.4417   1.0200   3.4445   0.4025   0.0766   0.0144   0.4860
#> 368:    0.4420   1.0199   3.4446   0.4028   0.0767   0.0144   0.4861
#> 369:    0.4421   1.0200   3.4447   0.4029   0.0768   0.0144   0.4860
#> 370:    0.4421   1.0201   3.4446   0.4027   0.0768   0.0143   0.4860
#> 371:    0.4419   1.0201   3.4445   0.4028   0.0768   0.0143   0.4859
#> 372:    0.4420   1.0201   3.4445   0.4031   0.0767   0.0144   0.4857
#> 373:    0.4421   1.0200   3.4446   0.4033   0.0767   0.0144   0.4856
#> 374:    0.4422   1.0199   3.4446   0.4034   0.0766   0.0144   0.4857
#> 375:    0.4420   1.0198   3.4446   0.4032   0.0766   0.0144   0.4856
#> 376:    0.4420   1.0198   3.4446   0.4032   0.0766   0.0143   0.4855
#> 377:    0.4419   1.0199   3.4445   0.4029   0.0765   0.0144   0.4855
#> 378:    0.4417   1.0201   3.4445   0.4028   0.0764   0.0144   0.4856
#> 379:    0.4417   1.0201   3.4445   0.4030   0.0764   0.0144   0.4855
#> 380:    0.4415   1.0202   3.4444   0.4031   0.0765   0.0144   0.4856
#> 381:    0.4414   1.0202   3.4444   0.4034   0.0766   0.0144   0.4856
#> 382:    0.4415   1.0202   3.4444   0.4036   0.0766   0.0144   0.4856
#> 383:    0.4416   1.0203   3.4445   0.4037   0.0767   0.0144   0.4856
#> 384:    0.4416   1.0203   3.4444   0.4037   0.0767   0.0144   0.4859
#> 385:    0.4415   1.0203   3.4444   0.4035   0.0767   0.0144   0.4860
#> 386:    0.4414   1.0203   3.4444   0.4036   0.0768   0.0144   0.4861
#> 387:    0.4414   1.0203   3.4443   0.4040   0.0769   0.0144   0.4860
#> 388:    0.4414   1.0202   3.4443   0.4044   0.0769   0.0144   0.4860
#> 389:    0.4413   1.0202   3.4443   0.4046   0.0770   0.0144   0.4860
#> 390:    0.4413   1.0204   3.4442   0.4044   0.0770   0.0144   0.4862
#> 391:    0.4413   1.0206   3.4442   0.4039   0.0771   0.0144   0.4862
#> 392:    0.4414   1.0207   3.4443   0.4037   0.0771   0.0144   0.4864
#> 393:    0.4414   1.0208   3.4443   0.4037   0.0771   0.0144   0.4864
#> 394:    0.4413   1.0208   3.4443   0.4040   0.0770   0.0145   0.4865
#> 395:    0.4413   1.0208   3.4443   0.4040   0.0770   0.0145   0.4865
#> 396:    0.4415   1.0206   3.4445   0.4041   0.0770   0.0145   0.4866
#> 397:    0.4416   1.0206   3.4445   0.4043   0.0769   0.0145   0.4867
#> 398:    0.4416   1.0205   3.4446   0.4044   0.0769   0.0145   0.4868
#> 399:    0.4418   1.0205   3.4447   0.4044   0.0768   0.0145   0.4869
#> 400:    0.4419   1.0205   3.4447   0.4047   0.0769   0.0145   0.4869
#> 401:    0.4421   1.0205   3.4448   0.4049   0.0769   0.0145   0.4868
#> 402:    0.4421   1.0204   3.4448   0.4048   0.0770   0.0145   0.4867
#> 403:    0.4422   1.0204   3.4449   0.4050   0.0771   0.0145   0.4867
#> 404:    0.4425   1.0203   3.4450   0.4051   0.0771   0.0146   0.4867
#> 405:    0.4426   1.0202   3.4451   0.4054   0.0771   0.0146   0.4868
#> 406:    0.4432   1.0201   3.4451   0.4059   0.0771   0.0146   0.4870
#> 407:    0.4432   1.0202   3.4450   0.4062   0.0770   0.0146   0.4870
#> 408:    0.4433   1.0201   3.4450   0.4062   0.0770   0.0146   0.4869
#> 409:    0.4433   1.0201   3.4451   0.4060   0.0769   0.0146   0.4869
#> 410:    0.4434   1.0201   3.4451   0.4059   0.0769   0.0146   0.4869
#> 411:    0.4435   1.0201   3.4451   0.4055   0.0770   0.0145   0.4870
#> 412:    0.4438   1.0200   3.4451   0.4055   0.0771   0.0145   0.4871
#> 413:    0.4438   1.0200   3.4452   0.4052   0.0771   0.0145   0.4873
#> 414:    0.4439   1.0199   3.4453   0.4050   0.0771   0.0145   0.4873
#> 415:    0.4442   1.0198   3.4454   0.4051   0.0772   0.0145   0.4874
#> 416:    0.4443   1.0197   3.4455   0.4053   0.0772   0.0145   0.4875
#> 417:    0.4444   1.0197   3.4455   0.4053   0.0772   0.0145   0.4876
#> 418:    0.4445   1.0198   3.4455   0.4055   0.0772   0.0145   0.4877
#> 419:    0.4447   1.0198   3.4456   0.4057   0.0773   0.0145   0.4877
#> 420:    0.4448   1.0199   3.4456   0.4058   0.0772   0.0145   0.4877
#> 421:    0.4447   1.0199   3.4456   0.4060   0.0772   0.0145   0.4877
#> 422:    0.4448   1.0199   3.4456   0.4062   0.0772   0.0145   0.4878
#> 423:    0.4447   1.0199   3.4457   0.4064   0.0772   0.0145   0.4878
#> 424:    0.4447   1.0199   3.4457   0.4066   0.0772   0.0145   0.4879
#> 425:    0.4446   1.0198   3.4456   0.4064   0.0773   0.0145   0.4879
#> 426:    0.4444   1.0198   3.4456   0.4063   0.0773   0.0146   0.4879
#> 427:    0.4442   1.0198   3.4456   0.4064   0.0772   0.0146   0.4879
#> 428:    0.4441   1.0198   3.4455   0.4066   0.0771   0.0146   0.4879
#> 429:    0.4441   1.0198   3.4455   0.4069   0.0771   0.0146   0.4879
#> 430:    0.4441   1.0199   3.4455   0.4072   0.0771   0.0146   0.4879
#> 431:    0.4441   1.0200   3.4455   0.4073   0.0770   0.0146   0.4880
#> 432:    0.4442   1.0200   3.4455   0.4073   0.0770   0.0146   0.4880
#> 433:    0.4443   1.0202   3.4456   0.4072   0.0770   0.0146   0.4881
#> 434:    0.4442   1.0202   3.4455   0.4073   0.0769   0.0146   0.4881
#> 435:    0.4442   1.0202   3.4454   0.4074   0.0769   0.0146   0.4881
#> 436:    0.4443   1.0202   3.4454   0.4075   0.0769   0.0146   0.4882
#> 437:    0.4443   1.0201   3.4454   0.4074   0.0769   0.0146   0.4884
#> 438:    0.4443   1.0200   3.4455   0.4071   0.0770   0.0146   0.4885
#> 439:    0.4444   1.0198   3.4455   0.4068   0.0770   0.0146   0.4886
#> 440:    0.4443   1.0197   3.4456   0.4065   0.0770   0.0146   0.4886
#> 441:    0.4441   1.0197   3.4456   0.4063   0.0770   0.0146   0.4886
#> 442:    0.4442   1.0197   3.4456   0.4062   0.0770   0.0146   0.4886
#> 443:    0.4443   1.0197   3.4456   0.4062   0.0771   0.0146   0.4887
#> 444:    0.4444   1.0195   3.4456   0.4062   0.0771   0.0146   0.4889
#> 445:    0.4444   1.0194   3.4456   0.4061   0.0771   0.0146   0.4890
#> 446:    0.4445   1.0194   3.4456   0.4059   0.0771   0.0146   0.4890
#> 447:    0.4447   1.0194   3.4457   0.4059   0.0771   0.0146   0.4889
#> 448:    0.4449   1.0193   3.4458   0.4060   0.0771   0.0146   0.4888
#> 449:    0.4449   1.0193   3.4458   0.4061   0.0772   0.0146   0.4888
#> 450:    0.4451   1.0192   3.4459   0.4058   0.0772   0.0146   0.4889
#> 451:    0.4454   1.0190   3.4460   0.4058   0.0772   0.0146   0.4889
#> 452:    0.4455   1.0189   3.4461   0.4054   0.0772   0.0146   0.4890
#> 453:    0.4456   1.0189   3.4461   0.4055   0.0772   0.0146   0.4890
#> 454:    0.4457   1.0189   3.4462   0.4054   0.0772   0.0146   0.4889
#> 455:    0.4459   1.0189   3.4462   0.4055   0.0772   0.0146   0.4888
#> 456:    0.4458   1.0189   3.4462   0.4055   0.0772   0.0146   0.4888
#> 457:    0.4457   1.0189   3.4462   0.4054   0.0772   0.0146   0.4889
#> 458:    0.4457   1.0190   3.4462   0.4052   0.0772   0.0146   0.4889
#> 459:    0.4457   1.0189   3.4463   0.4051   0.0772   0.0146   0.4889
#> 460:    0.4456   1.0188   3.4463   0.4048   0.0772   0.0146   0.4889
#> 461:    0.4456   1.0188   3.4463   0.4048   0.0771   0.0146   0.4889
#> 462:    0.4458   1.0188   3.4463   0.4047   0.0771   0.0146   0.4889
#> 463:    0.4458   1.0189   3.4463   0.4046   0.0771   0.0146   0.4888
#> 464:    0.4460   1.0188   3.4464   0.4047   0.0771   0.0146   0.4888
#> 465:    0.4461   1.0187   3.4465   0.4047   0.0771   0.0146   0.4887
#> 466:    0.4461   1.0185   3.4466   0.4046   0.0771   0.0146   0.4887
#> 467:    0.4463   1.0184   3.4466   0.4046   0.0770   0.0146   0.4887
#> 468:    0.4463   1.0183   3.4467   0.4043   0.0770   0.0147   0.4888
#> 469:    0.4463   1.0184   3.4467   0.4042   0.0770   0.0147   0.4887
#> 470:    0.4463   1.0185   3.4467   0.4042   0.0770   0.0147   0.4887
#> 471:    0.4462   1.0185   3.4467   0.4040   0.0769   0.0147   0.4887
#> 472:    0.4460   1.0185   3.4466   0.4038   0.0769   0.0147   0.4888
#> 473:    0.4460   1.0185   3.4466   0.4039   0.0769   0.0147   0.4888
#> 474:    0.4460   1.0186   3.4465   0.4042   0.0769   0.0147   0.4888
#> 475:    0.4459   1.0185   3.4465   0.4042   0.0770   0.0147   0.4888
#> 476:    0.4459   1.0185   3.4465   0.4044   0.0770   0.0147   0.4888
#> 477:    0.4458   1.0184   3.4465   0.4045   0.0769   0.0147   0.4887
#> 478:    0.4457   1.0185   3.4465   0.4045   0.0770   0.0147   0.4887
#> 479:    0.4456   1.0185   3.4465   0.4044   0.0770   0.0147   0.4887
#> 480:    0.4457   1.0184   3.4466   0.4044   0.0770   0.0147   0.4886
#> 481:    0.4457   1.0183   3.4466   0.4041   0.0770   0.0147   0.4886
#> 482:    0.4458   1.0183   3.4466   0.4041   0.0770   0.0147   0.4885
#> 483:    0.4460   1.0183   3.4466   0.4040   0.0771   0.0147   0.4885
#> 484:    0.4460   1.0183   3.4466   0.4040   0.0771   0.0147   0.4884
#> 485:    0.4461   1.0183   3.4467   0.4042   0.0770   0.0147   0.4884
#> 486:    0.4462   1.0183   3.4467   0.4044   0.0770   0.0147   0.4885
#> 487:    0.4463   1.0182   3.4467   0.4049   0.0770   0.0147   0.4885
#> 488:    0.4466   1.0181   3.4467   0.4053   0.0770   0.0147   0.4885
#> 489:    0.4467   1.0182   3.4468   0.4055   0.0770   0.0148   0.4885
#> 490:    0.4468   1.0182   3.4468   0.4055   0.0770   0.0148   0.4884
#> 491:    0.4469   1.0182   3.4469   0.4054   0.0770   0.0148   0.4885
#> 492:    0.4471   1.0182   3.4469   0.4054   0.0770   0.0148   0.4885
#> 493:    0.4473   1.0183   3.4469   0.4053   0.0769   0.0148   0.4884
#> 494:    0.4472   1.0183   3.4469   0.4049   0.0769   0.0148   0.4884
#> 495:    0.4471   1.0182   3.4469   0.4046   0.0769   0.0148   0.4884
#> 496:    0.4471   1.0183   3.4469   0.4044   0.0769   0.0148   0.4884
#> 497:    0.4471   1.0183   3.4469   0.4042   0.0769   0.0148   0.4883
#> 498:    0.4470   1.0182   3.4468   0.4043   0.0768   0.0148   0.4883
#> 499:    0.4471   1.0182   3.4469   0.4046   0.0768   0.0148   0.4881
#> 500:    0.4472   1.0182   3.4469   0.4044   0.0768   0.0148   0.4880

#> Calculating covariance matrix

#> 

#> Warning: Bounds are ignored in SAEM

print(f)

#> ── nlmixr SAEM(Solved); OBJF not calculated fit ────────────────────────────────
#> 
#>  Gaussian/Laplacian Likelihoods: AIC() or $objf etc. 
#>  FOCEi CWRES & Likelihoods: addCwres() 
#> 
#> ── Time (sec $time): ───────────────────────────────────────────────────────────
#> 
#>          saem   .time covariance   other
#> elapsed 1.125 0.30507      0.015 0.47493
#> 
#> ── Population Parameters ($parFixed or $parFixedDf): ───────────────────────────
#> 
#>        Parameter  Est.     SE %RSE Back-transformed(95%CI) BSV(CV%) Shrink(SD)%
#> tka       Log Ka 0.447  0.194 43.4       1.56 (1.07, 2.29)     70.6    -0.900% 
#> tcl       Log Cl  1.02  0.087 8.54       2.77 (2.33, 3.28)     28.3      4.17% 
#> tv         log V  3.45 0.0426 1.24       31.4 (28.9, 34.1)     12.2      9.77% 
#> add.sd           0.699                               0.699                     
#>  
#>   Covariance Type ($covMethod): linFim
#>   Fixed parameter correlations in $cor
#>   No correlations in between subject variability (BSV) matrix
#>   Full BSV covariance ($omega) or correlation ($omegaR; diagonals=SDs) 
#>   Distribution stats (mean/skewness/kurtosis/p-value) available in $shrink 

# Now add the tables

f <- addTable(f)

#> Calculating residuals/tables

#> done

print(f)

#> ── nlmixr SAEM(Solved); OBJF not calculated fit ────────────────────────────────
#> 
#>  Gaussian/Laplacian Likelihoods: AIC() or $objf etc. 
#>  FOCEi CWRES & Likelihoods: addCwres() 
#> 
#> ── Time (sec $time): ───────────────────────────────────────────────────────────
#> 
#>          saem   .time covariance   other table
#> elapsed 1.125 0.30507      0.015 0.47493  0.01
#> 
#> ── Population Parameters ($parFixed or $parFixedDf): ───────────────────────────
#> 
#>        Parameter  Est.     SE %RSE Back-transformed(95%CI) BSV(CV%) Shrink(SD)%
#> tka       Log Ka 0.447  0.194 43.4       1.56 (1.07, 2.29)     70.6    -0.900% 
#> tcl       Log Cl  1.02  0.087 8.54       2.77 (2.33, 3.28)     28.3      4.17% 
#> tv         log V  3.45 0.0426 1.24       31.4 (28.9, 34.1)     12.2      9.77% 
#> add.sd           0.699                               0.699                     
#>  
#>   Covariance Type ($covMethod): linFim
#>   Fixed parameter correlations in $cor
#>   No correlations in between subject variability (BSV) matrix
#>   Full BSV covariance ($omega) or correlation ($omegaR; diagonals=SDs) 
#>   Distribution stats (mean/skewness/kurtosis/p-value) available in $shrink 
#> 
#> ── Fit Data (object is a modified tibble): ─────────────────────────────────────
#> # A tibble: 132 x 16
#>   ID     TIME    DV  PRED    RES IPRED   IRES  IWRES eta.ka eta.cl   eta.v    ka
#>   <fct> <dbl> <dbl> <dbl>  <dbl> <dbl>  <dbl>  <dbl>  <dbl>  <dbl>   <dbl> <dbl>
#> 1 1     0      0.74  0     0.74   0     0.74   1.06   0.105 -0.481 -0.0783  1.74
#> 2 1     0.25   2.84  3.26 -0.419  3.85 -1.01  -1.45   0.105 -0.481 -0.0783  1.74
#> 3 1     0.570  6.57  5.84  0.729  6.79 -0.221 -0.316  0.105 -0.481 -0.0783  1.74
#> # … with 129 more rows, and 4 more variables: cl <dbl>, v <dbl>, tad <dbl>,
#> #   dosenum <dbl>

# }