Configure an SAEM model

Configure an SAEM model by generating an input list to the SAEM model function

configsaem(
  model,
  data,
  inits,
  mcmc = list(niter = c(200, 300), nmc = 3, nu = c(2, 2, 2)),
  ODEopt = list(atol = 1e-06, rtol = 1e-04, method = "lsoda", transitAbs = FALSE,
    maxeval = 1e+05),
  distribution = c("normal", "poisson", "binomial"),
  addProp = c("combined2", "combined1"),
  seed = 99,
  fixed = NULL,
  DEBUG = 0,
  tol = 1e-04,
  itmax = 100L,
  type = c("nelder-mead", "newuoa"),
  lambdaRange = 3,
  powRange = 10
)

Arguments

model
a compiled saem model by gen_saem_user_fn()

data
input data

inits
initial values

mcmc
a list of various mcmc options

ODEopt
optional ODE solving options

distribution
one of c("normal","poisson","binomial")

addProp
one of "combined1" and "combined2"; These are the two forms of additive+proportional errors supported by monolix/nonmem:

combined1: transform(y)=transform(f)+(a+b*f^c)*eps

combined2: transform(y)=transform(f)+(a^2+b^2*f^(2c))*eps

seed
seed for random number generator

fixed
a character vector of fixed effect only parameters (no random effects attached) to be fixed

DEBUG
Integer determining if debugging is enabled

tol
This is the tolerance for the regression models used for complex residual errors (ie add+prop etc)

itmax
This is the maximum number of iterations for the regression models used for complex residual errors. The number of iterations is itmax*number of parameters

type
indicates the type of optimization for the residuals; Can be one of c("nelder-mead", "newuoa")

lambdaRange
This indicates the range that Box-Cox and Yeo-Johnson parameters are constrained to be; The default is 3 indicating the range (-3,3)

powRange
This indicates the range that powers can take for residual errors; By default this is 10 indicating the range is c(1/10, 10) or c(0.1,10)

Details

Fit a generalized nonlinear mixed-effect model by he Stochastic Approximation Expectation-Maximization (SAEM) algorithm

Author

Wenping Wang & Matthew Fidler

Examples

# \donttest{

# In this ODE system we simply specify the ODEs

ode <- "d/dt(depot) =-KA*depot;
        d/dt(centr) = KA*depot - KE*centr;"
m1 <- RxODE(ode, modName = "m1")

#>  

# In this ode System, we also specify the concentration as C2 = centr/V

ode <- "C2 = centr/V;
      d/dt(depot) =-KA*depot;
      d/dt(centr) = KA*depot - KE*centr;"
m2 = RxODE(ode, modName="m2")

#>  

PKpars <- function() {
  CL <- exp(lCL)
  V <- exp(lV)
  KA <- exp(lKA)
  KE <- CL / V
}

PRED <- function() centr / V
PRED2 <- function() C2

saem_fit <- gen_saem_user_fn(model = m1, PKpars, pred = PRED)

#>  

#> Warning: this was not generated from nlmixr, loading into symengine

#> → creating SAEM model...

#> → pruning branches (`if`/`else`)...

#> ✔ done

#> → loading into symengine environment...

#> ✔ done

# Can also use PRED2
saem_fit <- gen_saem_user_fn(model=m2, PKpars, pred=PRED2)

#>  

#> Warning: this was not generated from nlmixr, loading into symengine

#> → creating SAEM model...

#> → pruning branches (`if`/`else`)...

#> ✔ done

#> → loading into symengine environment...

#> ✔ done

# You can also use the nlmixr UI to run this model and call the lower level functions

one.compartment <- function() {
ini({
  tka <- 0.45 # Log Ka
  tcl <- 1 # Log Cl
  tv <- 3.45    # Log V
  eta.ka ~ 0.6
  eta.cl ~ 0.3
  eta.v ~ 0.1
  add.sd <- 0.7
  wt.est <- 0.0
})
model({
  ka <- exp(tka + eta.ka)
  cl <- exp(tcl + eta.cl)
  v <- exp(tv + eta.v + wt.est * WT)
  d/dt(depot) = -ka * depot
  d/dt(center) = ka * depot - cl / v * center
  cp = center / v
  cp ~ add(add.sd)
})
}
fit  <- nlmixr(one.compartment, theo_sd, "saem")

#> ℹ parameter labels from comments will be replaced by 'label()'

#>  

#> → generate SAEM model

#> ✔ done

#> 1:    0.1675   0.9655   3.3527   0.0007   0.5700   0.2850   0.0950   5.3940
#> 2:    0.2385   0.9543   3.0659   0.0057   0.5415   0.2707   0.0902   1.6774
#> 3:    0.4555   1.0337   3.3566   0.0015   0.5144   0.2572   0.0857   1.0735
#> 4:    0.4497   1.0299   2.9211   0.0078   0.4887   0.2444   0.0815   0.7586
#> 5:    0.4722   1.0300   2.7501   0.0101   0.5009   0.2321   0.0774   0.5925
#> 6:    0.4552   1.0097   2.7337   0.0105   0.4892   0.2205   0.0735   0.5602
#> 7:    0.4386   1.0180   2.6822   0.0109   0.5013   0.2095   0.0698   0.5043
#> 8:    0.4285   1.0141   2.7920   0.0096   0.4962   0.1990   0.0663   0.4986
#> 9:    0.4389   0.9899   2.8373   0.0090   0.4714   0.1891   0.0630   0.4902
#> 10:    0.5005   1.0170   2.8014   0.0094   0.4919   0.1796   0.0599   0.5068
#> 11:    0.4993   1.0143   2.9150   0.0079   0.4972   0.1706   0.0569   0.5130
#> 12:    0.4375   1.0217   2.8136   0.0093   0.4724   0.1621   0.0540   0.5077
#> 13:    0.4381   1.0241   2.8946   0.0078   0.4487   0.1540   0.0513   0.4883
#> 14:    0.4607   1.0313   3.0309   0.0060   0.4431   0.1463   0.0488   0.5074
#> 15:    0.4786   0.9986   2.8903   0.0081   0.4680   0.1390   0.0463   0.4911
#> 16:    0.4578   0.9954   3.0002   0.0064   0.4446   0.1320   0.0440   0.4850
#> 17:    0.4413   0.9990   3.0805   0.0054   0.4224   0.1254   0.0418   0.4953
#> 18:    0.4593   1.0004   3.0068   0.0064   0.4013   0.1192   0.0397   0.4944
#> 19:    0.4386   0.9842   3.0746   0.0056   0.3812   0.1132   0.0377   0.5035
#> 20:    0.4352   0.9981   3.0180   0.0064   0.3622   0.1075   0.0358   0.5040
#> 21:    0.4629   1.0153   3.0023   0.0066   0.4005   0.1022   0.0341   0.4712
#> 22:    0.5040   1.0142   2.9701   0.0071   0.3972   0.0971   0.0324   0.4763
#> 23:    0.4786   1.0109   2.8724   0.0086   0.3773   0.0922   0.0307   0.4679
#> 24:    0.4702   1.0112   2.8830   0.0084   0.3584   0.0876   0.0292   0.4628
#> 25:    0.4944   0.9770   2.9138   0.0081   0.3683   0.0832   0.0277   0.4788
#> 26:    0.5040   1.0200   3.0862   0.0054   0.3499   0.0791   0.0264   0.4839
#> 27:    0.4800   1.0161   3.0246   0.0064   0.4188   0.0751   0.0250   0.4738
#> 28:    0.5324   1.0240   3.0417   0.0062   0.5126   0.0821   0.0238   0.4977
#> 29:    0.5377   1.0113   3.0438   0.0063   0.4869   0.0780   0.0226   0.4862
#> 30:    0.4920   1.0141   3.1429   0.0046   0.4804   0.0780   0.0215   0.5032
#> 31:    0.4702   1.0090   3.0221   0.0062   0.4564   0.0807   0.0204   0.4771
#> 32:    0.4436   1.0212   2.9917   0.0066   0.4336   0.0766   0.0194   0.4994
#> 33:    0.5128   1.0119   2.8679   0.0086   0.5005   0.0728   0.0186   0.4666
#> 34:    0.4768   1.0005   2.8176   0.0093   0.4755   0.0691   0.0177   0.4729
#> 35:    0.4486   0.9953   2.7705   0.0099   0.4601   0.0657   0.0168   0.4890
#> 36:    0.4628   1.0188   2.7788   0.0098   0.4510   0.0624   0.0159   0.4708
#> 37:    0.5096   0.9996   2.7410   0.0102   0.4963   0.0710   0.0152   0.4787
#> 38:    0.5140   1.0130   2.9304   0.0076   0.4818   0.0774   0.0144   0.4731
#> 39:    0.4979   1.0345   2.8641   0.0085   0.4820   0.0735   0.0137   0.4606
#> 40:    0.4833   1.0351   2.9858   0.0065   0.4579   0.0698   0.0130   0.4629
#> 41:    0.4485   1.0239   3.0009   0.0064   0.4350   0.0771   0.0123   0.4566
#> 42:    0.4368   1.0262   3.0519   0.0058   0.4133   0.0793   0.0150   0.4631
#> 43:    0.4732   1.0138   2.9883   0.0066   0.4296   0.0754   0.0143   0.4678
#> 44:    0.4735   0.9765   3.0598   0.0058   0.4275   0.0716   0.0136   0.4686
#> 45:    0.4828   0.9900   3.1378   0.0048   0.4263   0.0777   0.0130   0.4823
#> 46:    0.4930   1.0005   3.0003   0.0067   0.4704   0.0991   0.0123   0.4763
#> 47:    0.4934   0.9900   2.9956   0.0067   0.5616   0.0942   0.0125   0.4780
#> 48:    0.4186   1.0111   2.9555   0.0071   0.5335   0.0895   0.0119   0.4808
#> 49:    0.4345   1.0012   2.9864   0.0067   0.5068   0.0850   0.0136   0.4714
#> 50:    0.4286   0.9770   2.9315   0.0075   0.4815   0.0807   0.0130   0.4990
#> 51:    0.4531   1.0069   2.8697   0.0084   0.4621   0.0767   0.0131   0.4864
#> 52:    0.4869   1.0143   2.9186   0.0077   0.4390   0.0729   0.0124   0.4981
#> 53:    0.4762   0.9856   3.0190   0.0063   0.4170   0.0692   0.0142   0.4804
#> 54:    0.4505   0.9862   2.9063   0.0079   0.3962   0.0699   0.0149   0.4565
#> 55:    0.4384   1.0032   2.8821   0.0083   0.3764   0.0664   0.0177   0.4746
#> 56:    0.4549   1.0269   2.9269   0.0077   0.3688   0.0701   0.0179   0.4705
#> 57:    0.4522   1.0311   2.8636   0.0085   0.3626   0.0666   0.0209   0.4660
#> 58:    0.4687   1.0052   2.9142   0.0079   0.3522   0.0632   0.0198   0.4629
#> 59:    0.4635   0.9945   2.8357   0.0090   0.3865   0.0625   0.0188   0.4979
#> 60:    0.5350   0.9684   2.9140   0.0082   0.4388   0.0643   0.0179   0.4713
#> 61:    0.4968   0.9876   2.8835   0.0086   0.4169   0.0640   0.0170   0.4854
#> 62:    0.4879   0.9789   2.9447   0.0077   0.3960   0.0608   0.0185   0.4917
#> 63:    0.5009   0.9952   2.7875   0.0098   0.4236   0.0627   0.0176   0.4890
#> 64:    0.4942   0.9886   2.7543   0.0102   0.4215   0.0625   0.0167   0.4660
#> 65:    0.4449   1.0332   2.8188   0.0091   0.4185   0.0699   0.0159   0.4506
#> 66:    0.4677   1.0136   2.8221   0.0090   0.5087   0.0676   0.0180   0.4408
#> 67:    0.4603   1.0272   2.8231   0.0090   0.5156   0.0642   0.0182   0.4514
#> 68:    0.4526   1.0080   2.8408   0.0088   0.4898   0.0610   0.0238   0.4707
#> 69:    0.5105   0.9688   2.8175   0.0094   0.4653   0.0586   0.0300   0.5309
#> 70:    0.4905   0.9644   2.8786   0.0087   0.4500   0.0557   0.0285   0.4985
#> 71:    0.4813   0.9739   2.7580   0.0101   0.4562   0.0529   0.0271   0.4925
#> 72:    0.4731   0.9905   2.8347   0.0090   0.4333   0.0502   0.0257   0.4878
#> 73:    0.4949   0.9959   2.8761   0.0083   0.5842   0.0517   0.0244   0.5196
#> 74:    0.4500   0.9937   2.9217   0.0079   0.5550   0.0570   0.0232   0.4856
#> 75:    0.5261   1.0021   2.9559   0.0075   0.5272   0.0542   0.0220   0.4584
#> 76:    0.5005   1.0005   2.9779   0.0071   0.5009   0.0644   0.0209   0.4600
#> 77:    0.4858   1.0153   2.8886   0.0083   0.4758   0.0775   0.0199   0.4733
#> 78:    0.4506   1.0059   2.8803   0.0082   0.4520   0.0736   0.0189   0.4931
#> 79:    0.4517   1.0006   2.8766   0.0083   0.4294   0.0735   0.0179   0.4859
#> 80:    0.4403   1.0179   2.8346   0.0089   0.4080   0.0724   0.0170   0.4407
#> 81:    0.4844   1.0004   2.8846   0.0082   0.3904   0.0687   0.0162   0.4352
#> 82:    0.4681   1.0229   2.7971   0.0096   0.4155   0.0703   0.0154   0.4817
#> 83:    0.4473   1.0034   2.7599   0.0101   0.3947   0.0770   0.0146   0.4872
#> 84:    0.4306   1.0153   2.9246   0.0075   0.3750   0.0731   0.0159   0.4853
#> 85:    0.4765   1.0350   2.9694   0.0070   0.4082   0.0765   0.0151   0.4720
#> 86:    0.4522   1.0363   2.8433   0.0087   0.4028   0.0727   0.0147   0.4560
#> 87:    0.4525   1.0183   2.8646   0.0085   0.4078   0.0691   0.0140   0.4617
#> 88:    0.4935   1.0254   2.8462   0.0085   0.4737   0.0773   0.0134   0.4618
#> 89:    0.4448   1.0218   2.9197   0.0074   0.4500   0.0735   0.0145   0.4667
#> 90:    0.4306   1.0382   2.8909   0.0077   0.4275   0.0782   0.0138   0.4564
#> 91:    0.3824   1.0360   2.9506   0.0069   0.4113   0.0763   0.0131   0.4655
#> 92:    0.4295   1.0436   2.8976   0.0078   0.3907   0.0749   0.0125   0.5029
#> 93:    0.4248   1.0449   2.8759   0.0081   0.3799   0.0712   0.0139   0.4788
#> 94:    0.3856   1.0315   2.8921   0.0077   0.3609   0.0676   0.0170   0.4720
#> 95:    0.4109   1.0319   2.8823   0.0081   0.3656   0.0642   0.0186   0.4709
#> 96:    0.4352   1.0408   2.8656   0.0082   0.4480   0.0654   0.0182   0.4404
#> 97:    0.4403   1.0393   2.8219   0.0087   0.4256   0.0776   0.0173   0.4572
#> 98:    0.4414   1.0432   2.9066   0.0076   0.4043   0.0737   0.0191   0.4720
#> 99:    0.4099   1.0552   2.8546   0.0081   0.4007   0.0700   0.0181   0.4749
#> 100:    0.4039   1.0538   2.8088   0.0089   0.3806   0.0716   0.0180   0.4746
#> 101:    0.4090   1.0175   2.7833   0.0092   0.3936   0.0680   0.0171   0.5052
#> 102:    0.3923   1.0163   2.8063   0.0089   0.4340   0.0651   0.0181   0.4897
#> 103:    0.4447   1.0033   2.9571   0.0069   0.4545   0.0669   0.0172   0.4691
#> 104:    0.3928   1.0358   3.0184   0.0059   0.4318   0.0778   0.0164   0.4805
#> 105:    0.4233   1.0342   3.0717   0.0052   0.4102   0.0834   0.0155   0.4470
#> 106:    0.4453   1.0257   2.9591   0.0071   0.3897   0.0792   0.0148   0.4469
#> 107:    0.4392   1.0173   3.0102   0.0061   0.3702   0.0752   0.0140   0.4639
#> 108:    0.4604   1.0289   2.9906   0.0065   0.4469   0.0715   0.0133   0.4716
#> 109:    0.4346   1.0171   3.0411   0.0059   0.4245   0.0679   0.0127   0.4750
#> 110:    0.4413   1.0162   2.9595   0.0071   0.4181   0.0712   0.0120   0.4712
#> 111:    0.4366   1.0199   3.0294   0.0060   0.3972   0.0745   0.0114   0.4805
#> 112:    0.4557   1.0182   2.9170   0.0076   0.4283   0.0771   0.0116   0.4722
#> 113:    0.4326   1.0227   2.9328   0.0075   0.4069   0.0732   0.0140   0.4666
#> 114:    0.4991   1.0041   2.9245   0.0078   0.3872   0.0695   0.0187   0.4864
#> 115:    0.4888   1.0245   2.9274   0.0077   0.4209   0.0661   0.0205   0.5085
#> 116:    0.4434   1.0136   2.8371   0.0090   0.3999   0.0628   0.0195   0.4900
#> 117:    0.4500   1.0108   2.8218   0.0090   0.4202   0.0596   0.0185   0.5022
#> 118:    0.4910   1.0276   2.8377   0.0086   0.5280   0.0582   0.0176   0.5002
#> 119:    0.4615   1.0209   2.9607   0.0068   0.5016   0.0592   0.0167   0.5096
#> 120:    0.4542   1.0254   2.9359   0.0074   0.4765   0.0644   0.0159   0.4599
#> 121:    0.4288   1.0332   2.9251   0.0075   0.4833   0.0640   0.0171   0.4804
#> 122:    0.4184   1.0425   2.9381   0.0071   0.4592   0.0786   0.0162   0.4929
#> 123:    0.4269   1.0388   2.9504   0.0069   0.4414   0.0747   0.0169   0.4822
#> 124:    0.4125   1.0339   2.9409   0.0071   0.4235   0.0709   0.0161   0.4471
#> 125:    0.4383   1.0078   2.9099   0.0075   0.4056   0.0674   0.0153   0.4615
#> 126:    0.4163   1.0399   2.8940   0.0077   0.3907   0.0663   0.0145   0.4991
#> 127:    0.4151   1.0522   2.8321   0.0084   0.4446   0.0629   0.0138   0.4994
#> 128:    0.4084   1.0616   2.8203   0.0086   0.4224   0.0786   0.0131   0.5093
#> 129:    0.4172   1.0281   2.8617   0.0080   0.5528   0.0903   0.0126   0.4498
#> 130:    0.4364   1.0063   2.8550   0.0083   0.5252   0.0858   0.0137   0.4432
#> 131:    0.4397   1.0049   2.8769   0.0081   0.4989   0.0815   0.0167   0.4497
#> 132:    0.4595   0.9800   2.9485   0.0073   0.4827   0.0774   0.0185   0.4622
#> 133:    0.4812   1.0040   2.9354   0.0076   0.4586   0.0736   0.0180   0.4621
#> 134:    0.4943   0.9926   2.8385   0.0090   0.4676   0.0699   0.0171   0.4597
#> 135:    0.4564   1.0076   2.8649   0.0084   0.4834   0.0851   0.0162   0.4832
#> 136:    0.4185   1.0424   2.7174   0.0103   0.4887   0.0852   0.0154   0.4843
#> 137:    0.4558   1.0508   2.7010   0.0106   0.4642   0.0809   0.0146   0.4810
#> 138:    0.4719   1.0265   2.7275   0.0104   0.4410   0.0769   0.0176   0.4911
#> 139:    0.4422   1.0297   2.7573   0.0099   0.4361   0.0731   0.0233   0.4665
#> 140:    0.4767   1.0384   2.8248   0.0089   0.4923   0.0752   0.0221   0.4841
#> 141:    0.4484   1.0288   2.8346   0.0088   0.4814   0.0715   0.0210   0.4658
#> 142:    0.4909   1.0264   2.8358   0.0089   0.4699   0.0679   0.0200   0.4391
#> 143:    0.4697   1.0130   2.8215   0.0091   0.4464   0.0645   0.0190   0.4624
#> 144:    0.4821   1.0159   2.9147   0.0077   0.4928   0.0613   0.0180   0.4460
#> 145:    0.4752   1.0104   2.9716   0.0071   0.4682   0.0610   0.0180   0.4949
#> 146:    0.4931   1.0136   2.9738   0.0069   0.5069   0.0590   0.0171   0.4858
#> 147:    0.4499   1.0325   2.9245   0.0075   0.5049   0.0744   0.0162   0.4489
#> 148:    0.4739   1.0122   2.8112   0.0092   0.5897   0.0794   0.0154   0.4381
#> 149:    0.5099   1.0147   2.8612   0.0084   0.5602   0.0769   0.0146   0.4733
#> 150:    0.4431   1.0082   2.8802   0.0083   0.5322   0.0815   0.0139   0.4811
#> 151:    0.4443   1.0036   2.8651   0.0085   0.5056   0.0867   0.0132   0.4619
#> 152:    0.4652   1.0110   2.8180   0.0091   0.3853   0.0669   0.0152   0.4673
#> 153:    0.4664   1.0209   2.8186   0.0090   0.4875   0.0583   0.0151   0.4442
#> 154:    0.4445   0.9914   2.8976   0.0079   0.4233   0.0680   0.0115   0.4585
#> 155:    0.4804   1.0062   2.9316   0.0076   0.5024   0.0743   0.0089   0.4717
#> 156:    0.4847   1.0135   2.8440   0.0087   0.4894   0.0695   0.0104   0.4802
#> 157:    0.4446   1.0078   2.8650   0.0084   0.4995   0.0643   0.0111   0.4850
#> 158:    0.4526   1.0346   2.8954   0.0079   0.4670   0.0785   0.0155   0.4688
#> 159:    0.4533   1.0227   2.8645   0.0084   0.4033   0.0689   0.0112   0.4405
#> 160:    0.4467   1.0079   2.9767   0.0068   0.3896   0.0565   0.0110   0.4695
#> 161:    0.4686   0.9711   2.9462   0.0074   0.4436   0.0613   0.0143   0.4838
#> 162:    0.4753   1.0110   2.7951   0.0096   0.3268   0.0638   0.0148   0.4900
#> 163:    0.4920   0.9964   2.7251   0.0105   0.4837   0.0506   0.0166   0.4744
#> 164:    0.4591   1.0113   2.7562   0.0101   0.4917   0.0525   0.0154   0.4841
#> 165:    0.4510   1.0394   2.9627   0.0071   0.3929   0.0638   0.0141   0.4628
#> 166:    0.4700   0.9968   2.9662   0.0071   0.3709   0.0626   0.0159   0.4985
#> 167:    0.4788   0.9984   3.0201   0.0064   0.3974   0.0676   0.0144   0.4977
#> 168:    0.4402   0.9991   2.9667   0.0071   0.3728   0.0615   0.0139   0.4820
#> 169:    0.4389   1.0392   3.0152   0.0062   0.3569   0.0590   0.0142   0.4899
#> 170:    0.4354   1.0284   3.0124   0.0063   0.3992   0.0566   0.0156   0.4995
#> 171:    0.4060   1.0191   2.8784   0.0079   0.4730   0.0700   0.0089   0.4985
#> 172:    0.4359   1.0154   2.9410   0.0072   0.4118   0.0736   0.0083   0.4589
#> 173:    0.4268   1.0468   2.9924   0.0064   0.4022   0.0758   0.0071   0.4925
#> 174:    0.4339   1.0544   2.9408   0.0072   0.3334   0.0829   0.0071   0.4807
#> 175:    0.3981   1.0083   2.9416   0.0073   0.3090   0.0789   0.0085   0.5098
#> 176:    0.4107   0.9918   2.9511   0.0073   0.2975   0.0948   0.0090   0.5077
#> 177:    0.4371   0.9972   2.8745   0.0083   0.4524   0.0880   0.0109   0.4745
#> 178:    0.4398   1.0140   2.8355   0.0088   0.3630   0.0844   0.0129   0.5193
#> 179:    0.4607   1.0108   2.9135   0.0077   0.3179   0.0724   0.0149   0.5098
#> 180:    0.4440   0.9950   2.9009   0.0079   0.3959   0.0688   0.0125   0.4775
#> 181:    0.4361   1.0181   2.9379   0.0072   0.3526   0.0665   0.0134   0.4783
#> 182:    0.4043   1.0598   2.9452   0.0069   0.3642   0.0662   0.0147   0.4630
#> 183:    0.4100   1.0157   2.9058   0.0076   0.3394   0.0789   0.0110   0.4525
#> 184:    0.4251   1.0343   2.8954   0.0077   0.4204   0.0734   0.0101   0.4707
#> 185:    0.4017   1.0417   2.8241   0.0085   0.3666   0.0807   0.0124   0.4924
#> 186:    0.3746   1.0276   2.8605   0.0080   0.3228   0.0694   0.0111   0.4671
#> 187:    0.3982   1.0371   2.8695   0.0079   0.3986   0.0825   0.0109   0.4755
#> 188:    0.3822   1.0484   2.8290   0.0086   0.3320   0.0813   0.0094   0.4728
#> 189:    0.3802   1.0464   2.8037   0.0090   0.3347   0.0793   0.0094   0.4864
#> 190:    0.4066   1.0222   2.7947   0.0091   0.3717   0.0649   0.0099   0.4949
#> 191:    0.4262   1.0317   2.9034   0.0076   0.4218   0.0619   0.0116   0.4873
#> 192:    0.4436   1.0299   2.8064   0.0089   0.4764   0.0640   0.0107   0.4778
#> 193:    0.3951   1.0492   2.8222   0.0089   0.4169   0.0818   0.0095   0.4783
#> 194:    0.4092   1.0219   2.8236   0.0087   0.4336   0.0920   0.0088   0.4760
#> 195:    0.3885   1.0157   2.8651   0.0080   0.4493   0.0974   0.0084   0.4604
#> 196:    0.4308   1.0340   2.8197   0.0086   0.5379   0.0844   0.0072   0.4500
#> 197:    0.4522   1.0415   2.8880   0.0077   0.5985   0.0821   0.0082   0.5016
#> 198:    0.3736   1.0387   2.9671   0.0064   0.4252   0.0863   0.0080   0.5198
#> 199:    0.3988   1.0281   2.9994   0.0063   0.4389   0.0937   0.0069   0.5039
#> 200:    0.4004   1.0156   2.9917   0.0062   0.3984   0.0973   0.0058   0.5378
#> 201:    0.4089   1.0176   2.9983   0.0061   0.4173   0.0958   0.0063   0.5191
#> 202:    0.4087   1.0234   2.9915   0.0062   0.4194   0.0953   0.0062   0.5121
#> 203:    0.4021   1.0268   2.9788   0.0065   0.4248   0.0945   0.0060   0.5114
#> 204:    0.4123   1.0236   2.9827   0.0065   0.4161   0.0912   0.0062   0.5082
#> 205:    0.4128   1.0254   2.9750   0.0066   0.4143   0.0905   0.0061   0.5074
#> 206:    0.4187   1.0250   2.9756   0.0066   0.4167   0.0885   0.0062   0.5047
#> 207:    0.4164   1.0241   2.9678   0.0067   0.4202   0.0874   0.0064   0.5043
#> 208:    0.4189   1.0239   2.9637   0.0068   0.4205   0.0849   0.0066   0.5008
#> 209:    0.4185   1.0257   2.9597   0.0069   0.4157   0.0835   0.0070   0.5009
#> 210:    0.4170   1.0267   2.9572   0.0069   0.4144   0.0823   0.0074   0.4972
#> 211:    0.4173   1.0252   2.9544   0.0069   0.4119   0.0819   0.0076   0.4948
#> 212:    0.4189   1.0230   2.9511   0.0070   0.4082   0.0821   0.0077   0.4958
#> 213:    0.4184   1.0211   2.9466   0.0071   0.4040   0.0822   0.0078   0.4962
#> 214:    0.4158   1.0209   2.9433   0.0071   0.3992   0.0813   0.0080   0.4967
#> 215:    0.4143   1.0209   2.9385   0.0072   0.4007   0.0807   0.0082   0.4968
#> 216:    0.4156   1.0209   2.9366   0.0072   0.4030   0.0793   0.0084   0.4960
#> 217:    0.4158   1.0215   2.9354   0.0072   0.4018   0.0790   0.0085   0.4960
#> 218:    0.4143   1.0228   2.9364   0.0072   0.4088   0.0793   0.0085   0.4958
#> 219:    0.4132   1.0230   2.9405   0.0071   0.4096   0.0790   0.0086   0.4955
#> 220:    0.4117   1.0242   2.9444   0.0070   0.4114   0.0793   0.0087   0.4957
#> 221:    0.4116   1.0260   2.9470   0.0070   0.4120   0.0790   0.0087   0.4946
#> 222:    0.4117   1.0280   2.9507   0.0069   0.4103   0.0788   0.0088   0.4942
#> 223:    0.4110   1.0286   2.9522   0.0069   0.4099   0.0787   0.0089   0.4935
#> 224:    0.4123   1.0287   2.9489   0.0070   0.4122   0.0781   0.0090   0.4927
#> 225:    0.4141   1.0280   2.9510   0.0069   0.4128   0.0779   0.0091   0.4920
#> 226:    0.4168   1.0276   2.9532   0.0069   0.4125   0.0781   0.0091   0.4919
#> 227:    0.4163   1.0279   2.9545   0.0069   0.4102   0.0785   0.0092   0.4909
#> 228:    0.4153   1.0276   2.9592   0.0068   0.4083   0.0787   0.0092   0.4913
#> 229:    0.4155   1.0273   2.9625   0.0068   0.4062   0.0783   0.0092   0.4918
#> 230:    0.4161   1.0269   2.9663   0.0067   0.4045   0.0781   0.0093   0.4918
#> 231:    0.4175   1.0269   2.9679   0.0067   0.4060   0.0780   0.0094   0.4912
#> 232:    0.4201   1.0272   2.9686   0.0067   0.4084   0.0775   0.0095   0.4911
#> 233:    0.4218   1.0266   2.9702   0.0067   0.4106   0.0770   0.0095   0.4915
#> 234:    0.4219   1.0256   2.9694   0.0067   0.4124   0.0768   0.0096   0.4915
#> 235:    0.4225   1.0256   2.9676   0.0067   0.4123   0.0767   0.0097   0.4906
#> 236:    0.4243   1.0257   2.9673   0.0067   0.4133   0.0769   0.0098   0.4895
#> 237:    0.4242   1.0259   2.9671   0.0067   0.4135   0.0770   0.0098   0.4897
#> 238:    0.4243   1.0264   2.9647   0.0068   0.4140   0.0772   0.0098   0.4899
#> 239:    0.4251   1.0262   2.9615   0.0068   0.4153   0.0769   0.0098   0.4898
#> 240:    0.4252   1.0264   2.9611   0.0068   0.4184   0.0765   0.0099   0.4890
#> 241:    0.4255   1.0266   2.9607   0.0068   0.4184   0.0762   0.0099   0.4891
#> 242:    0.4259   1.0263   2.9598   0.0068   0.4202   0.0760   0.0099   0.4884
#> 243:    0.4260   1.0257   2.9582   0.0069   0.4218   0.0756   0.0100   0.4886
#> 244:    0.4266   1.0257   2.9575   0.0069   0.4231   0.0755   0.0101   0.4885
#> 245:    0.4264   1.0260   2.9553   0.0069   0.4225   0.0755   0.0101   0.4882
#> 246:    0.4273   1.0259   2.9527   0.0070   0.4232   0.0756   0.0102   0.4882
#> 247:    0.4281   1.0258   2.9526   0.0070   0.4252   0.0755   0.0102   0.4881
#> 248:    0.4285   1.0258   2.9540   0.0069   0.4256   0.0756   0.0102   0.4883
#> 249:    0.4288   1.0260   2.9546   0.0069   0.4245   0.0754   0.0102   0.4885
#> 250:    0.4293   1.0258   2.9558   0.0069   0.4244   0.0752   0.0103   0.4887
#> 251:    0.4305   1.0255   2.9569   0.0069   0.4242   0.0751   0.0103   0.4884
#> 252:    0.4313   1.0253   2.9594   0.0069   0.4234   0.0751   0.0103   0.4880
#> 253:    0.4311   1.0255   2.9618   0.0068   0.4230   0.0751   0.0103   0.4877
#> 254:    0.4317   1.0254   2.9623   0.0068   0.4231   0.0753   0.0103   0.4877
#> 255:    0.4322   1.0251   2.9633   0.0068   0.4219   0.0754   0.0103   0.4875
#> 256:    0.4321   1.0246   2.9624   0.0068   0.4218   0.0757   0.0103   0.4870
#> 257:    0.4325   1.0239   2.9619   0.0069   0.4227   0.0758   0.0103   0.4868
#> 258:    0.4324   1.0237   2.9608   0.0069   0.4219   0.0762   0.0103   0.4869
#> 259:    0.4322   1.0230   2.9601   0.0069   0.4222   0.0767   0.0102   0.4871
#> 260:    0.4322   1.0225   2.9597   0.0069   0.4217   0.0766   0.0102   0.4871
#> 261:    0.4335   1.0220   2.9581   0.0069   0.4215   0.0766   0.0102   0.4869
#> 262:    0.4340   1.0218   2.9568   0.0070   0.4204   0.0767   0.0103   0.4868
#> 263:    0.4340   1.0213   2.9555   0.0070   0.4217   0.0771   0.0103   0.4866
#> 264:    0.4343   1.0213   2.9534   0.0070   0.4229   0.0771   0.0103   0.4864
#> 265:    0.4345   1.0210   2.9521   0.0070   0.4233   0.0774   0.0103   0.4865
#> 266:    0.4343   1.0209   2.9503   0.0071   0.4231   0.0773   0.0103   0.4862
#> 267:    0.4341   1.0205   2.9484   0.0071   0.4225   0.0775   0.0103   0.4864
#> 268:    0.4339   1.0206   2.9465   0.0071   0.4223   0.0777   0.0103   0.4863
#> 269:    0.4343   1.0203   2.9451   0.0071   0.4219   0.0778   0.0103   0.4863
#> 270:    0.4335   1.0204   2.9442   0.0071   0.4223   0.0779   0.0103   0.4866
#> 271:    0.4336   1.0202   2.9421   0.0072   0.4237   0.0778   0.0103   0.4866
#> 272:    0.4340   1.0199   2.9411   0.0072   0.4242   0.0781   0.0104   0.4866
#> 273:    0.4349   1.0198   2.9409   0.0072   0.4247   0.0782   0.0104   0.4865
#> 274:    0.4352   1.0198   2.9387   0.0072   0.4247   0.0782   0.0104   0.4865
#> 275:    0.4351   1.0197   2.9375   0.0073   0.4243   0.0782   0.0105   0.4862
#> 276:    0.4349   1.0196   2.9355   0.0073   0.4242   0.0783   0.0105   0.4861
#> 277:    0.4353   1.0199   2.9322   0.0073   0.4251   0.0780   0.0105   0.4860
#> 278:    0.4356   1.0201   2.9301   0.0074   0.4253   0.0781   0.0105   0.4858
#> 279:    0.4356   1.0205   2.9284   0.0074   0.4248   0.0782   0.0105   0.4858
#> 280:    0.4360   1.0202   2.9263   0.0074   0.4251   0.0783   0.0105   0.4854
#> 281:    0.4362   1.0204   2.9259   0.0074   0.4251   0.0782   0.0105   0.4856
#> 282:    0.4358   1.0206   2.9264   0.0074   0.4248   0.0781   0.0105   0.4857
#> 283:    0.4354   1.0207   2.9269   0.0074   0.4248   0.0780   0.0105   0.4858
#> 284:    0.4354   1.0206   2.9268   0.0074   0.4253   0.0781   0.0105   0.4856
#> 285:    0.4356   1.0204   2.9274   0.0074   0.4258   0.0782   0.0105   0.4851
#> 286:    0.4361   1.0202   2.9281   0.0074   0.4261   0.0781   0.0105   0.4851
#> 287:    0.4367   1.0199   2.9292   0.0074   0.4265   0.0782   0.0105   0.4851
#> 288:    0.4373   1.0199   2.9306   0.0074   0.4273   0.0783   0.0105   0.4850
#> 289:    0.4374   1.0199   2.9323   0.0073   0.4270   0.0782   0.0105   0.4849
#> 290:    0.4377   1.0200   2.9335   0.0073   0.4265   0.0782   0.0105   0.4850
#> 291:    0.4380   1.0204   2.9356   0.0073   0.4264   0.0783   0.0106   0.4850
#> 292:    0.4381   1.0207   2.9382   0.0072   0.4263   0.0782   0.0106   0.4850
#> 293:    0.4382   1.0207   2.9401   0.0072   0.4263   0.0781   0.0106   0.4847
#> 294:    0.4384   1.0205   2.9406   0.0072   0.4257   0.0778   0.0106   0.4848
#> 295:    0.4382   1.0205   2.9397   0.0072   0.4254   0.0777   0.0106   0.4846
#> 296:    0.4382   1.0201   2.9389   0.0072   0.4258   0.0776   0.0106   0.4845
#> 297:    0.4388   1.0202   2.9381   0.0072   0.4261   0.0775   0.0107   0.4844
#> 298:    0.4389   1.0203   2.9361   0.0073   0.4258   0.0774   0.0107   0.4841
#> 299:    0.4391   1.0202   2.9343   0.0073   0.4254   0.0774   0.0107   0.4839
#> 300:    0.4394   1.0201   2.9324   0.0073   0.4255   0.0774   0.0107   0.4837
#> 301:    0.4392   1.0201   2.9307   0.0074   0.4255   0.0773   0.0107   0.4838
#> 302:    0.4392   1.0201   2.9293   0.0074   0.4255   0.0771   0.0107   0.4839
#> 303:    0.4396   1.0199   2.9284   0.0074   0.4258   0.0769   0.0108   0.4837
#> 304:    0.4398   1.0195   2.9282   0.0074   0.4261   0.0768   0.0108   0.4836
#> 305:    0.4395   1.0196   2.9272   0.0074   0.4260   0.0768   0.0108   0.4833
#> 306:    0.4389   1.0195   2.9272   0.0074   0.4253   0.0766   0.0108   0.4833
#> 307:    0.4390   1.0194   2.9269   0.0074   0.4250   0.0765   0.0108   0.4835
#> 308:    0.4393   1.0192   2.9268   0.0074   0.4248   0.0764   0.0108   0.4834
#> 309:    0.4399   1.0194   2.9260   0.0074   0.4248   0.0764   0.0108   0.4832
#> 310:    0.4401   1.0194   2.9252   0.0075   0.4250   0.0764   0.0108   0.4833
#> 311:    0.4400   1.0194   2.9248   0.0075   0.4248   0.0765   0.0108   0.4834
#> 312:    0.4400   1.0193   2.9244   0.0075   0.4242   0.0766   0.0108   0.4834
#> 313:    0.4401   1.0192   2.9242   0.0075   0.4236   0.0766   0.0108   0.4833
#> 314:    0.4404   1.0191   2.9241   0.0075   0.4241   0.0767   0.0108   0.4831
#> 315:    0.4405   1.0190   2.9243   0.0075   0.4240   0.0766   0.0108   0.4831
#> 316:    0.4406   1.0190   2.9245   0.0075   0.4237   0.0766   0.0109   0.4831
#> 317:    0.4403   1.0192   2.9242   0.0075   0.4236   0.0766   0.0109   0.4830
#> 318:    0.4398   1.0193   2.9232   0.0075   0.4232   0.0766   0.0108   0.4830
#> 319:    0.4397   1.0194   2.9222   0.0075   0.4234   0.0766   0.0108   0.4830
#> 320:    0.4393   1.0198   2.9217   0.0075   0.4234   0.0766   0.0108   0.4833
#> 321:    0.4391   1.0200   2.9221   0.0075   0.4232   0.0767   0.0108   0.4836
#> 322:    0.4386   1.0202   2.9225   0.0075   0.4232   0.0766   0.0108   0.4836
#> 323:    0.4387   1.0203   2.9226   0.0075   0.4235   0.0767   0.0108   0.4836
#> 324:    0.4387   1.0204   2.9217   0.0075   0.4234   0.0766   0.0108   0.4835
#> 325:    0.4387   1.0205   2.9211   0.0075   0.4233   0.0765   0.0108   0.4835
#> 326:    0.4385   1.0205   2.9200   0.0075   0.4228   0.0763   0.0109   0.4833
#> 327:    0.4389   1.0207   2.9188   0.0075   0.4235   0.0762   0.0109   0.4833
#> 328:    0.4389   1.0208   2.9183   0.0075   0.4236   0.0762   0.0109   0.4830
#> 329:    0.4391   1.0208   2.9185   0.0075   0.4237   0.0763   0.0109   0.4828
#> 330:    0.4391   1.0211   2.9183   0.0075   0.4242   0.0763   0.0110   0.4828
#> 331:    0.4392   1.0208   2.9178   0.0075   0.4247   0.0762   0.0110   0.4827
#> 332:    0.4393   1.0207   2.9179   0.0075   0.4248   0.0762   0.0110   0.4828
#> 333:    0.4394   1.0206   2.9184   0.0075   0.4249   0.0761   0.0110   0.4829
#> 334:    0.4394   1.0205   2.9187   0.0075   0.4244   0.0761   0.0110   0.4831
#> 335:    0.4393   1.0205   2.9191   0.0075   0.4245   0.0761   0.0110   0.4830
#> 336:    0.4393   1.0205   2.9194   0.0075   0.4246   0.0760   0.0110   0.4830
#> 337:    0.4395   1.0204   2.9198   0.0075   0.4248   0.0759   0.0110   0.4831
#> 338:    0.4394   1.0202   2.9205   0.0075   0.4248   0.0760   0.0110   0.4832
#> 339:    0.4394   1.0201   2.9213   0.0075   0.4249   0.0760   0.0110   0.4831
#> 340:    0.4396   1.0201   2.9219   0.0075   0.4253   0.0760   0.0110   0.4829
#> 341:    0.4398   1.0201   2.9227   0.0075   0.4254   0.0759   0.0110   0.4828
#> 342:    0.4401   1.0201   2.9233   0.0075   0.4256   0.0758   0.0111   0.4827
#> 343:    0.4402   1.0202   2.9235   0.0075   0.4256   0.0758   0.0111   0.4825
#> 344:    0.4402   1.0202   2.9237   0.0075   0.4261   0.0758   0.0111   0.4826
#> 345:    0.4401   1.0202   2.9234   0.0075   0.4265   0.0758   0.0111   0.4827
#> 346:    0.4401   1.0202   2.9234   0.0075   0.4269   0.0758   0.0111   0.4826
#> 347:    0.4403   1.0202   2.9233   0.0075   0.4275   0.0757   0.0111   0.4826
#> 348:    0.4404   1.0204   2.9234   0.0075   0.4278   0.0758   0.0111   0.4828
#> 349:    0.4406   1.0205   2.9242   0.0075   0.4280   0.0758   0.0111   0.4830
#> 350:    0.4406   1.0205   2.9251   0.0074   0.4280   0.0758   0.0111   0.4831
#> 351:    0.4407   1.0205   2.9260   0.0074   0.4282   0.0757   0.0111   0.4829
#> 352:    0.4407   1.0206   2.9262   0.0074   0.4281   0.0757   0.0111   0.4828
#> 353:    0.4408   1.0208   2.9261   0.0074   0.4278   0.0757   0.0111   0.4828
#> 354:    0.4406   1.0207   2.9261   0.0074   0.4275   0.0757   0.0112   0.4828
#> 355:    0.4406   1.0207   2.9262   0.0074   0.4274   0.0757   0.0112   0.4829
#> 356:    0.4406   1.0208   2.9263   0.0074   0.4273   0.0757   0.0112   0.4827
#> 357:    0.4406   1.0208   2.9270   0.0074   0.4272   0.0757   0.0112   0.4828
#> 358:    0.4406   1.0208   2.9272   0.0074   0.4269   0.0758   0.0112   0.4829
#> 359:    0.4407   1.0207   2.9280   0.0074   0.4264   0.0758   0.0112   0.4829
#> 360:    0.4406   1.0208   2.9281   0.0074   0.4259   0.0758   0.0112   0.4829
#> 361:    0.4405   1.0207   2.9280   0.0074   0.4260   0.0757   0.0112   0.4828
#> 362:    0.4405   1.0208   2.9281   0.0074   0.4260   0.0755   0.0112   0.4827
#> 363:    0.4407   1.0207   2.9280   0.0074   0.4259   0.0755   0.0112   0.4827
#> 364:    0.4406   1.0205   2.9282   0.0074   0.4256   0.0754   0.0112   0.4827
#> 365:    0.4406   1.0205   2.9279   0.0074   0.4256   0.0754   0.0112   0.4827
#> 366:    0.4407   1.0205   2.9273   0.0074   0.4254   0.0754   0.0112   0.4825
#> 367:    0.4408   1.0205   2.9268   0.0074   0.4253   0.0753   0.0112   0.4824
#> 368:    0.4410   1.0203   2.9271   0.0074   0.4255   0.0754   0.0112   0.4824
#> 369:    0.4410   1.0205   2.9274   0.0074   0.4255   0.0754   0.0112   0.4824
#> 370:    0.4408   1.0205   2.9271   0.0074   0.4252   0.0754   0.0112   0.4824
#> 371:    0.4406   1.0206   2.9268   0.0074   0.4250   0.0754   0.0112   0.4824
#> 372:    0.4407   1.0206   2.9264   0.0074   0.4251   0.0754   0.0112   0.4823
#> 373:    0.4408   1.0206   2.9260   0.0074   0.4252   0.0753   0.0112   0.4822
#> 374:    0.4410   1.0205   2.9259   0.0074   0.4250   0.0754   0.0112   0.4821
#> 375:    0.4409   1.0205   2.9257   0.0074   0.4248   0.0753   0.0112   0.4820
#> 376:    0.4410   1.0205   2.9256   0.0074   0.4247   0.0754   0.0112   0.4820
#> 377:    0.4408   1.0206   2.9256   0.0074   0.4243   0.0753   0.0112   0.4821
#> 378:    0.4407   1.0209   2.9256   0.0074   0.4240   0.0753   0.0112   0.4823
#> 379:    0.4404   1.0209   2.9260   0.0074   0.4239   0.0753   0.0112   0.4824
#> 380:    0.4400   1.0210   2.9261   0.0074   0.4238   0.0754   0.0112   0.4825
#> 381:    0.4400   1.0210   2.9262   0.0074   0.4239   0.0755   0.0112   0.4826
#> 382:    0.4403   1.0210   2.9265   0.0074   0.4241   0.0755   0.0112   0.4827
#> 383:    0.4404   1.0211   2.9272   0.0074   0.4243   0.0755   0.0112   0.4826
#> 384:    0.4402   1.0212   2.9276   0.0074   0.4242   0.0757   0.0112   0.4827
#> 385:    0.4402   1.0212   2.9281   0.0074   0.4241   0.0757   0.0112   0.4826
#> 386:    0.4402   1.0212   2.9289   0.0074   0.4242   0.0757   0.0112   0.4826
#> 387:    0.4403   1.0212   2.9294   0.0074   0.4247   0.0758   0.0112   0.4826
#> 388:    0.4403   1.0211   2.9299   0.0074   0.4249   0.0758   0.0112   0.4826
#> 389:    0.4402   1.0211   2.9307   0.0074   0.4250   0.0758   0.0112   0.4825
#> 390:    0.4401   1.0212   2.9312   0.0073   0.4247   0.0759   0.0113   0.4826
#> 391:    0.4401   1.0213   2.9318   0.0073   0.4243   0.0759   0.0113   0.4826
#> 392:    0.4403   1.0214   2.9321   0.0073   0.4243   0.0760   0.0113   0.4826
#> 393:    0.4403   1.0215   2.9320   0.0073   0.4244   0.0760   0.0113   0.4827
#> 394:    0.4403   1.0215   2.9313   0.0073   0.4247   0.0759   0.0113   0.4827
#> 395:    0.4404   1.0214   2.9308   0.0074   0.4247   0.0758   0.0113   0.4827
#> 396:    0.4407   1.0212   2.9302   0.0074   0.4247   0.0759   0.0113   0.4827
#> 397:    0.4408   1.0211   2.9301   0.0074   0.4246   0.0758   0.0113   0.4828
#> 398:    0.4408   1.0211   2.9299   0.0074   0.4246   0.0759   0.0113   0.4830
#> 399:    0.4410   1.0211   2.9298   0.0074   0.4246   0.0759   0.0113   0.4831
#> 400:    0.4410   1.0212   2.9291   0.0074   0.4248   0.0760   0.0113   0.4831
#> 401:    0.4410   1.0212   2.9287   0.0074   0.4249   0.0760   0.0113   0.4831
#> 402:    0.4411   1.0211   2.9281   0.0074   0.4249   0.0761   0.0113   0.4830
#> 403:    0.4412   1.0212   2.9275   0.0074   0.4251   0.0761   0.0114   0.4830
#> 404:    0.4412   1.0212   2.9270   0.0074   0.4254   0.0762   0.0114   0.4829
#> 405:    0.4413   1.0212   2.9272   0.0074   0.4255   0.0762   0.0113   0.4829
#> 406:    0.4416   1.0212   2.9272   0.0074   0.4257   0.0761   0.0113   0.4829
#> 407:    0.4418   1.0212   2.9274   0.0074   0.4259   0.0761   0.0113   0.4828
#> 408:    0.4418   1.0212   2.9278   0.0074   0.4258   0.0760   0.0113   0.4827
#> 409:    0.4417   1.0212   2.9282   0.0074   0.4254   0.0760   0.0113   0.4827
#> 410:    0.4418   1.0213   2.9283   0.0074   0.4250   0.0760   0.0113   0.4828
#> 411:    0.4419   1.0215   2.9279   0.0074   0.4248   0.0760   0.0113   0.4829
#> 412:    0.4421   1.0215   2.9281   0.0074   0.4250   0.0761   0.0113   0.4829
#> 413:    0.4421   1.0215   2.9283   0.0074   0.4248   0.0761   0.0113   0.4831
#> 414:    0.4421   1.0215   2.9287   0.0074   0.4247   0.0761   0.0113   0.4832
#> 415:    0.4423   1.0214   2.9283   0.0074   0.4245   0.0761   0.0113   0.4832
#> 416:    0.4425   1.0213   2.9279   0.0074   0.4245   0.0760   0.0113   0.4832
#> 417:    0.4424   1.0214   2.9282   0.0074   0.4245   0.0760   0.0113   0.4833
#> 418:    0.4425   1.0214   2.9282   0.0074   0.4250   0.0760   0.0113   0.4834
#> 419:    0.4426   1.0214   2.9282   0.0074   0.4251   0.0760   0.0113   0.4833
#> 420:    0.4427   1.0214   2.9288   0.0074   0.4252   0.0760   0.0113   0.4833
#> 421:    0.4428   1.0214   2.9291   0.0074   0.4254   0.0760   0.0113   0.4833
#> 422:    0.4430   1.0213   2.9294   0.0074   0.4258   0.0760   0.0114   0.4833
#> 423:    0.4430   1.0213   2.9295   0.0074   0.4262   0.0759   0.0114   0.4833
#> 424:    0.4429   1.0213   2.9292   0.0074   0.4263   0.0758   0.0114   0.4833
#> 425:    0.4427   1.0212   2.9286   0.0074   0.4262   0.0758   0.0114   0.4833
#> 426:    0.4425   1.0212   2.9280   0.0074   0.4262   0.0757   0.0114   0.4831
#> 427:    0.4425   1.0212   2.9275   0.0074   0.4265   0.0756   0.0114   0.4830
#> 428:    0.4426   1.0212   2.9271   0.0074   0.4269   0.0756   0.0114   0.4829
#> 429:    0.4427   1.0211   2.9269   0.0074   0.4272   0.0755   0.0114   0.4830
#> 430:    0.4429   1.0211   2.9274   0.0074   0.4275   0.0755   0.0115   0.4830
#> 431:    0.4432   1.0211   2.9281   0.0074   0.4277   0.0754   0.0115   0.4829
#> 432:    0.4433   1.0210   2.9282   0.0074   0.4277   0.0754   0.0115   0.4831
#> 433:    0.4435   1.0211   2.9287   0.0074   0.4277   0.0753   0.0115   0.4832
#> 434:    0.4435   1.0211   2.9288   0.0074   0.4278   0.0754   0.0115   0.4831
#> 435:    0.4434   1.0211   2.9293   0.0074   0.4280   0.0754   0.0115   0.4831
#> 436:    0.4435   1.0211   2.9302   0.0074   0.4279   0.0754   0.0115   0.4833
#> 437:    0.4434   1.0211   2.9311   0.0074   0.4277   0.0754   0.0115   0.4834
#> 438:    0.4436   1.0210   2.9319   0.0074   0.4272   0.0754   0.0115   0.4835
#> 439:    0.4436   1.0208   2.9324   0.0073   0.4269   0.0754   0.0115   0.4836
#> 440:    0.4435   1.0206   2.9328   0.0073   0.4266   0.0754   0.0115   0.4837
#> 441:    0.4435   1.0205   2.9327   0.0073   0.4263   0.0754   0.0115   0.4837
#> 442:    0.4435   1.0203   2.9327   0.0073   0.4262   0.0755   0.0115   0.4835
#> 443:    0.4436   1.0203   2.9330   0.0073   0.4260   0.0755   0.0115   0.4835
#> 444:    0.4437   1.0202   2.9330   0.0073   0.4260   0.0755   0.0115   0.4836
#> 445:    0.4437   1.0201   2.9334   0.0073   0.4258   0.0755   0.0115   0.4837
#> 446:    0.4437   1.0200   2.9338   0.0073   0.4256   0.0755   0.0115   0.4838
#> 447:    0.4438   1.0200   2.9339   0.0073   0.4256   0.0756   0.0115   0.4837
#> 448:    0.4439   1.0200   2.9342   0.0073   0.4257   0.0756   0.0115   0.4837
#> 449:    0.4439   1.0200   2.9348   0.0073   0.4256   0.0757   0.0115   0.4837
#> 450:    0.4439   1.0200   2.9352   0.0073   0.4253   0.0757   0.0115   0.4836
#> 451:    0.4440   1.0198   2.9354   0.0073   0.4254   0.0757   0.0115   0.4836
#> 452:    0.4440   1.0197   2.9355   0.0073   0.4251   0.0757   0.0115   0.4836
#> 453:    0.4441   1.0196   2.9358   0.0073   0.4251   0.0757   0.0115   0.4836
#> 454:    0.4441   1.0196   2.9363   0.0073   0.4250   0.0757   0.0115   0.4836
#> 455:    0.4443   1.0196   2.9367   0.0073   0.4249   0.0757   0.0115   0.4834
#> 456:    0.4443   1.0197   2.9365   0.0073   0.4248   0.0757   0.0115   0.4833
#> 457:    0.4441   1.0197   2.9363   0.0073   0.4245   0.0757   0.0115   0.4834
#> 458:    0.4439   1.0198   2.9362   0.0073   0.4241   0.0757   0.0115   0.4835
#> 459:    0.4438   1.0197   2.9362   0.0073   0.4240   0.0756   0.0115   0.4836
#> 460:    0.4437   1.0196   2.9359   0.0073   0.4238   0.0756   0.0115   0.4837
#> 461:    0.4437   1.0196   2.9352   0.0073   0.4235   0.0755   0.0115   0.4837
#> 462:    0.4438   1.0196   2.9350   0.0073   0.4233   0.0755   0.0115   0.4836
#> 463:    0.4438   1.0196   2.9349   0.0073   0.4232   0.0754   0.0115   0.4836
#> 464:    0.4438   1.0196   2.9348   0.0073   0.4232   0.0754   0.0115   0.4835
#> 465:    0.4438   1.0195   2.9343   0.0073   0.4232   0.0754   0.0115   0.4834
#> 466:    0.4438   1.0195   2.9339   0.0073   0.4232   0.0753   0.0115   0.4834
#> 467:    0.4440   1.0194   2.9335   0.0073   0.4230   0.0753   0.0115   0.4835
#> 468:    0.4440   1.0193   2.9333   0.0073   0.4225   0.0754   0.0115   0.4836
#> 469:    0.4441   1.0193   2.9333   0.0073   0.4223   0.0753   0.0115   0.4836
#> 470:    0.4441   1.0194   2.9333   0.0073   0.4219   0.0754   0.0116   0.4836
#> 471:    0.4440   1.0195   2.9331   0.0074   0.4216   0.0754   0.0116   0.4836
#> 472:    0.4439   1.0195   2.9332   0.0074   0.4213   0.0753   0.0116   0.4836
#> 473:    0.4439   1.0195   2.9328   0.0074   0.4214   0.0753   0.0116   0.4835
#> 474:    0.4439   1.0196   2.9325   0.0074   0.4215   0.0753   0.0116   0.4835
#> 475:    0.4439   1.0196   2.9325   0.0074   0.4215   0.0754   0.0116   0.4836
#> 476:    0.4439   1.0195   2.9323   0.0074   0.4215   0.0754   0.0116   0.4835
#> 477:    0.4439   1.0195   2.9324   0.0074   0.4217   0.0754   0.0115   0.4836
#> 478:    0.4438   1.0195   2.9323   0.0074   0.4217   0.0755   0.0115   0.4836
#> 479:    0.4437   1.0194   2.9323   0.0074   0.4217   0.0755   0.0115   0.4837
#> 480:    0.4437   1.0194   2.9323   0.0074   0.4216   0.0755   0.0115   0.4837
#> 481:    0.4438   1.0193   2.9321   0.0074   0.4214   0.0755   0.0115   0.4837
#> 482:    0.4439   1.0193   2.9318   0.0074   0.4215   0.0755   0.0115   0.4837
#> 483:    0.4440   1.0192   2.9318   0.0074   0.4214   0.0755   0.0115   0.4837
#> 484:    0.4440   1.0193   2.9317   0.0074   0.4214   0.0755   0.0115   0.4837
#> 485:    0.4442   1.0193   2.9314   0.0074   0.4215   0.0755   0.0115   0.4837
#> 486:    0.4443   1.0194   2.9314   0.0074   0.4216   0.0755   0.0115   0.4837
#> 487:    0.4443   1.0193   2.9314   0.0074   0.4219   0.0756   0.0115   0.4837
#> 488:    0.4444   1.0192   2.9314   0.0074   0.4222   0.0756   0.0115   0.4837
#> 489:    0.4445   1.0193   2.9315   0.0074   0.4223   0.0756   0.0115   0.4838
#> 490:    0.4444   1.0194   2.9317   0.0074   0.4222   0.0756   0.0115   0.4837
#> 491:    0.4446   1.0193   2.9318   0.0074   0.4221   0.0756   0.0115   0.4838
#> 492:    0.4448   1.0193   2.9320   0.0074   0.4221   0.0756   0.0115   0.4837
#> 493:    0.4448   1.0193   2.9322   0.0074   0.4221   0.0756   0.0115   0.4837
#> 494:    0.4446   1.0193   2.9327   0.0074   0.4218   0.0756   0.0115   0.4837
#> 495:    0.4445   1.0194   2.9329   0.0074   0.4217   0.0756   0.0115   0.4838
#> 496:    0.4444   1.0195   2.9328   0.0074   0.4216   0.0756   0.0115   0.4837
#> 497:    0.4443   1.0195   2.9328   0.0074   0.4214   0.0756   0.0115   0.4836
#> 498:    0.4443   1.0194   2.9330   0.0074   0.4216   0.0755   0.0115   0.4836
#> 499:    0.4445   1.0193   2.9335   0.0073   0.4219   0.0755   0.0115   0.4835
#> 500:    0.4447   1.0192   2.9335   0.0074   0.4218   0.0755   0.0115   0.4834

#> Calculating covariance matrix

#> 

#> → creating full model...

#> → pruning branches (`if`/`else`)...

#> ✔ done

#> → loading into symengine environment...

#> ✔ done

#> → compiling EBE model...

#>  

#> ✔ done

#> Needed Covariates:

#> [1] "WT"

#> Calculating residuals/tables

#> done

fit

#> ── nlmixr SAEM(ODE); OBJF not calculated fit ───────────────────────────────────
#> 
#>  Gaussian/Laplacian Likelihoods: AIC(x) or x$objf etc. 
#>  FOCEi CWRES & Likelihoods: addCwres(x) 
#> 
#> ── Time (sec x$time): ──────────────────────────────────────────────────────────
#> 
#>           saem   setup table covariance   other
#> elapsed 10.777 2.36556 0.014      0.012 0.78244
#> 
#> ── Population Parameters (x$parFixed or x$parFixedDf): ─────────────────────────
#> 
#>        Parameter    Est.      SE %RSE     Back-transformed(95%CI) BSV(CV%)
#> tka       Log Ka   0.445   0.198 44.5            1.56 (1.06, 2.3)     72.4
#> tcl       Log Cl    1.02  0.0863 8.47           2.77 (2.34, 3.28)     28.0
#> tv         Log V    2.93   0.288 9.82             18.8 (10.7, 33)     10.7
#> add.sd             0.695                                    0.695         
#> wt.est           0.00735 0.00409 55.6 0.00735 (-0.000664, 0.0154)         
#>        Shrink(SD)%
#> tka       -0.577% 
#> tcl         4.07% 
#> tv          13.1% 
#> add.sd            
#> wt.est            
#>  
#>   Covariance Type (x$covMethod): linFim
#>   Some strong fixed parameter correlations exist (x$cor) :
#>        cor:tcl,tka     cor:tv,tka cor:wt.est,tka     cor:tv,tcl cor:wt.est,tcl 
#>       -0.0533          0.0175         0.00134         -0.0271         0.00821   
#>  cor:wt.est,tv 
#>        -0.991  
#>  
#> 
#>   No correlations in between subject variability (BSV) matrix
#>   Full BSV covariance (x$omega) or correlation (x$omegaR; diagonals=SDs) 
#>   Distribution stats (mean/skewness/kurtosis/p-value) available in x$shrink 
#> 
#> ── Fit Data (object x is a modified tibble): ───────────────────────────────────
#> # A tibble: 132 x 20
#>   ID     TIME    DV  PRED    RES IPRED   IRES  IWRES eta.ka eta.cl  eta.v    cp
#>   <fct> <dbl> <dbl> <dbl>  <dbl> <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl> <dbl>
#> 1 1     0      0.74  0     0.74   0     0.74   1.06   0.137 -0.497 -0.134  0   
#> 2 1     0.25   2.84  3.03 -0.190  3.88 -1.04  -1.50   0.137 -0.497 -0.134  3.88
#> 3 1     0.570  6.57  5.44  1.13   6.80 -0.234 -0.337  0.137 -0.497 -0.134  6.80
#> # … with 129 more rows, and 8 more variables: depot <dbl>, center <dbl>,
#> #   ka <dbl>, cl <dbl>, v <dbl>, tad <dbl>, dosenum <dbl>, WT <dbl>

# }